==> 3_MGB_KAS_global_mets_association.Rmd <==
---
title: "Regression analysis of global metabolites in PEGASUS"
author: "Yulu Chen"
date: "10/22/2023"
output: 
  html_document: 
    toc: true
    toc_float: 
      collapsed: false
      smooth_scroll: false
    code_folding: hide
editor_options: 
  chunk_output_type: console
---

```{r, include = FALSE}
##### 1. Setup #####
## Packages

pkg <- c("tidyverse", "data.table", "here", "survival", "gtools", "tableone", "DT", "sjPlot", 
         "sjmisc", "grid", "gridExtra", "ggpubr", "table1", "readxl", "ggvenn", "patchwork")
for (p in pkg) {
        if (require(p, character.only = T)) {
                print(paste0(p, " loaded successfully"))
        } else {
                install.packages(p)
                require(p, character.only = T)
                print(paste0(p, " downloaded and loaded successfully"))
        }
}

## Paths
pegasus_dir <- "/udd/reyul/mets_ratio_CodeReview/data/PEGASUS/"

dat_dir <- str_c(pegasus_dir, "2_assembly_global_targeted/")
fig_dir <- str_c("/udd/reyul/mets_ratio_CodeReview/", "figures/3_PEGASUS_global_mets_association/")
res_dir <- str_c("/udd/reyul/mets_ratio_CodeReview/", "results/3_PEGASUS_global_mets_association/")

## Hard-coded numbers

missing_thld <- 0.30
alpha_thld <- 0.05
z_95 <- qnorm(p = 0.975)

## Functions
pvalue <- function(x, ...) {
    # Construct vectors of data y, and groups (strata) g
    y <- unlist(x)
    g <- factor(rep(1:length(x), times=sapply(x, length)))
    if (is.numeric(y)) {
        # For numeric variables, perform a standard 2-sample t-test
        p <- as.data.frame(summary(aov(y ~ g))[[1]])[1,5]
    } else {
        # For categorical variables, perform a chi-squared test of independence
        p <- chisq.test(table(y, g))$p.value
    }
    # Format the p-value, using an HTML entity for the less-than sign.
    # The initial empty string places the output on the line below the variable label.
    c("", sub("<", "&lt;", format.pval(p, digits=6, eps=0.000001)))
}

```


# 1. Load data
Using the data generated by Yulu and Mengna, including the phenotype/EMR data, global metabolomics data (HILIC+, Lipid+, and Lipid-), targeted sphingolipids data, targeted microbial data

global metabolomics data:
- remove metatolites with sample CV >= 25%
- remove metabolites with >= 30% missing in mets
- impute with half min
- log10 transformation & pareto-scaling

targeted sphingolipids:  remove metabolites with >= 30% missing in mets & Standardize

targeted microbial: remove metabolites with >= 30% missing in mets & Standardize

targeted steroids: remove metabolites with >= 30% missing in mets & Standardize

```{r}

load(here(dat_dir, "all_platforms_data_updated.RData"), verbose = T)

```

# 2. Table 1
## 2.1 trim data
- _trim: subjects missing bmi & matched subjects removed  

```{r}
nrow(pheno_mets) #1132

stratum_to_rm <- pheno_mets[is.na(bmi), unique(Stratum)]
stratum_to_rm <- c(stratum_to_rm, pheno_mets[, .N, Stratum][N == 1, Stratum])
print(str_c("Number of strata removed due to missing BMI: ", 
            length(stratum_to_rm)))
pheno_mets_trim <- pheno_mets[!(Stratum %in% stratum_to_rm), ]

```

## 2.2 Table 1
basic demographic
```{r}
table1(~ Age + bmi + Gender + Race_3cat + Ethnicity_2cat + Nonsmoker | SampleGroup, data = pheno_mets_trim, overall = T, extra.col=list(`P-value`=pvalue))
```

Medication 

- ICS prescriptions x year before serum collection
- OCS prescription x year before serum collection
- Asthma exacerbation  x year before serum collection 
```{r}
table1(~ ics_trim_totnum_1y +  ics_trim_totnum_2y + ics_trim_totnum_3y + ics_trim_totnum_5y + ics_trim_totnum_10y + ics_trim_totnum +
         ocs_trim_totnum_1y +  ocs_trim_totnum_2y + ocs_trim_totnum_3y + ocs_trim_totnum_5y + ocs_trim_totnum_10y + ocs_trim_totnum + 
         asthma_exacerbation_trim_totnum_1y +  asthma_exacerbation_trim_totnum_2y + asthma_exacerbation_trim_totnum_3y + 
         asthma_exacerbation_trim_totnum_5y + asthma_exacerbation_trim_totnum_10y + asthma_exacerbation_trim_totnum +
         ics_trim_totnum_5y_after + ocs_trim_totnum_5y_after | SampleGroup, data = pheno_mets_trim, overall = T,
       extra.col=list(`P-value`=pvalue))
```

Lab test
- Eosinophil count closest to serum collection, K/uL
- Neutrophil count closest to serum collection, K/uL
- Log10(IgE) closest to serum collection, KU/L
```{r}
table1(~ EOScount_median_6M + EOScount_median_12M + EOScount_median_24M + 
         EOSpct_median_6M + EOSpct_median_12M + EOSpct_median_24M + eos + 
         NEUcount_median_6M + NEUcount_median_12M + NEUcount_median_24M + 
         NEUpct_median_6M + NEUpct_median_12M + NEUpct_median_24M + neut +
         log10(ige) | SampleGroup, data = pheno_mets_trim, overall = T,
       extra.col=list(`P-value`=pvalue))

t.test(EOScount_median_6M ~ SampleGroup, pheno_mets_trim)
t.test(EOScount_median_12M ~ SampleGroup, pheno_mets_trim)
t.test(EOScount_median_24M ~ SampleGroup, pheno_mets_trim)
t.test(EOSpct_median_6M ~ SampleGroup, pheno_mets_trim)
t.test(EOSpct_median_12M ~ SampleGroup, pheno_mets_trim)
t.test(EOSpct_median_24M ~ SampleGroup, pheno_mets_trim)
t.test(eos ~ SampleGroup, pheno_mets_trim)

t.test(NEUcount_median_6M ~ SampleGroup, pheno_mets_trim)
t.test(NEUcount_median_12M ~ SampleGroup, pheno_mets_trim)
t.test(NEUcount_median_24M ~ SampleGroup, pheno_mets_trim)
t.test(NEUpct_median_6M ~ SampleGroup, pheno_mets_trim)
t.test(NEUpct_median_12M ~ SampleGroup, pheno_mets_trim)
t.test(NEUpct_median_24M ~ SampleGroup, pheno_mets_trim)
t.test(neut ~ SampleGroup, pheno_mets_trim)

t.test(log10(ige) ~ SampleGroup, pheno_mets_trim)
```

Lung function

```{r}
table1( ~ fev1_prebd_pctpred + FEV1_median_6M + FEV1_median_12M + FEV1_median_24M +
          fvc_prebd_pctpred + FVC_median_6M + FVC_median_12M + FVC_median_24M +
          fev1fvc_prebd_pctpred + FEV1_FVC_median_6M + FEV1_FVC_median_12M + FEV1_FVC_median_24M| SampleGroup, data = pheno_mets_trim, overall = T,
       extra.col=list(`P-value`=pvalue))
```


# 3. Conditional Logistic Regression model functions
```{r}
clogi_res_fnc <- function(outc, mets_list, dat, mets_info, add_covar = "", 
                          first_cols = c("feat_id", "tentative_annotation", "beta", "pval", "fdr_bh", "or", "lower95", "upper95", 
                                         "med_feat", "wo_ms2", "rt_min", "mz", "feature", "platform", 
                                         "totaln", "casen", "cv", "pct_na_met"), 
                          na_thld = missing_thld) {
        
        clogi_mdl_tmp <- function(met) {
                
                fit <- eval(parse(text = str_c("clogit(", outc, " ~ ", met, " + strata(Stratum)", 
                                               add_covar, ", data = ", dat, ")")))
                
                coef <- summary(fit)$coefficients
                coef <- coef[rownames(coef) == met, ]

                totaln <- summary(fit)$n
                casen <- summary(fit)$nevent
                
                output <- t(c(outc, met, coef, totaln, casen)) %>% as.data.table()
                colnames(output) <- c("outc", "feat_id", "beta", "or", "se", "zval", "pval", 
                                      "totaln", "casen")
                output[, (colnames(output)[-c(1:2)]) := lapply(.SD, as.numeric), .SDcols = colnames(output)[-c(1:2)]]
                
                return(output)
        }
        
        res <- lapply(mets_list, function(x) {clogi_mdl_tmp(met = x)}) %>% rbindlist()
        
        eval(parse(text = str_c("mets_info <- copy(", mets_info, ")")))
        res <- merge(res, mets_info, by = "feat_id")
        res[pct_na_met <= na_thld, fdr_bh := p.adjust(pval, method = "BH")] 
        res[, ':='(lower95 = exp(beta - z_95 * se), 
                   upper95 = exp(beta + z_95 * se))] 
        
        setcolorder(res, first_cols)
        res <- res[order(pval)]
        return(res)
}

```

# 4. Conduct association analysis with Asthma
## 4.1 Association
Keep Age, Gender, race are same in two groups. No smokers. 
model: Asthma ~ strata + bmi
```{r}
res_4 <- clogi_res_fnc(outc = "asthma_num", 
                       mets_list = mets_list_analysis, 
                       dat = "pheno_mets_trim", 
                       mets_info = "mets_info_all", 
                       add_covar = " + bmi")

## Save results
saveRDS(res_4, file = here(res_dir, "res_4.rds"))


```

## 4.2 Effective number of test (ENT)
```{r}
mets_list_global <- c(hilic_pos_mets_info[is.na(qc_rm) & pct_na_met <= 0.30, feat_id], 
                      lipid_pos_mets_info[is.na(qc_rm) & pct_na_met <= 0.30, feat_id], 
                      lipid_neg_mets_info[is.na(qc_rm) & pct_na_met <= 0.30, feat_id])

pca <- prcomp(pheno_mets[, ..mets_list_global])

print(str_c("Number of PCs needed to account for 80% of all variance in global metabolome: ", 
            which(cumsum(pca$sdev^2)/sum(pca$sdev^2) >= 0.8)[1]))
print(str_c("ENT80% + bonferroni corrected p-value: ", 
            round(0.05 / which(cumsum(pca$sdev^2)/sum(pca$sdev^2) >= 0.8)[1], 10)))

print(str_c("ENT80% + bonferroni corrected -log10(p-value): ", 
            round(-log10(0.05 / which(cumsum(pca$sdev^2)/sum(pca$sdev^2) >= 0.8)[1]), 3)))
print(str_c("Number of PCs needed to account for 90% of all variance in global metabolome: ", 
            which(cumsum(pca$sde^2)/sum(pca$sdev^2) >= 0.9)[1]))

print(str_c("ENT90% + bonferroni corrected p-value: ", 
            round(0.05 / which(cumsum(pca$sdev^2)/sum(pca$sdev^2) >= 0.9)[1], 10)))
print(str_c("ENT90% + bonferroni corrected -log10(p-value): ", 
            round(-log10(0.05 / which(cumsum(pca$sdev^2)/sum(pca$sdev^2) >= 0.9)[1]), 3)))

print(str_c("ENT95% + bonferroni corrected p-value: ", 
            round(0.05 / which(cumsum(pca$sdev^2)/sum(pca$sdev^2) >= 0.95)[1], 10)))
print(str_c("ENT95% + bonferroni corrected -log10(p-value): ", 
            round(-log10(0.05 / which(cumsum(pca$sdev^2)/sum(pca$sdev^2) >= 0.95)[1]), 3)))
```

## 4.3 Summary the significant results
significant mets identified in conditional logistic regression:
0.05: global 274
0.01: global 135
ENT95% + borferroni corrected  global 201
```{r}
length(res_4[tentative_annotation != "Unknown" & pval < 0.05 & platform != "Targeted-sphingolipids" & platform != "Targeted-microbial-mets" & platform != "Targeted-steroids", feat_id]) #274
length(res_4[tentative_annotation != "Unknown" & pval < 0.01 & platform != "Targeted-sphingolipids" & platform != "Targeted-microbial-mets" & platform != "Targeted-steroids", feat_id]) #135
length(res_4[pval < 1.07E-04 & platform != "Targeted-sphingolipids" & platform != "Targeted-microbial-mets" & platform != "Targeted-steroids", feat_id])

```

## 4.4 Plot the significant global mets
```{r}
# load the significant mets by manually annotating
sig_global_mets_df <- read_excel(here(res_dir, "asthma_sig_global_mets_manual_3.xlsx"), sheet = 3)

sig_global_xeno_df <- read_excel(here(res_dir, "asthma_sig_global_mets_manual_3.xlsx"), sheet = 4)

sig_global_unknown_df <- read_excel(here(res_dir, "asthma_sig_global_mets_manual_3.xlsx"), sheet = 5)

# add confidence interval
sig_global_mets_df_updated <- merge(sig_global_mets_df, res_4[,c("feat_id", "lower95", "upper95")], by = "feat_id", all.x = T)
sig_global_xeno_df_updated <- merge(sig_global_xeno_df, res_4[,c("feat_id", "lower95", "upper95")], by = "feat_id", all.x = T)
sig_global_unknown_df_updated <- merge(sig_global_unknown_df, res_4[,c("feat_id", "lower95", "upper95")], by = "feat_id", all.x = T)

summary(sig_global_mets_df_updated$lower95)
summary(sig_global_mets_df_updated$upper95)

# plot all
sig_global_mets_df_updated$tentative_superpathway[sig_global_mets_df_updated$tentative_superpathway == "Cofactors and Vitamins"] <- "Cofactors & Vitamins"

sig_global_mets_df_updated$or <- log(sig_global_mets_df_updated$or)
sig_global_mets_df_updated$lower95 <- log(sig_global_mets_df_updated$lower95)
sig_global_mets_df_updated$upper95 <- log(sig_global_mets_df_updated$upper95)

sig_mets_gplot_1 <- ggplot(sig_global_mets_df_updated, 
                           aes(reorder(annotation, -or), y = or, color = or >0)) +
   scale_colour_manual(name = '', values = setNames(c('red','blue'),c(T, F)))+
   geom_rect(aes(fill = tentative_subpathway),xmin = -Inf,xmax = Inf,
             ymin = -Inf,ymax = Inf,alpha = 0.3, linejoin = "round") +
   facet_grid(tentative_subpathway ~ factor(tentative_superpathway, levels = c('Amino Acid', 'Lipid', 'Carbohydrate', 'Cofactors & Vitamins', 'Nucleotide', 'Unknown')), scales = "free_y", space = "free_y", labeller = label_wrap_gen()) +
   geom_errorbar(aes(ymin = lower95, ymax = upper95), width = 0.25) +
   geom_pointrange(aes(ymin = lower95, ymax = upper95), size = 0.25) +
   geom_hline(yintercept = 0, linetype = 2) +
   labs(
     x = "Metabolite",
     y = "log(OR) (95% Confidence Interval)"
   ) +
   theme_minimal() +
   theme_bw() +
   theme(
     title = element_text(size = 16),
     strip.text.x = element_text(size = 12),
     strip.text.y = element_text(size = 10),
     axis.title = element_text(size = 12),
     axis.text = element_text(size = 9),
     legend.position = "",
     legend.title = element_text(size = 12),
     legend.text = element_text(size = 12),
    strip.text.y.right = element_text(angle = 0)) +
   coord_flip() + scale_fill_viridis_d(option = "D", direction = -1, begin = 0, end = 1, alpha = 0.1)

ggsave(sig_mets_gplot_1, 
       filename = here(fig_dir, "sig_mets_summary_gplot_all.png"),
       width = 14.5,
       height = 18, dpi = 300)


# plot need
## annotated the unknown pathway
sig_global_mets_df_updated$tentative_superpathway[sig_global_mets_df_updated$feat_id == "F299_hilic_pos"] <- "Amino Acid"
sig_global_mets_df_updated$tentative_subpathway[sig_global_mets_df_updated$feat_id == "F299_hilic_pos"] <- "Tryptophan Metabolism"

sig_global_mets_df_updated$tentative_superpathway[sig_global_mets_df_updated$feat_id == "F83_hilic_pos"] <- "Amino Acid"
sig_global_mets_df_updated$tentative_subpathway[sig_global_mets_df_updated$feat_id == "F83_hilic_pos"] <- "Glycine, Serine and Threonine Metabolism"


## remove Cofactors & Vitamins, Nucleotide, and Unknown
table(sig_global_mets_df_updated$tentative_superpathway)
sig_global_mets_df_updated_1 <- sig_global_mets_df_updated[sig_global_mets_df_updated$tentative_superpathway != "Unknown" & sig_global_mets_df_updated$tentative_superpathway != "Cofactors & Vitamins" & sig_global_mets_df_updated$tentative_superpathway != "Nucleotide",]

## remove no-overlap sub-pathway
table(sig_global_mets_df_updated_1$tentative_subpathway)
sig_global_mets_df_updated_1 <- sig_global_mets_df_updated_1[sig_global_mets_df_updated_1$tentative_subpathway != "Diacylglycerol" & sig_global_mets_df_updated_1$tentative_subpathway != "N-acylethanolamines" & sig_global_mets_df_updated_1$tentative_subpathway != "Polyamine Metabolism" &
sig_global_mets_df_updated_1$tentative_subpathway != "Primary Bile Acid Metabolism" & sig_global_mets_df_updated_1$tentative_subpathway != "Terpene lactone" & 
sig_global_mets_df_updated_1$tentative_subpathway != "Tyrosine Metabolism",]

## combined some subpathway together
table(sig_global_mets_df_updated_1$tentative_subpathway)
sig_global_mets_df_updated_1$tentative_subpathway_updated <- sig_global_mets_df_updated_1$tentative_subpathway

sig_global_mets_df_updated_1$tentative_subpathway_updated[sig_global_mets_df_updated_1$tentative_subpathway_updated == "Androgenic Steroids" | sig_global_mets_df_updated_1$tentative_subpathway_updated == "Corticosteroids" | sig_global_mets_df_updated_1$tentative_subpathway_updated == "Estrogenic Steroids" | sig_global_mets_df_updated_1$tentative_subpathway_updated == "Sterol"] <- "Steroid-related metabolites"

sig_global_mets_df_updated_1$tentative_subpathway_updated[sig_global_mets_df_updated_1$tentative_subpathway_updated == "Sphingolipid Synthesis" | sig_global_mets_df_updated_1$tentative_subpathway_updated == "Sphingosines"] <- "Sphingolipid-related metabolites"

table(sig_global_mets_df_updated_1$tentative_subpathway_updated)

sig_mets_gplot_2 <- ggplot(sig_global_mets_df_updated_1, 
                           aes(reorder(annotation, -or), y = or, color = or >0)) +
   scale_colour_manual(name = '', values = setNames(c('red','blue'),c(T, F)))+
   geom_rect(aes(),xmin = -Inf,xmax = Inf,
             ymin = -Inf,ymax = Inf, alpha = 0, color = 'black', linejoin = "round") +
   facet_grid(tentative_subpathway_updated ~ factor(tentative_superpathway, levels = c('Amino Acid', 'Lipid', 'Carbohydrate')), scales = "free_y", space = "free_y", labeller = label_wrap_gen()) +
   geom_errorbar(aes(ymin = lower95, ymax = upper95), width = 0.25) +
   geom_pointrange(aes(ymin = lower95, ymax = upper95), size = 0.25) +
   geom_hline(yintercept = 0, linetype = 2) +
   labs(
     x = "Metabolite",
     y = "log(OR) (95% Confidence Interval)"
   ) +
   theme_minimal() +
   theme_bw() +
   theme(
     title = element_text(size = 16),
     strip.text.x = element_text(size = 12),
     strip.text.y = element_text(size = 10),
     axis.title = element_text(size = 12),
     axis.text = element_text(size = 9),
     legend.position = "",
     legend.title = element_text(size = 12),
     legend.text = element_text(size = 12),
    strip.text.y.right = element_text(angle = 0)) +
   coord_flip() + scale_fill_viridis_d(option = "D", direction = -1, begin = 0, end = 1, alpha = 0.1)

ggsave(sig_mets_gplot_2, 
       filename = here(fig_dir, "sig_mets_summary_gplot_overlap.png"),
       width = 12,
       height = 14, dpi = 300)
```


## 4.5 Plot the significant targeted mets
```{r}
# sphingolipids
sig_targeted_sph <- res_4[pval < 0.05 & platform == "Targeted-sphingolipids",]
sig_targeted_sph$common_name <- c("Sphingosine 1P",
                                  "Sphinganine 1P",
                                  "Sphingosine",
                                  "Sphinganine",
                                  "Sphingosine(d18:2)",
                                  "Cer(d18:1/18:0)",
                                  "Sphingosine-1P(d18:2-P)",
                                  "Cer 1P(d18:1/24:0)",
                                  "Cer 1P(d18:1/24:1)",
                                  "HexCer(d18:1/24:0(15Z))",
                                  "HexCer(d18:1/18:0)",
                                  "SM(d16:1/24:0)",
                                  "DhCer(d18:0/24:1)",
                                  "SM(d18:2/24:0)",
                                  "HexCer(d18:1/16:0)")

sph_mets_info$`SUPER PATHWAY` <- "sphingolipid metabolism"
sph_mets_info[, `SUB PATHWAY` := case_when(grepl("^HexCer|LacCer", feat_id) ~ "Hex/LacCeramide", 
                                    grepl("^SM", feat_id) ~ "Sphingomyelin", 
                                    grepl("^Sph|Sphd|S1P", feat_id) ~ "Sphingosine/S1P", 
                                    grepl("^Cer1P", feat_id) ~ "Ceramide-1-phosphate", 
                                    grepl("^Cer[1-2][0-9]", feat_id) ~ "Ceramide", 
                                    grepl("^DhCer", feat_id) ~ "Dihydroceramide", 
                                    grepl("^Spa|Spa1P", feat_id) ~ "Sphinganine/Spa1P", 
                                    feat_id == "Glucosph" ~ "Hexosylsphingosine", 
                                    feat_id == "X3_KS" ~ "X3_Ketosphinganine")]

colnames(sph_mets_info)[1] <- "feat_id"

sig_targeted_sph <- merge(sig_targeted_sph, sph_mets_info[,c("feat_id", "SUPER PATHWAY", "SUB PATHWAY")], by = "feat_id", all.x = T)

p_sph <- ggplot(sig_targeted_sph, 
                aes(reorder(common_name, -or), y = or, color = or >1)) +
   scale_colour_manual(name = '', values = setNames(c('red','blue'),c(T, F)))+
   geom_rect(aes(fill = `SUB PATHWAY`),xmin = -Inf,xmax = Inf,
             ymin = -Inf,ymax = Inf, alpha = 0, color = 'black', linejoin = "round") +
   facet_grid(`SUB PATHWAY` ~ `SUPER PATHWAY`, scales = "free_y", space = "free_y", labeller = label_wrap_gen()) +
   geom_errorbar(aes(ymin = lower95, ymax = upper95), width = 0.25) +
   geom_pointrange(aes(ymin = lower95, ymax = upper95), size = 0.25) +
   geom_hline(yintercept = 1, linetype = 2) +
   labs(
     x = "",
     y = ""
   ) +
   theme_minimal() +
   theme_bw() +
   theme(
     title = element_text(size = 16),
     strip.text.x = element_text(size = 12),
     strip.text.y = element_text(size = 10),
     axis.title = element_text(size = 12),
     axis.text = element_text(size = 9),
     legend.position = "",
     legend.title = element_text(size = 12),
     legend.text = element_text(size = 12),
     strip.text.y.right = element_text(angle = 0),
     strip.background = element_rect(fill = "mistyrose", colour = "black")) +
   coord_flip() + scale_fill_viridis_d(option = "D", direction = -1, begin = 0, end = 1, alpha = 0.1) +
   scale_y_continuous(limits = c(0.2, 1.6))


ggsave(p_sph, 
       filename = here(fig_dir, "sig_targeted_sph.png"),
       width = 6,
       height = 8, dpi = 300)

# microbial-related metabolites
sig_targeted_microb <- res_4[pval < 0.05 & platform == "Targeted-microbial-mets",]
sig_targeted_microb <- merge(sig_targeted_microb, microb_mets_info[,c("feat_id", "SUPER_PATHWAY", "SUB_PATHWAY")], by = "feat_id", all.x = T)

p_microb <- ggplot(sig_targeted_microb, 
                   aes(reorder(tentative_annotation, -or), y = or, color = or >1)) +
   scale_colour_manual(name = '', values = setNames(c('red','blue'),c(T, F)))+
   geom_rect(aes(fill = `SUB_PATHWAY`),xmin = -Inf,xmax = Inf,
             ymin = -Inf,ymax = Inf, alpha = 0, color = 'black', linejoin = "round") +
   facet_grid(`SUB_PATHWAY` ~ `SUPER_PATHWAY`, scales = "free_y", space = "free_y", labeller = label_wrap_gen()) +
   geom_errorbar(aes(ymin = lower95, ymax = upper95), width = 0.25) +
   geom_pointrange(aes(ymin = lower95, ymax = upper95), size = 0.25) +
   geom_hline(yintercept = 1, linetype = 2) +
   labs(
     x = "",
     y = ""
   ) +
   theme_minimal() +
   theme_bw() +
   theme(
     title = element_text(size = 16),
     strip.text.x = element_text(size = 12),
     strip.text.y = element_text(size = 10),
     axis.title = element_text(size = 12),
     axis.text = element_text(size = 9),
     legend.position = "",
     legend.title = element_text(size = 12),
     legend.text = element_text(size = 12),
     strip.text.y.right = element_text(angle = 0),
     strip.background = element_rect(fill = "lightcyan", colour = "black")) +
   coord_flip() + scale_fill_viridis_d(option = "D", direction = -1, begin = 0, end = 1, alpha = 0.1) +
   scale_y_continuous(limits = c(0.2, 1.6))


ggsave(p_microb, 
       filename = here(fig_dir, "sig_targeted_microb.png"),
       width = 6,
       height = 8, dpi = 300)

# steroid-rwlated metabolites
sig_targeted_str <- res_4[pval < 0.05 & platform == "Targeted-steroids",]
sig_targeted_str <- merge(sig_targeted_str, str_mets_info[,c("feat_id", "SUPER PATHWAY", "SUB PATHWAY")], by = "feat_id", all.x = T)

p_str <- ggplot(sig_targeted_str, 
                aes(reorder(tentative_annotation, -or), y = or, color = or >1)) +
   scale_colour_manual(name = '', values = setNames(c('red','blue'),c(T, F)))+
   geom_rect(aes(fill = `SUB PATHWAY`),xmin = -Inf,xmax = Inf,
             ymin = -Inf,ymax = Inf, alpha = 0, color = 'black', linejoin = "round") +
   facet_grid(`SUB PATHWAY` ~ `SUPER PATHWAY`, scales = "free_y", space = "free_y", labeller = label_wrap_gen()) +
   geom_errorbar(aes(ymin = lower95, ymax = upper95), width = 0.25) +
   geom_pointrange(aes(ymin = lower95, ymax = upper95), size = 0.25) +
   geom_hline(yintercept = 1, linetype = 2) +
   labs(
     x = "",
     y = ""
   ) +
   theme_minimal() +
   theme_bw() +
   theme(
     title = element_text(size = 16),
     strip.text.x = element_text(size = 12),
     strip.text.y = element_text(size = 10),
     axis.title = element_text(size = 12),
     axis.text = element_text(size = 9),
     legend.position = "",
     legend.title = element_text(size = 12),
     legend.text = element_text(size = 12),
     strip.text.y.right = element_text(angle = 0),
     strip.background = element_rect(fill = "lemonchiffon", colour = "black")) +
   coord_flip() + scale_fill_viridis_d(option = "D", direction = -1, begin = 0, end = 1, alpha = 0.1) +
   scale_y_continuous(limits = c(0.2, 1.6))


ggsave(p_str, 
       filename = here(fig_dir, "sig_targeted_str.png"),
       width = 6,
       height = 8, dpi = 300)


# combined the plots
combined_plots <- p_str + p_microb + p_sph

ggsave(combined_plots, 
       filename = here(fig_dir, "sig_targeted_all.png"),
       width = 20,
       height = 6, dpi = 300)

```

# 5. Save Data
```{r}
save(res_4,  
     file = str_c(res_dir, "PEGASUS_all_association_asthma_results.RData"))

write.csv(sig_global_mets_df_updated, file = str_c(res_dir, "PEGASUS_sig_global_mets_df_updated.csv"), row.names = FALSE)
write.csv(sig_global_xeno_df_updated, file = str_c(res_dir, "PEGASUS_sig_global_xeno_df_updated.csv"), row.names = FALSE)
write.csv(sig_global_unknown_df_updated, file = str_c(res_dir, "PEGASUS_sig_global_unknown_df_updated.csv"), row.names = FALSE)

write.csv(sig_global_mets_df_updated_1, file = str_c(res_dir, "PEGASUS_sig_mets_df_overlap.csv"), row.names = FALSE)

write.csv(sig_targeted_sph, file = str_c(res_dir, "PEGASUS_sig_targeted_sph.csv"), row.names = FALSE)
write.csv(sig_targeted_str, file = str_c(res_dir, "PEGASUS_sig_targeted_str.csv"), row.names = FALSE)
write.csv(sig_targeted_microb, file = str_c(res_dir, "PEGASUS_sig_targeted_microb.csv"), row.names = FALSE)


```

# Session info
```{r}
sessionInfo()

```
