==> 4_MGBB_asthma_global_mets_association.Rmd <==
---
title: "Regression analysis of global metabolites in MGBB"
author: "Yulu Chen"
date: "10/20/2023"
output: 
  html_document: 
    toc: true
    toc_float: 
      collapsed: false
      smooth_scroll: false
    code_folding: hide
editor_options: 
  chunk_output_type: console
---

```{r, include = FALSE}
##### 1. Setup #####
## Packages

pkg <- c("tidyverse", "data.table", "here", "survival", "gtools", "tableone", "DT", "sjPlot", "sjmisc", "grid", "gridExtra", "ggpubr", "table1", "readxl", "openxlsx", "ggrepel", "gghighlight")
for (p in pkg) {
        if (require(p, character.only = T)) {
                print(paste0(p, " loaded successfully"))
        } else {
                install.packages(p)
                require(p, character.only = T)
                print(paste0(p, " downloaded and loaded successfully"))
        }
}

## Paths
dat_dir <- "/udd/reprk/projects/PartnersBiobank_asthma_metabolomics/code_review/results/"

fig_dir <- str_c("/udd/reyul/mets_ratio_CodeReview/", "figures/4_MGBB_global_mets_association/")
res_dir <- str_c("/udd/reyul/mets_ratio_CodeReview/", "results/4_MGBB_global_mets_association/")

pegasus_dir <- "/udd/reyul/mets_ratio_CodeReview/data/PEGASUS/"

## Hard-coded numbers

missing_thld <- 0.30
sig_digits <- 3
alpha_thld <- 0.05
z_95 <- qnorm(p = 0.975)

## Functions
pvalue <- function(x, ...) {
    # Construct vectors of data y, and groups (strata) g
    y <- unlist(x)
    g <- factor(rep(1:length(x), times=sapply(x, length)))
    if (is.numeric(y)) {
        # For numeric variables, perform a standard 2-sample t-test
        p <- as.data.frame(summary(aov(y ~ g))[[1]])[1,5]
    } else {
        # For categorical variables, perform a chi-squared test of independence
        p <- chisq.test(table(y, g))$p.value
    }
    # Format the p-value, using an HTML entity for the less-than sign.
    # The initial empty string places the output on the line below the variable label.
    c("", sub("<", "&lt;", format.pval(p, digits=6, eps=0.000001)))
}

```

# 1. Read MGBB data

## 1.1 Load mets and pheno data

```{r}
load(str_c(dat_dir, "samp_mets_final_processed_with_pheno_MGBBA_1628803433.RData"), verbose = T)
```

## 1.2 Table1

basic demographic
```{r}
# change the race characters to be 3
final_data$Race_3cat <- with(final_data, ifelse(RACE_cat == "White", "White",
                                         ifelse(RACE_cat == "Black", "Black", "Others")))

# Typical adjustments: Age + Gender + Race + BMI + smoking

table1(~ Age + BMI + Sex + Race_3cat + Ethnicity + Smoking | ASTHMA, data = final_data, overall = T, extra.col=list(`P-value`=pvalue))

```

## 1.3 Effective number of test (ENT)
```{r}
mets_list <- metabolite_summary2$met_ID
length(mets_list) #904

final_data <- as.data.table(final_data)
pca <- prcomp(final_data[, ..mets_list])

print(str_c("Number of PCs needed to account for 80% of all variance in global metabolome: ", 
            which(cumsum(pca$sdev*2)/sum(pca$sdev^2) >= 0.8)[1]))
print(str_c("ENT80% + bonferroni corrected p-value: ", 
            round((0.05 / which(cumsum(pca$sdev*2)/sum(pca$sdev^2) >= 0.8)[1]), 6)))
print(str_c("ENT80% + bonferroni corrected -log10(p-value): ", 
            round(-log10(0.05 / which(cumsum(pca$sdev*2)/sum(pca$sdev^2) >= 0.8)[1]), 3)))


print(str_c("Number of PCs needed to account for 90% of all variance in global metabolome: ", 
            which(cumsum(pca$sdev*2)/sum(pca$sdev^2) >= 0.9)[1]))
print(str_c("ENT90% + bonferroni corrected p-value: ", 
            round((0.05 / which(cumsum(pca$sdev*2)/sum(pca$sdev^2) >= 0.9)[1]), 6)))
print(str_c("ENT90% + bonferroni corrected -log10(p-value): ", 
            round(-log10(0.05 / which(cumsum(pca$sdev*2)/sum(pca$sdev^2) >= 0.9)[1]), 3)))

```

# 2. logistic regression

```{r}

logi_res_fnc <- function(mets_list, outc, dat, mets_info, add_covar = "", 
                             first_cols = c("met_ID", "SUPER.PATHWAY", "SUB.PATHWAY", "BIOCHEMICAL",
                                            "beta", "pval", "or",
                                            "lowerCI", "upperCI")) {
  
        logi_mdl_tmp <- function(met) {

                eval(parse(text = str_c("fit <- glm(", outc, " ~ ", met, 
                                        add_covar, ", data = ", dat, ", family = 'binomial')")))
                
                coef <- summary(fit)$coef %>% as.data.frame() %>% rownames_to_column() %>% as.data.table()
                coef <- coef[rowname == met, ]
                
                conf <- confint.default(fit) %>% as.data.frame() %>% rownames_to_column() %>% as.data.table()
                conf <- conf[rowname == met, -c("rowname")]
                
                output <- cbind("Asthma", coef,conf)

                setnames(output, c("outc", "met_ID", "beta", "se", "zval", "pval","lowerCI","upperCI"))
                output[, (colnames(output)[-c(1:2)]) := lapply(.SD, as.numeric), .SDcols = colnames(output)[-c(1:2)]]
                
                return(output)
        }
        
        res <- lapply(mets_list, function(x) {logi_mdl_tmp(met = x)}) %>% rbindlist()
        res <- merge(res, metabolite_summary2, by = "met_ID", all.x = T)

        res[, ':='(neg_log10pval = -log10(pval),
                   fdr_bh = p.adjust(pval, method = "BH"),
                   or = exp(beta),
                   lower95 = exp(beta - z_95 * se),
                   upper95 = exp(beta + z_95 * se))]

        setcolorder(res, first_cols)
        return(res)
}

```

Asthma ~ met + Age + Gender + Race + BMI + Smoking

610 subjects

## 2.1 single met
```{r}
# Sex + RACE_cat + Age + BMI + Smoking | ASTHMA

asthma_logi <- logi_res_fnc(outc = "ASTHMA", mets_list = mets_list, add_covar = "+ Sex + Race_3cat + Age + BMI + Smoking", dat = "final_data")

asthma_logi[pval < 0.05][order(pval)] %>% datatable(filter = "top") %>% formatRound(colnames(asthma_logi)[colnames(asthma_logi) %in% c("beta", "pval", "fdr_bh", "or", "se", "zval", "neg_log10pval","lowerCI","upperCI")], digits=5)

asthma_logi[pval < 0.000382][order(pval)] %>% datatable(filter = "top") %>% formatRound(colnames(asthma_logi)[colnames(asthma_logi) %in% c("beta", "pval", "fdr_bh", "or", "se", "zval", "neg_log10pval","lowerCI","upperCI")], digits=5)

```

## 2.2 summary the significant results
significant mets identified in logistic regression:
0.05: global 195

```{r}
sig_mets_df <- asthma_logi[pval < 0.05,]
dim(sig_mets_df)[1] #195

sig_mets_df$SUPER.PATHWAY[sig_mets_df$SUPER.PATHWAY == ""] <- "Unknown"

sig_global_mets_df <- sig_mets_df[sig_mets_df$SUPER.PATHWAY != "Xenobiotics" & sig_mets_df$SUPER.PATHWAY != "Unknown",]
dim(sig_global_mets_df)[1] #112

sig_xenobiotics_mets_df <- sig_mets_df[sig_mets_df$SUPER.PATHWAY == "Xenobiotics",]
dim(sig_xenobiotics_mets_df)[1] #19

sig_unknown_mets_df <- sig_mets_df[sig_mets_df$SUPER.PATHWAY == "Unknown",]
dim(sig_unknown_mets_df)[1] #64

```

# 4. Create the significant mets figure
## 4.1 Plot MGBB
```{r}
sig_global_mets_df$or <- log(sig_global_mets_df$or)

# plot all
sig_global_mets_df$SUPER.PATHWAY[sig_global_mets_df$SUPER.PATHWAY == "Cofactors and Vitamins"] <- "Cofactors & Vitamins"

sig_mets_gplot_1 <- ggplot(sig_global_mets_df, aes(reorder(BIOCHEMICAL, -or), y = or, color = or >0)) +
   scale_colour_manual(name = '', values = setNames(c('red','blue'),c(T, F)))+
   geom_rect(aes(fill = SUB.PATHWAY),xmin = -Inf,xmax = Inf,
             ymin = -Inf,ymax = Inf,alpha = 0.3, linejoin = "round") +
   facet_grid(SUB.PATHWAY ~ factor(SUPER.PATHWAY, levels = c('Amino Acid', 'Lipid', 'Carbohydrate', 'Cofactors & Vitamins', 'Nucleotide', 'Peptide')), scales = "free_y", space = "free_y", labeller = label_wrap_gen()) +
   geom_errorbar(aes(ymin = log(exp(lowerCI)), ymax = log(exp(upperCI))), width = 0.25) +
   geom_pointrange(aes(ymin = log(exp(lowerCI)), ymax = log(exp(upperCI))), size = 0.25) +
   geom_hline(yintercept = 0, linetype = 2) +
   labs(
     x = "Metabolite",
     y = "log(OR) (95% Confidence Interval)"
   ) +
   theme_minimal() +
   theme_bw() +
   theme(
     title = element_text(size = 16),
     strip.text.x = element_text(size = 12),
     strip.text.y = element_text(size = 10),
     axis.title = element_text(size = 12),
     axis.text = element_text(size = 9),
     legend.position = "",
     legend.title = element_text(size = 12),
     legend.text = element_text(size = 12),
    strip.text.y.right = element_text(angle = 0)) +
   coord_flip() + scale_fill_viridis_d(option = "D", direction = -1, begin = 0, end = 1, alpha = 0.1)

ggsave(sig_mets_gplot_1, 
       filename = here(fig_dir, "sig_mets_summary_gplot_all.png.png"),
       width = 18,
       height = 25, dpi = 300)


# plot need
## remove Cofactors & Vitamins, Nucleotide, and Peptide
table(sig_global_mets_df$SUPER.PATHWAY)
sig_global_mets_df_updated <- sig_global_mets_df[sig_global_mets_df$SUPER.PATHWAY != "Cofactors & Vitamins" & sig_global_mets_df$SUPER.PATHWAY != "Nucleotide" & sig_global_mets_df$SUPER.PATHWAY != "Peptide",]

## remove no-overlap sub-pathway
table(sig_global_mets_df_updated$SUB.PATHWAY)
sig_global_mets_df_updated <- sig_global_mets_df_updated[sig_global_mets_df_updated$SUB.PATHWAY != "Alanine and Aspartate Metabolism" & sig_global_mets_df_updated$SUB.PATHWAY != "Endocannabinoid" & sig_global_mets_df_updated$SUB.PATHWAY != "Fatty Acid Metabolism(Acyl Carnitine)" &
sig_global_mets_df_updated$SUB.PATHWAY != "Fatty Acid, Dicarboxylate" & sig_global_mets_df_updated$SUB.PATHWAY != "Fatty Acid, Dihydroxy" & 
sig_global_mets_df_updated$SUB.PATHWAY != "Fatty Acid, Monohydroxy" &
sig_global_mets_df_updated$SUB.PATHWAY != "Glutamate Metabolism" &
sig_global_mets_df_updated$SUB.PATHWAY != "Glutathione Metabolism" &
sig_global_mets_df_updated$SUB.PATHWAY != "Glycolysis, Gluconeogenesis, and Pyruvate Metabolism" &
sig_global_mets_df_updated$SUB.PATHWAY != "Guanidino and Acetamido Metabolism" &
sig_global_mets_df_updated$SUB.PATHWAY != "Histidine Metabolism" &
sig_global_mets_df_updated$SUB.PATHWAY != "Leucine, Isoleucine and Valine Metabolism" &
sig_global_mets_df_updated$SUB.PATHWAY != "Lysine Metabolism" &
sig_global_mets_df_updated$SUB.PATHWAY != "Pentose Metabolism" &
sig_global_mets_df_updated$SUB.PATHWAY != "Phenylalanine Metabolism" &
sig_global_mets_df_updated$SUB.PATHWAY != "Polyamine Metabolism" &
sig_global_mets_df_updated$SUB.PATHWAY != "Primary Bile Acid Metabolism" &
sig_global_mets_df_updated$SUB.PATHWAY != "Secondary Bile Acid Metabolism" &
sig_global_mets_df_updated$SUB.PATHWAY != "Tyrosine Metabolism" &
sig_global_mets_df_updated$SUB.PATHWAY != "Urea cycle; Arginine and Proline Metabolism",]

## combined some subpathway together
table(sig_global_mets_df_updated$SUB.PATHWAY)
sig_global_mets_df_updated$SUB.PATHWAY.updated <- sig_global_mets_df_updated$SUB.PATHWAY

sig_global_mets_df_updated$SUB.PATHWAY.updated[sig_global_mets_df_updated$SUB.PATHWAY.updated == "Androgenic Steroids" | sig_global_mets_df_updated$SUB.PATHWAY.updated == "Corticosteroids" | sig_global_mets_df_updated$SUB.PATHWAY.updated == "Pregnenolone Steroids" | sig_global_mets_df_updated$SUB.PATHWAY.updated == "Progestin Steroids" |
sig_global_mets_df_updated$SUB.PATHWAY.updated == "Sterol"] <- "Steroid-related metabolites"

table(sig_global_mets_df_updated$SUB.PATHWAY.updated)

sig_mets_gplot_2 <- ggplot(sig_global_mets_df_updated, 
                           aes(reorder(BIOCHEMICAL, -or), y = or, color = or >0)) +
   scale_colour_manual(name = '', values = setNames(c('red','blue'),c(T, F)))+
   geom_rect(aes(),xmin = -Inf,xmax = Inf,
             ymin = -Inf,ymax = Inf, alpha = 0, color = 'black', linejoin = "round") +
   facet_grid(SUB.PATHWAY.updated ~ factor(SUPER.PATHWAY, levels = c('Amino Acid', 'Lipid', 'Carbohydrate')), scales = "free_y", space = "free_y", labeller = label_wrap_gen()) +
   geom_errorbar(aes(ymin = log(lower95), ymax = log(upper95)), width = 0.25) +
   geom_pointrange(aes(ymin = log(lower95), ymax = log(upper95)), size = 0.25) +
   geom_hline(yintercept = 0, linetype = 2) +
   labs(
     x = "Metabolite",
     y = "log(OR) (95% Confidence Interval)"
   ) +
   theme_minimal() +
   theme_bw() +
   theme(
     title = element_text(size = 16),
     strip.text.x = element_text(size = 12),
     strip.text.y = element_text(size = 10),
     axis.title = element_text(size = 12),
     axis.text = element_text(size = 9),
     legend.position = "",
     legend.title = element_text(size = 12),
     legend.text = element_text(size = 12),
    strip.text.y.right = element_text(angle = 0)) +
   coord_flip() + scale_fill_viridis_d(option = "D", direction = -1, begin = 0, end = 1, alpha = 0.1) +
   scale_y_continuous(limits = c(-2, 2))

ggsave(sig_mets_gplot_2, 
       filename = here(fig_dir, "sig_mets_summary_gplot_overlap.png"),
       width = 14,
       height = 16, dpi = 300)
```



## 4.2 Plot PEGASUS and MGBB
```{r}
# plot need for PEGASUS and MGBB
table(sig_global_mets_df_updated$SUB.PATHWAY.updated)

# add PEGASUS data
PEGASUS_sig_mets_df_overlap_df <- fread(str_c("/udd/reyul/mets_ratio_CodeReview/results/3_PEGASUS_global_mets_association/", "PEGASUS_sig_mets_df_overlap.csv"))

PEGASUS_sig_mets_df_overlap_df <- PEGASUS_sig_mets_df_overlap_df[,c("feat_id", "annotation", "tentative_superpathway",
                                                                    "tentative_subpathway_updated", "or", 
                                                                    "lower95", "upper95")]
colnames(PEGASUS_sig_mets_df_overlap_df) <- c("feat_id", "annotation", 
                                              "superpathway", "subpathway", "or", "lower95", "upper95")

PEGASUS_sig_mets_df_overlap_df$cohort <- "PEGASUS"
MGBB_sig_mets_df_overlap_df <- sig_global_mets_df_updated[,c("met_ID", "BIOCHEMICAL", 
                                                             "SUPER.PATHWAY", "SUB.PATHWAY.updated",
                                                             "or", "lower95", "upper95")]
colnames(MGBB_sig_mets_df_overlap_df) <- c("feat_id", "annotation", 
                                           "superpathway", "subpathway", "or", "lower95", "upper95")
MGBB_sig_mets_df_overlap_df$cohort <- "MGBB"

MGBB_sig_mets_df_overlap_df$lower95 <- log(MGBB_sig_mets_df_overlap_df$lower95)
MGBB_sig_mets_df_overlap_df$upper95 <- log(MGBB_sig_mets_df_overlap_df$upper95)

PEGASUS_MGBB_sig_mets_overlap_df <- rbind(PEGASUS_sig_mets_df_overlap_df, MGBB_sig_mets_df_overlap_df)

PEGASUS_MGBB_sig_mets_overlap_df$cohort <- factor(PEGASUS_MGBB_sig_mets_overlap_df$cohort, levels = c("PEGASUS", "MGBB"))

sig_mets_gplot_3 <- ggplot(PEGASUS_MGBB_sig_mets_overlap_df, 
                           aes(reorder(annotation, -or), y = or, color = or >0)) +
   scale_colour_manual(name = '', values = setNames(c('red','blue'),c(T, F)))+
   geom_rect(aes(),xmin = -Inf,xmax = Inf,
             ymin = -Inf,ymax = Inf, alpha = 0, color = 'black', linejoin = "round") +
   facet_grid(subpathway*cohort ~ factor(superpathway, levels = c('Amino Acid', 'Lipid', 'Carbohydrate')), scales = "free_y", space = "free_y", labeller = label_wrap_gen()) +
   geom_errorbar(aes(ymin = lower95, ymax = upper95), width = 0.25) +
   geom_pointrange(aes(ymin = lower95, ymax = upper95), size = 0.25) +
   geom_hline(yintercept = 0, linetype = 2) +
   labs(
     x = "Metabolite",
     y = "log(OR) (95% Confidence Interval)"
   ) +
   theme_minimal() +
   theme_bw() +
   theme(
     title = element_text(size = 16),
     strip.text.x = element_text(size = 12),
     strip.text.y = element_text(size = 10),
     axis.title = element_text(size = 12),
     axis.text = element_text(size = 9),
     legend.position = "",
     legend.title = element_text(size = 12),
     legend.text = element_text(size = 12),
    strip.text.y.right = element_text(angle = 0)) +
   coord_flip() + scale_fill_viridis_d(option = "D", direction = -1, begin = 0, end = 1, alpha = 0.1)

ggsave(sig_mets_gplot_3, 
       filename = here(fig_dir, "sig_mets_summary_gplot_overlap_PEGASUS_MGBB.png"),
       width = 16,
       height = 20, dpi = 300)
```

# 5. Save data
```{r}
save(asthma_logi,  
     file = str_c(res_dir, "MGBB_all_association_asthma_results.RData"))

write.csv(sig_global_mets_df, file = str_c(res_dir, "MGBB_sig_global_mets_df.csv"), row.names = FALSE)
write.csv(sig_xenobiotics_mets_df, file = str_c(res_dir, "MGBB_sig_global_xeno_df.csv"), row.names = FALSE)
write.csv(sig_unknown_mets_df, file = str_c(res_dir, "MGBB_sig_global_unknown_df.csv"), row.names = FALSE)

```

# Session info
```{r}
sessionInfo()

```
