==> 5_MGB_KAS_targeted_mets_association_phenotypes.Rmd <==
---
title: "Regression analysis of targeted metabolites in PEGASUS"
subtitle: "asthma phenotypes -- ige, fev1, fvc, fev1.fvc"
author: "Yulu Chen"
date: "10/24/2023"
output: 
  html_document: 
    toc: true
    toc_float: 
      collapsed: false
      smooth_scroll: false
    code_folding: hide
editor_options: 
  chunk_output_type: console
---

```{r, include = FALSE}
##### 1. Setup #####
## Packages

pkg <- c("tidyverse", "data.table", "here", "survival", "gtools", "tableone", "DT", "sjPlot", 
         "sjmisc", "grid", "gridExtra", "ggpubr", "table1", "readxl", "ggvenn")
for (p in pkg) {
        if (require(p, character.only = T)) {
                print(paste0(p, " loaded successfully"))
        } else {
                install.packages(p)
                require(p, character.only = T)
                print(paste0(p, " downloaded and loaded successfully"))
        }
}

## Paths
pegasus_dir <- "/udd/reyul/mets_ratio_CodeReview/data/PEGASUS/"

dat_dir <- str_c(pegasus_dir, "2_assembly_global_targeted/")
fig_dir <- str_c("/udd/reyul/mets_ratio_CodeReview/", "figures/5_PEGASUS_targeted_mets_association/")
res_dir <- str_c("/udd/reyul/mets_ratio_CodeReview/", "results/5_PEGASUS_targeted_mets_association/")

## Hard-coded numbers

missing_thld <- 0.30
alpha_thld <- 0.05
z_95 <- qnorm(p = 0.975)

## Functions
pvalue <- function(x, ...) {
    # Construct vectors of data y, and groups (strata) g
    y <- unlist(x)
    g <- factor(rep(1:length(x), times=sapply(x, length)))
    if (is.numeric(y)) {
        # For numeric variables, perform a standard 2-sample t-test
        p <- as.data.frame(summary(aov(y ~ g))[[1]])[1,5]
    } else {
        # For categorical variables, perform a chi-squared test of independence
        p <- chisq.test(table(y, g))$p.value
    }
    # Format the p-value, using an HTML entity for the less-than sign.
    # The initial empty string places the output on the line below the variable label.
    c("", sub("<", "&lt;", format.pval(p, digits=6, eps=0.000001)))
}

```


# 1. Load data
Using the data generated by Yulu and Mengna, including the phenotype/EMR data, global metabolomics data (HILIC+, Lipid+, and Lipid-), targeted sphingolipids data, targeted microbial data

global metabolomics data:
- remove metatolites with sample CV >= 25%
- remove metabolites with >= 30% missing in mets
- impute with half min
- log10 transformation & pareto-scaling

targeted sphingolipids:  remove metabolites with >= 30% missing in mets & Standardize

targeted microbial: remove metabolites with >= 30% missing in mets & Standardize

targeted steroids: remove metabolites with >= 30% missing in mets & Standardize

```{r}

load(here(dat_dir, "all_platforms_data_updated.RData"), verbose = T)

targeted_sph_list <- sph_mets_info[pct_na_met <= 0.30, feat_id]
targeted_microb_list <- microb_mets_info[pct_na_met <= 0.30, feat_id]
targeted_str_list <- str_mets_info[pct_na_met <= 0.30, feat_id]

sph_mets_info$`SUPER PATHWAY` <- "sphingolipid metabolism"
sph_mets_info[, `SUB PATHWAY` := case_when(grepl("^HexCer|LacCer", feat_id) ~ "Hex/LacCeramide", 
                                    grepl("^SM", feat_id) ~ "Sphingomyelin", 
                                    grepl("^Sph|Sphd|S1P", feat_id) ~ "Sphingosine/S1P", 
                                    grepl("^Cer1P", feat_id) ~ "Ceramide-1-phosphate", 
                                    grepl("^Cer[1-2][0-9]", feat_id) ~ "Ceramide", 
                                    grepl("^DhCer", feat_id) ~ "Dihydroceramide", 
                                    grepl("^Spa|Spa1P", feat_id) ~ "Sphinganine/Spa1P", 
                                    feat_id == "Glucosph" ~ "Hexosylsphingosine", 
                                    feat_id == "X3_KS" ~ "X3_Ketosphinganine")]

sph_mets_info$tentative_annotation <- c("Cer(d18:1/10:0)", "Cer(d18:1/12:0)", "Cer(d18:1/14:0)", 
                                        "Cer(d16:1/16:0)", "Cer(d18:1/16:0)", "Cer(d18:2/16:0)", 
                                        "Cer(d18:1/17:0)", "Cer(d16:1/18:0)", "Cer(d18:1/18:0)", 
                                        "Cer(d18:2/18:0)", "Cer(d18:1/18:1)", "Cer-1P(d18:1/16:0)", 
                                        "Cer-1P(d18:1/24:0)", "Cer-1P(d18:1/24:1)", "Cer(d18:1/20:0)",
                                        "Cer(d18:1/20:1)", "Cer(d18:1/22:0)", "Cer(d16:1/24:0)", 
                                        "Cer(d18:1/24:0)", "Cer(d18:2/24:0)", "Cer(d16:1/24:1)", 
                                        "Cer(d18:1/24:1)", "Cer(d18:2/24:1)", "Cer(d18:1/26:0)",
                                        "Cer(d18:1/26:1)", "DhCer(d18:0/16:0)", "DhCer(d18:0/18:0)",
                                        "DhCer(d18:0/24:0)", "DhCer(d18:0/24:1)", "Glucosph", 
                                        "HexCer(d18:1/12:0)", "HexCer(d18:1/14:0)", "HexCer(d18:1/16:0)", 
                                        "HexCer(d18:1/18:0)", "HexCer(d18:1/18:1)", "HexCer(d18:1/20:0)",
                                        "HexCer(d18:1/22:0)", "HexCer(d18:1/24:0)", "HexCer(d18:1/24:1)",
                                        "LacCer(d18:1/12:0)", "LacCer(d18:1/14:0)", "LacCer(d18:1/16:0)",
                                        "LacCer(d18:1/17:0)", "LacCer(d18:1/18:0)", "LacCer(d18:1/18:1)", 
                                        "LacCer(d18:1/20:0)", "LacCer(d18:1/22:0)", "LacCer(d18:1/24:0)",
                                        "LacCer(d18:1/24:1)", "Sphingosine-1P(d16:1)", 
                                        "Sphingosine-1P(d17:1)", "Sphingosine-1P", "Sphingosine-1P(d18:2)", 
                                        "SM(d18:1/12:0)",  "SM(d18:1/14:0)", "SM(d16:1/16:0)", 
                                        "SM(d18:1/16:0)", "SM(d18:2/16:0)", "SM(d18:1/17:0)", 
                                        "SM(d16:1/18:0)", "SM(d18:1/18:0)", "SM(d18:2/18:0)", 
                                        "SM(d18:1/18:1)", "SM(d18:1/20:0)", "SM(d18:2/22:0)",
                                        "SM(d18:1/22:0)", "SM(d16:1/24:0)", "SM(d18:1/24:0)", 
                                        "SM(d18:2/24:0)", "SM(d16:1/24:1)",  "SM(d18:1/24:1)", 
                                        "SM(d18:2/24:1)", "Sphinganine-1P", "Sphinganine", 
                                        "Sphingosine(d16:1)", "Sphingosine", "Sphingosine(d18:2)",
                                        "3-Dehydrosphinganine")
```

# 2 Trim data
- _trim: subjects missing bmi & matched subjects removed  

```{r}
nrow(pheno_mets) #1132

stratum_to_rm <- pheno_mets[is.na(bmi), unique(Stratum)]
stratum_to_rm <- c(stratum_to_rm, pheno_mets[, .N, Stratum][N == 1, Stratum])
print(str_c("Number of strata removed due to missing BMI: ", 
            length(stratum_to_rm)))
pheno_mets_trim <- pheno_mets[!(Stratum %in% stratum_to_rm), ]

```

## 2.1 heatmap for steroids
```{r}
# ratio with missing percentatge <= 30%
list_analysis <- str_mets_info[pct_na_met <= 0.3, ]$feat_id

length(list_analysis) #16

# get the ratio data
ratio_column <- c("Biobank_Subject_ID", list_analysis)
str_feat_data = pheno_mets_trim[, ..ratio_column] # n = 540
str_feat_data <- str_feat_data[complete_cases(str_feat_data),] # n = 898
str_subject_ID <- str_feat_data$Biobank_Subject_ID

# Standardize data for clustering
str_feat_data[, (list_analysis) := lapply(.SD, function(x) scale(x, center = T, scale = T)), .SDcols = list_analysis]

clust_df <- as.data.frame(str_feat_data)
rownames(clust_df) <- clust_df$Biobank_Subject_ID
clust_df <- clust_df[,-1]
clust_df <- transpose(clust_df)

colnames(clust_df) <- str_subject_ID

str_name <- str_mets_info[feat_id %in% list_analysis,`Component Name`]

rownames(clust_df) <- str_name

## heatmap
str_cor_df <- as.data.frame(t(clust_df))
str_cor_df <- str_cor_df %>% 
                as.matrix %>%
                cor

# Perform hierarchical (**thus NOT kmeans; see comment below**) clustering 'outside' ComplexHeatmap using hclust

hr <- hclust(dist(str_cor_df, method = "euclidean"), method="ward.D2") 
hc <- hclust(dist(t(str_cor_df), method = "euclidean"), method="ward.D2")


png(file=str_c(fig_dir, "str_cor_pheatmap.png"), width = 5, height = 7, units = "in", res = 300)
pheatmap(str_cor_df, scale = "none", show_rownames=F, show_colnames=T,
	cluster_rows=hr, cluster_cols=hc,
	main = "Steroid correlation",
	treeheight_row = 0)
dev.off()


```


## 2.2 heatmap for sphingolipids
```{r}
# ratio with missing percentatge <= 30%
list_analysis <- sph_mets_info[pct_na_met <= 0.3, ]$feat_id

length(list_analysis) #78

# get the ratio data
ratio_column <- c("Biobank_Subject_ID", list_analysis)
sph_feat_data = pheno_mets_trim[, ..ratio_column] # n = 540
sph_feat_data <- sph_feat_data[complete_cases(sph_feat_data),] # n = 975
sph_subject_ID <- sph_feat_data$Biobank_Subject_ID

# Standardize data for clustering
sph_feat_data[, (list_analysis) := lapply(.SD, function(x) scale(x, center = T, scale = T)), .SDcols = list_analysis]

clust_df <- as.data.frame(sph_feat_data)
rownames(clust_df) <- clust_df$Biobank_Subject_ID
clust_df <- clust_df[,-1]
clust_df <- transpose(clust_df)

colnames(clust_df) <- sph_subject_ID

sph_name <- sph_mets_info[feat_id %in% list_analysis,tentative_annotation]

rownames(clust_df) <- sph_name

## heatmap
sph_cor_df <- as.data.frame(t(clust_df))
sph_cor_df <- sph_cor_df %>% 
                as.matrix %>%
                cor

# Perform hierarchical (**thus NOT kmeans; see comment below**) clustering 'outside' ComplexHeatmap using hclust

hr <- hclust(dist(sph_cor_df, method = "euclidean"), method="ward.D2") 
hc <- hclust(dist(t(sph_cor_df), method = "euclidean"), method="ward.D2")


png(file=str_c(fig_dir, "sph_cor_pheatmap.png"), width = 15, height = 16.5, units = "in", res = 300)
pheatmap(sph_cor_df, scale = "none", show_rownames=F, show_colnames=T,
	cluster_rows=hr, cluster_cols=hc,
	main = "Sphingolipid correlation",
	treeheight_row = 0)
dev.off()


```

# 3. Regression model function
## 3.1 Conditional Logistic Regression model functions
```{r}
clogi_res_fnc <- function(outc, mets_list, dat, mets_info, add_covar = "", 
                          first_cols = c("feat_id", "tentative_annotation", "beta", "pval", "fdr_bh", "or", "lower95", "upper95", 
                                         "med_feat", "wo_ms2", "rt_min", "mz", "feature", "platform", 
                                         "totaln", "casen", "cv", "pct_na_met"), 
                          na_thld = missing_thld) {
        
        clogi_mdl_tmp <- function(met) {
                
                fit <- eval(parse(text = str_c("clogit(", outc, " ~ ", met, " + strata(Stratum)", 
                                               add_covar, ", data = ", dat, ")")))
                
                coef <- summary(fit)$coefficients
                coef <- coef[rownames(coef) == met, ]

                totaln <- summary(fit)$n
                casen <- summary(fit)$nevent
                
                output <- t(c(outc, met, coef, totaln, casen)) %>% as.data.table()
                colnames(output) <- c("outc", "feat_id", "beta", "or", "se", "zval", "pval", 
                                      "totaln", "casen")
                output[, (colnames(output)[-c(1:2)]) := lapply(.SD, as.numeric), .SDcols = colnames(output)[-c(1:2)]]
                
                return(output)
        }
        
        res <- lapply(mets_list, function(x) {clogi_mdl_tmp(met = x)}) %>% rbindlist()
        
        eval(parse(text = str_c("mets_info <- copy(", mets_info, ")")))
        res <- merge(res, mets_info, by = "feat_id")
        # res[pct_na_met <= na_thld, fdr_bh := p.adjust(pval, method = "BH")] 
        res[, ':='(lower95 = exp(beta - z_95 * se), 
                   upper95 = exp(beta + z_95 * se))] 
        res[, ':='(fdr = p.adjust(pval, method = "fdr"), 
                   fdr_bh = p.adjust(pval, method = "BH"))]
        setcolorder(res, first_cols)
        res <- res[order(pval)]
        return(res)
}

```

## 3.2 Linear regression model functions
```{r}
linreg_res_fnc <- function(outc, mets_list, dat, mets_info, add_covar = "", 
                           first_cols = c("feat_id", "outc", "beta", "pval", "lower95", "upper95", "fdr", "fdr_bh"), 
                           na_thld = missing_thld) {
        
        linear_mdl_tmp <- function(met) {
                
                fit <- eval(parse(text = str_c("lm(", outc, " ~ ", met, 
                                               add_covar, ", data = ", dat, ")")))
                
                coef <- summary(fit)$coefficients %>% as.data.frame() %>% rownames_to_column() %>% as.data.table()
                coef <- coef[rowname == met, ]

                output <- cbind(outc, coef)
                colnames(output) <- c("outc", "feat_id", "beta", "se", "tval", "pval")
                output[, (colnames(output)[-c(1:2)]) := lapply(.SD, as.numeric), .SDcols = colnames(output)[-c(1:2)]]
                
                return(output)
        }
        
        res <- lapply(mets_list, function(x) {linear_mdl_tmp(met = x)}) %>% rbindlist()
        
        # eval(parse(text = str_c("mets_info <- copy(", mets_info, ")")))
        # res <- merge(res, mets_info, by = "feat_id")
        res[, ':='(fdr = p.adjust(pval, method = "fdr"), 
                   fdr_bh = p.adjust(pval, method = "BH"))]
        res[, totaln := eval(parse(text = str_c(dat, "[!is.na(", outc, "), .N]")))]
        res[, ':='(lower95 = beta - z_95 * se, 
                   upper95 = beta + z_95 * se)]
        setcolorder(res, first_cols)
        res <- res[order(pval)]
        return(res)
}

```

## 3.3 Poisson regression
```{r}
poiss_res_fnc <- function(outc, mets_list, dat, add_covar = "",
                         first_cols = c("feat_id", 
                                        "outc", "beta", "pval", "fdr")) {
        
        logi_mdl_tmp <- function(met) {
                
                fit <- eval(parse(text = str_c("glm(", outc, " ~ ", met, add_covar, ", data = ", dat, ", family = 'quasipoisson')")))
                
                coef <- summary(fit)$coefficients %>% as.data.frame() %>% rownames_to_column() %>% as.data.table()
                coef <- coef[rowname == met, ]

                output <- cbind(outc, coef)
                colnames(output) <- c("outc", "feat_id", "beta", "se", "zval", "pval")
                output[, (colnames(output)[-c(1:2)]) := lapply(.SD, as.numeric), .SDcols = colnames(output)[-c(1:2)]]
                
                return(output)
        }
        
        res <- lapply(mets_list, function(x) {logi_mdl_tmp(met = x)}) %>% rbindlist()
        
        # eval(parse(text = str_c("mets_info <- copy(", mets_info, ")")))
        # res <- merge(res, mets_info, by = "feat_id", all.x = T)
        res[, ':='(neg_log10_pval = -log10(pval), 
                   fdr = p.adjust(pval, method = "fdr"),
                   fdr_bh = p.adjust(pval, method = "BH"))]
        res[, ':='(lower95 = beta - z_95 * se, 
                   upper95 = beta + z_95 * se)]
        res[, totaln := eval(parse(text = str_c(dat, "[!is.na(", outc, "), .N]")))]
        
        setcolorder(res, first_cols)
        res <- res[order(pval)]
        return(res)
}



```

# 4. Conduct association analysis within Asthma Status
```{r}
dat_tmp <- pheno_mets_trim[Asthma_Status == TRUE,]

```

## 4.1 Log(IgE)
```{r}
dat_tmp$logige <- log(dat_tmp$ige)
sum(is.na(dat_tmp$logige)) #363
ggplot(dat_tmp, aes(x=logige)) + geom_histogram(color="darkblue", fill="lightblue")

```

### 4.1.1 targeted sph
```{r}
length(targeted_sph_list) #78

res_4.1.1 <- linreg_res_fnc(outc = "logige", 
                            mets_list = targeted_sph_list, 
                            dat = "dat_tmp", 
                            add_covar = " + Age_at_Collection + bmi + Gender + race + Ethnicity + ics_trim_totnum_5y")

res_4.1.1 <- merge(res_4.1.1, sph_mets_info, by = "feat_id", all.x = T)

## display results
datatable(res_4.1.1[pval <= 0.05, c("feat_id", "metabolite", "outc", "pval", "fdr", "beta", "lower95", "upper95", "SUB PATHWAY")], 
          filter = "top") %>% 
          formatSignif(columns = c("pval", "beta", "lower95", "upper95"), digits = 3)

res_4.1.1_sig <- res_4.1.1[pval <= 0.05,]
res_4.1.1_sig$Pval <- signif(res_4.1.1_sig$pval, 3)
res_4.1.1_sig <- res_4.1.1_sig[order(`SUB PATHWAY`, beta), ]

p_sph_log_ige <- ggplot(res_4.1.1_sig, 
                aes(reorder(tentative_annotation, -beta), y = beta, color = beta >0)) +
   scale_colour_manual(name = '', values = setNames(c('red','blue'),c(T, F)))+
   geom_rect(aes(fill = `SUB PATHWAY`),xmin = -Inf,xmax = Inf,
             ymin = -Inf,ymax = Inf, alpha = 0, color = 'black', linejoin = "round") +
   facet_grid(`SUB PATHWAY` ~ `SUPER PATHWAY`, scales = "free_y", space = "free_y", labeller = label_wrap_gen()) +
   geom_errorbar(aes(ymin = lower95, ymax = upper95), width = 0.25) +
   geom_pointrange(aes(ymin = lower95, ymax = upper95), size = 0.25) +
   geom_hline(yintercept = 0, linetype = 2) +
   labs(
     x = "",
     y = ""
   ) +
   theme_minimal() +
   theme_bw() +
   theme(
     title = element_text(size = 16),
     strip.text.x = element_text(size = 12, face = "bold"),
     strip.text.y = element_text(size = 10, face = "bold"),
     axis.title = element_text(size = 12),
     axis.text = element_text(size = 10, face = "bold"),
     axis.text.x = element_text(margin = margin(t = 10)),  # Adjust top margin of x-axis labels
     axis.text.y = element_text(margin = margin(r = 10)),   # Adjust right margin of y-axis labels
     legend.position = "",
     legend.title = element_text(size = 12),
     legend.text = element_text(size = 12),
     strip.text.y.right = element_text(angle = 0),
     strip.background = element_rect(fill = "mistyrose", colour = "black")) +
   coord_flip() + scale_fill_viridis_d(option = "D", direction = -1, begin = 0, end = 1, alpha = 0.1)
   # scale_y_continuous(limits = c(-1.5, 1.6))

```


### 4.1.2 targeted microb
```{r}
length(targeted_microb_list) #51

res_4.1.2 <- linreg_res_fnc(outc = "logige", 
                            mets_list = targeted_microb_list, 
                            dat = "dat_tmp", 
                            add_covar = " + Age_at_Collection + bmi + Gender + race + Ethnicity + ics_trim_totnum_5y")

res_4.1.2 <- merge(res_4.1.2, microb_mets_info, by = "feat_id", all.x = T)

## display results
datatable(res_4.1.2[pval <= 0.05, c("feat_id", "Component_Name", "outc", "pval", "fdr", "beta", "lower95", "upper95", "SUB_PATHWAY")], 
          filter = "top") %>% 
          formatSignif(columns = c("pval", "beta", "lower95", "upper95"), digits = 3)

res_4.1.2_sig <- res_4.1.2[pval <= 0.05,]
res_4.1.2_sig$Pval <- signif(res_4.1.2_sig$pval, 3)
res_4.1.2_sig <- res_4.1.2_sig[order(`SUB_PATHWAY`, beta), ]

p_microb_log_ige <- ggplot(res_4.1.2_sig, 
                   aes(reorder(tentative_annotation, -beta), y = beta, color = beta >0)) +
   scale_colour_manual(name = '', values = setNames(c('red','blue'),c(T, F)))+
   geom_rect(aes(fill = `SUB_PATHWAY`),xmin = -Inf,xmax = Inf,
             ymin = -Inf,ymax = Inf, alpha = 0, color = 'black', linejoin = "round") +
   facet_grid(`SUB_PATHWAY` ~ `SUPER_PATHWAY`, scales = "free_y", space = "free_y", labeller = label_wrap_gen()) +
   geom_errorbar(aes(ymin = lower95, ymax = upper95), width = 0.25) +
   geom_pointrange(aes(ymin = lower95, ymax = upper95), size = 0.25) +
   geom_hline(yintercept = 0, linetype = 2) +
   labs(
     x = "",
     y = ""
   ) +
   theme_minimal() +
   theme_bw() +
   theme(
     title = element_text(size = 16),
     strip.text.x = element_text(size = 12, face = "bold"),
     strip.text.y = element_text(size = 10, face = "bold"),
     axis.title = element_text(size = 12),
     axis.text = element_text(size = 10, face = "bold"),
     legend.position = "",
     legend.title = element_text(size = 12),
     legend.text = element_text(size = 12),
     strip.text.y.right = element_text(angle = 0),
     strip.background = element_rect(fill = "lightcyan", colour = "black")) +
   coord_flip() + scale_fill_viridis_d(option = "D", direction = -1, begin = 0, end = 1, alpha = 0.1)
   # scale_y_continuous(limits = c(-1.6, 4.1))


```


### 4.1.3 targeted str
```{r}
length(targeted_str_list) #16

res_4.1.3 <- linreg_res_fnc(outc = "logige", 
                            mets_list = targeted_str_list, 
                            dat = "dat_tmp", 
                            add_covar = " + Age_at_Collection + bmi + Gender + race + Ethnicity + ics_trim_totnum_5y")

res_4.1.3 <- merge(res_4.1.3, str_mets_info, by = "feat_id", all.x = T)

## display results
datatable(res_4.1.3[pval <= 0.05, c("feat_id", "Component Name", "outc", "pval", "fdr", "beta", "lower95", "upper95", "SUB PATHWAY")], 
          filter = "top") %>% 
          formatSignif(columns = c("pval", "beta", "lower95", "upper95"), digits = 3)

res_4.1.3_sig <- res_4.1.3[pval <= 0.05,]

res_4.1.3_sig$Pval <- signif(res_4.1.3_sig$pval, 3)
res_4.1.3_sig <- res_4.1.3_sig[order(`SUB PATHWAY`, beta), ]

p_str_log_ige <- ggplot(res_4.1.3_sig, 
                aes(reorder(tentative_annotation, -beta), y = beta, color = beta >0)) +
   scale_colour_manual(name = '', values = setNames(c('red','blue'),c(T, F)))+
   geom_rect(aes(fill = `SUB PATHWAY`),xmin = -Inf,xmax = Inf,
             ymin = -Inf,ymax = Inf, alpha = 0, color = 'black', linejoin = "round") +
   facet_grid(`SUB PATHWAY` ~ `SUPER PATHWAY`, scales = "free_y", space = "free_y", labeller = label_wrap_gen()) +
   geom_errorbar(aes(ymin = lower95, ymax = upper95), width = 0.25) +
   geom_pointrange(aes(ymin = lower95, ymax = upper95), size = 0.25) +
   geom_hline(yintercept = 0, linetype = 2) +
   labs(
     x = "",
     y = ""
   ) +
   theme_minimal() +
   theme_bw() +
   theme(
     title = element_text(size = 16),
     strip.text.x = element_text(size = 12, face = "bold"),
     strip.text.y = element_text(size = 10, face = "bold"),
     axis.title = element_text(size = 12),
     axis.text = element_text(size = 10, face = "bold"),
     legend.position = "",
     legend.title = element_text(size = 12),
     legend.text = element_text(size = 12),
     strip.text.y.right = element_text(angle = 0),
     strip.background = element_rect(fill = "lemonchiffon", colour = "black")) +
   coord_flip() + scale_fill_viridis_d(option = "D", direction = -1, begin = 0, end = 1, alpha = 0.1)
   # scale_y_continuous(limits = c(0.2, 1.6))

```


## 4.2 FEV1
```{r}
sum(is.na(dat_tmp$FEV1_most_recent)) #266
ggplot(dat_tmp, aes(x=FEV1_most_recent)) + geom_histogram(color="darkblue", fill="lightblue")

```

### 4.2.1 targeted sph
```{r}
length(targeted_sph_list) #78

res_4.2.1 <- linreg_res_fnc(outc = "FEV1_most_recent", 
                            mets_list = targeted_sph_list, 
                            dat = "dat_tmp", 
                            add_covar = " + Age_at_Collection + bmi + Gender + race + Ethnicity + ics_trim_totnum_5y")

res_4.2.1 <- merge(res_4.2.1, sph_mets_info, by = "feat_id", all.x = T)

## display results
datatable(res_4.2.1[pval <= 0.05, c("feat_id", "metabolite", "outc", "pval", "fdr", "beta", "lower95", "upper95", "SUB PATHWAY")], 
          filter = "top") %>% 
          formatSignif(columns = c("pval", "beta", "lower95", "upper95"), digits = 3)

res_4.2.1_sig <- res_4.2.1[pval <= 0.05,]
res_4.2.1_sig$Pval <- signif(res_4.2.1_sig$pval, 3)
res_4.2.1_sig <- res_4.2.1_sig[order(`SUB PATHWAY`, beta), ]

p_sph_fev1 <- ggplot(res_4.2.1_sig, 
                aes(reorder(tentative_annotation, -beta), y = beta, color = beta >0)) +
   scale_colour_manual(name = '', values = setNames(c('red','blue'),c(T, F)))+
   geom_rect(aes(fill = `SUB PATHWAY`),xmin = -Inf,xmax = Inf,
             ymin = -Inf,ymax = Inf, alpha = 0, color = 'black', linejoin = "round") +
   facet_grid(`SUB PATHWAY` ~ `SUPER PATHWAY`, scales = "free_y", space = "free_y", labeller = label_wrap_gen()) +
   geom_errorbar(aes(ymin = lower95, ymax = upper95), width = 0.25) +
   geom_pointrange(aes(ymin = lower95, ymax = upper95), size = 0.25) +
   geom_hline(yintercept = 0, linetype = 2) +
   labs(
     x = "",
     y = ""
   ) +
   theme_minimal() +
   theme_bw() +
   theme(
     title = element_text(size = 16),
     strip.text.x = element_text(size = 12, face = "bold"),
     strip.text.y = element_text(size = 10, face = "bold"),
     axis.title = element_text(size = 12),
     axis.text = element_text(size = 10, face = "bold"),
     axis.text.x = element_text(margin = margin(t = 10)),  # Adjust top margin of x-axis labels
     axis.text.y = element_text(margin = margin(r = 10)),   # Adjust right margin of y-axis labels
     legend.position = "",
     legend.title = element_text(size = 12),
     legend.text = element_text(size = 12),
     strip.text.y.right = element_text(angle = 0),
     strip.background = element_rect(fill = "mistyrose", colour = "black")) +
   coord_flip() + scale_fill_viridis_d(option = "D", direction = -1, begin = 0, end = 1, alpha = 0.1)
   # scale_y_continuous(limits = c(0.2, 1.6))

```


### 4.2.2 targeted microb
```{r}
length(targeted_microb_list) #51

res_4.2.2 <- linreg_res_fnc(outc = "FEV1_most_recent", 
                            mets_list = targeted_microb_list, 
                            dat = "dat_tmp", 
                            add_covar = " + Age_at_Collection + bmi + Gender + race + Ethnicity + ics_trim_totnum_5y")

res_4.2.2 <- merge(res_4.2.2, microb_mets_info, by = "feat_id", all.x = T)

## display results
datatable(res_4.2.2[pval <= 0.05, c("feat_id", "Component_Name", "outc", "pval", "fdr", "beta", "lower95", "upper95", "SUB_PATHWAY")], 
          filter = "top") %>% 
          formatSignif(columns = c("pval", "beta", "lower95", "upper95"), digits = 3)

res_4.2.2_sig <- res_4.2.2[pval <= 0.05,]
res_4.2.2_sig$Pval <- signif(res_4.2.2_sig$pval, 3)
res_4.2.2_sig <- res_4.2.2_sig[order(`SUB_PATHWAY`, beta), ]

p_microb_fev1 <- ggplot(res_4.2.2_sig, 
                   aes(reorder(tentative_annotation, -beta), y = beta, color = beta >0)) +
   scale_colour_manual(name = '', values = setNames(c('red','blue'),c(T, F)))+
   geom_rect(aes(fill = `SUB_PATHWAY`),xmin = -Inf,xmax = Inf,
             ymin = -Inf,ymax = Inf, alpha = 0, color = 'black', linejoin = "round") +
   facet_grid(`SUB_PATHWAY` ~ `SUPER_PATHWAY`, scales = "free_y", space = "free_y", labeller = label_wrap_gen()) +
   geom_errorbar(aes(ymin = lower95, ymax = upper95), width = 0.25) +
   geom_pointrange(aes(ymin = lower95, ymax = upper95), size = 0.25) +
   geom_hline(yintercept = 0, linetype = 2) +
   labs(
     x = "",
     y = ""
   ) +
   theme_minimal() +
   theme_bw() +
   theme(
     title = element_text(size = 16),
     strip.text.x = element_text(size = 12, face = "bold"),
     strip.text.y = element_text(size = 10, face = "bold"),
     axis.title = element_text(size = 12),
     axis.text = element_text(size = 10, face = "bold"),
     legend.position = "",
     legend.title = element_text(size = 12),
     legend.text = element_text(size = 12),
     strip.text.y.right = element_text(angle = 0),
     strip.background = element_rect(fill = "lightcyan", colour = "black")) +
   coord_flip() + scale_fill_viridis_d(option = "D", direction = -1, begin = 0, end = 1, alpha = 0.1)
   # scale_y_continuous(limits = c(-1.6, 4.1))

```


### 4.2.3 targeted str
```{r}
length(targeted_str_list) #16

res_4.2.3 <- linreg_res_fnc(outc = "FEV1_most_recent", 
                            mets_list = targeted_str_list, 
                            dat = "dat_tmp", 
                            add_covar = " + Age_at_Collection + bmi + Gender + race + Ethnicity + ics_trim_totnum_5y")

res_4.2.3 <- merge(res_4.2.3, str_mets_info, by = "feat_id", all.x = T)

## display results
datatable(res_4.2.3[pval <= 0.05, c("feat_id", "Component Name", "outc", "pval", "fdr", "beta", "lower95", "upper95", "SUB PATHWAY")], 
          filter = "top") %>% 
          formatSignif(columns = c("pval", "beta", "lower95", "upper95"), digits = 3)

res_4.2.3_sig <- res_4.2.3[pval <= 0.05,]

res_4.2.3_sig$Pval <- signif(res_4.2.3_sig$pval, 3)
res_4.2.3_sig <- res_4.2.3_sig[order(`SUB PATHWAY`, beta), ]

p_str_fev1 <- ggplot(res_4.2.3_sig, 
                aes(reorder(tentative_annotation, -beta), y = beta, color = beta >0)) +
   scale_colour_manual(name = '', values = setNames(c('red','blue'),c(T, F)))+
   geom_rect(aes(fill = `SUB PATHWAY`),xmin = -Inf,xmax = Inf,
             ymin = -Inf,ymax = Inf, alpha = 0, color = 'black', linejoin = "round") +
   facet_grid(`SUB PATHWAY` ~ `SUPER PATHWAY`, scales = "free_y", space = "free_y", labeller = label_wrap_gen()) +
   geom_errorbar(aes(ymin = lower95, ymax = upper95), width = 0.25) +
   geom_pointrange(aes(ymin = lower95, ymax = upper95), size = 0.25) +
   geom_hline(yintercept = 0, linetype = 2) +
   labs(
     x = "",
     y = ""
   ) +
   theme_minimal() +
   theme_bw() +
   theme(
     title = element_text(size = 16),
     strip.text.x = element_text(size = 12, face = "bold"),
     strip.text.y = element_text(size = 10, face = "bold"),
     axis.title = element_text(size = 12),
     axis.text = element_text(size = 10, face = "bold"),
     legend.position = "",
     legend.title = element_text(size = 12),
     legend.text = element_text(size = 12),
     strip.text.y.right = element_text(angle = 0),
     strip.background = element_rect(fill = "lemonchiffon", colour = "black")) +
   coord_flip() + scale_fill_viridis_d(option = "D", direction = -1, begin = 0, end = 1, alpha = 0.1)
   # scale_y_continuous(limits = c(0.2, 1.6))

```

## 4.3 FVC
```{r}
sum(is.na(dat_tmp$FVC_most_recent)) #268
ggplot(dat_tmp, aes(x=FVC_most_recent)) + geom_histogram(color="darkblue", fill="lightblue")

```

### 4.3.1 targeted sph
```{r}
length(targeted_sph_list) #78

res_4.3.1 <- linreg_res_fnc(outc = "FVC_most_recent", 
                            mets_list = targeted_sph_list, 
                            dat = "dat_tmp", 
                            add_covar = " + Age_at_Collection + bmi + Gender + race + Ethnicity + ics_trim_totnum_5y")

res_4.3.1 <- merge(res_4.3.1, sph_mets_info, by = "feat_id", all.x = T)

## display results
datatable(res_4.3.1[pval <= 0.05, c("feat_id", "metabolite", "outc", "pval", "fdr", "beta", "lower95", "upper95", "SUB PATHWAY")], 
          filter = "top") %>% 
          formatSignif(columns = c("pval", "beta", "lower95", "upper95"), digits = 3)

res_4.3.1_sig <- res_4.3.1[pval <= 0.05,]
res_4.3.1_sig$Pval <- signif(res_4.3.1_sig$pval, 3)
res_4.3.1_sig <- res_4.3.1_sig[order(`SUB PATHWAY`, beta), ]

p_sph_fvc <- ggplot(res_4.3.1_sig, 
                aes(reorder(tentative_annotation, -beta), y = beta, color = beta >0)) +
   scale_colour_manual(name = '', values = setNames(c('red','blue'),c(T, F)))+
   geom_rect(aes(fill = `SUB PATHWAY`),xmin = -Inf,xmax = Inf,
             ymin = -Inf,ymax = Inf, alpha = 0, color = 'black', linejoin = "round") +
   facet_grid(`SUB PATHWAY` ~ `SUPER PATHWAY`, scales = "free_y", space = "free_y", labeller = label_wrap_gen()) +
   geom_errorbar(aes(ymin = lower95, ymax = upper95), width = 0.25) +
   geom_pointrange(aes(ymin = lower95, ymax = upper95), size = 0.25) +
   geom_hline(yintercept = 0, linetype = 2) +
   labs(
     x = "",
     y = ""
   ) +
   theme_minimal() +
   theme_bw() +
   theme(
     title = element_text(size = 16),
     strip.text.x = element_text(size = 12, face = "bold"),
     strip.text.y = element_text(size = 10, face = "bold"),
     axis.title = element_text(size = 12),
     axis.text = element_text(size = 10, face = "bold"),
     axis.text.x = element_text(margin = margin(t = 10)),  # Adjust top margin of x-axis labels
     axis.text.y = element_text(margin = margin(r = 10)),   # Adjust right margin of y-axis labels
     legend.position = "",
     legend.title = element_text(size = 12),
     legend.text = element_text(size = 12),
     strip.text.y.right = element_text(angle = 0),
     strip.background = element_rect(fill = "mistyrose", colour = "black")) +
   coord_flip() + scale_fill_viridis_d(option = "D", direction = -1, begin = 0, end = 1, alpha = 0.1)
   # scale_y_continuous(limits = c(0.2, 1.6))

```


### 4.3.2 targeted microb
```{r}
length(targeted_microb_list) #51

res_4.3.2 <- linreg_res_fnc(outc = "FVC_most_recent", 
                            mets_list = targeted_microb_list, 
                            dat = "dat_tmp", 
                            add_covar = " + Age_at_Collection + bmi + Gender + race + Ethnicity + ics_trim_totnum_5y")

res_4.3.2 <- merge(res_4.3.2, microb_mets_info, by = "feat_id", all.x = T)

## display results
datatable(res_4.3.2[pval <= 0.05, c("feat_id", "Component_Name", "outc", "pval", "fdr", "beta", "lower95", "upper95", "SUB_PATHWAY")], 
          filter = "top") %>% 
          formatSignif(columns = c("pval", "beta", "lower95", "upper95"), digits = 3)

res_4.3.2_sig <- res_4.3.2[pval <= 0.05,]
res_4.3.2_sig$Pval <- signif(res_4.3.2_sig$pval, 3)
res_4.3.2_sig <- res_4.3.2_sig[order(`SUB_PATHWAY`, beta), ]

p_microb_fvc <- ggplot(res_4.3.2_sig, 
                   aes(reorder(tentative_annotation, -beta), y = beta, color = beta >0)) +
   scale_colour_manual(name = '', values = setNames(c('red','blue'),c(T, F)))+
   geom_rect(aes(fill = `SUB_PATHWAY`),xmin = -Inf,xmax = Inf,
             ymin = -Inf,ymax = Inf, alpha = 0, color = 'black', linejoin = "round") +
   facet_grid(`SUB_PATHWAY` ~ `SUPER_PATHWAY`, scales = "free_y", space = "free_y", labeller = label_wrap_gen()) +
   geom_errorbar(aes(ymin = lower95, ymax = upper95), width = 0.25) +
   geom_pointrange(aes(ymin = lower95, ymax = upper95), size = 0.25) +
   geom_hline(yintercept = 0, linetype = 2) +
   labs(
     x = "",
     y = ""
   ) +
   theme_minimal() +
   theme_bw() +
   theme(
     title = element_text(size = 16),
     strip.text.x = element_text(size = 12, face = "bold"),
     strip.text.y = element_text(size = 10, face = "bold"),
     axis.title = element_text(size = 12),
     axis.text = element_text(size = 10, face = "bold"),
     legend.position = "",
     legend.title = element_text(size = 12),
     legend.text = element_text(size = 12),
     strip.text.y.right = element_text(angle = 0),
     strip.background = element_rect(fill = "lightcyan", colour = "black")) +
   coord_flip() + scale_fill_viridis_d(option = "D", direction = -1, begin = 0, end = 1, alpha = 0.1)
   # scale_y_continuous(limits = c(-1.6, 4.1))

```


### 4.3.3 targeted str
```{r}
length(targeted_str_list) #16

res_4.3.3 <- linreg_res_fnc(outc = "FVC_most_recent", 
                            mets_list = targeted_str_list, 
                            dat = "dat_tmp", 
                            add_covar = " + Age_at_Collection + bmi + Gender + race + Ethnicity + ics_trim_totnum_5y")

res_4.3.3 <- merge(res_4.3.3, str_mets_info, by = "feat_id", all.x = T)

## display results
datatable(res_4.3.3[pval <= 0.05, c("feat_id", "Component Name", "outc", "pval", "fdr", "beta", "lower95", "upper95", "SUB PATHWAY")], 
          filter = "top") %>% 
          formatSignif(columns = c("pval", "beta", "lower95", "upper95"), digits = 3)

res_4.3.3_sig <- res_4.3.3[pval <= 0.05,]

res_4.3.3_sig$Pval <- signif(res_4.3.3_sig$pval, 3)
res_4.3.3_sig <- res_4.3.3_sig[order(`SUB PATHWAY`, beta), ]

p_str_fvc <- ggplot(res_4.3.3_sig, 
                aes(reorder(tentative_annotation, -beta), y = beta, color = beta >0)) +
   scale_colour_manual(name = '', values = setNames(c('red','blue'),c(T, F)))+
   geom_rect(aes(fill = `SUB PATHWAY`),xmin = -Inf,xmax = Inf,
             ymin = -Inf,ymax = Inf, alpha = 0, color = 'black', linejoin = "round") +
   facet_grid(`SUB PATHWAY` ~ `SUPER PATHWAY`, scales = "free_y", space = "free_y", labeller = label_wrap_gen()) +
   geom_errorbar(aes(ymin = lower95, ymax = upper95), width = 0.25) +
   geom_pointrange(aes(ymin = lower95, ymax = upper95), size = 0.25) +
   geom_hline(yintercept = 0, linetype = 2) +
   labs(
     x = "",
     y = ""
   ) +
   theme_minimal() +
   theme_bw() +
   theme(
     title = element_text(size = 16),
     strip.text.x = element_text(size = 12, face = "bold"),
     strip.text.y = element_text(size = 10, face = "bold"),
     axis.title = element_text(size = 12),
     axis.text = element_text(size = 10, face = "bold"),
     legend.position = "",
     legend.title = element_text(size = 12),
     legend.text = element_text(size = 12),
     strip.text.y.right = element_text(angle = 0),
     strip.background = element_rect(fill = "lemonchiffon", colour = "black")) +
   coord_flip() + scale_fill_viridis_d(option = "D", direction = -1, begin = 0, end = 1, alpha = 0.1)
   # scale_y_continuous(limits = c(0.2, 1.6))

```


## 4.4 FEV1/FVC
```{r}
sum(is.na(dat_tmp$FEV1_FVC_most_recent)) #268
ggplot(dat_tmp, aes(x=FEV1_FVC_most_recent)) + geom_histogram(color="darkblue", fill="lightblue")

```

### 4.4.1 targeted sph
```{r}
length(targeted_sph_list) #78

res_4.4.1 <- linreg_res_fnc(outc = "FEV1_FVC_most_recent", 
                            mets_list = targeted_sph_list, 
                            dat = "dat_tmp", 
                            add_covar = " + Age_at_Collection + bmi + Gender + race + Ethnicity + ics_trim_totnum_5y")

res_4.4.1 <- merge(res_4.4.1, sph_mets_info, by = "feat_id", all.x = T)

## display results
datatable(res_4.4.1[pval <= 0.05, c("feat_id", "metabolite", "outc", "pval", "fdr", "beta", "lower95", "upper95", "SUB PATHWAY")], 
          filter = "top") %>% 
          formatSignif(columns = c("pval", "beta", "lower95", "upper95"), digits = 3)

res_4.4.1_sig <- res_4.4.1[pval <= 0.05,]


```


### 4.4.2 targeted microb
```{r}
length(targeted_microb_list) #51

res_4.4.2 <- linreg_res_fnc(outc = "FEV1_FVC_most_recent", 
                            mets_list = targeted_microb_list, 
                            dat = "dat_tmp", 
                            add_covar = " + Age_at_Collection + bmi + Gender + race + Ethnicity + ics_trim_totnum_5y")

res_4.4.2 <- merge(res_4.4.2, microb_mets_info, by = "feat_id", all.x = T)

## display results
datatable(res_4.4.2[pval <= 0.05, c("feat_id", "Component_Name", "outc", "pval", "fdr", "beta", "lower95", "upper95", "SUB_PATHWAY")], 
          filter = "top") %>% 
          formatSignif(columns = c("pval", "beta", "lower95", "upper95"), digits = 3)

res_4.4.2_sig <- res_4.4.2[pval <= 0.05,]
res_4.4.2_sig$Pval <- signif(res_4.4.2_sig$pval, 3)
res_4.4.2_sig <- res_4.4.2_sig[order(`SUB_PATHWAY`, beta), ]

p_microb_fev1_fvc <- ggplot(res_4.4.2_sig, 
                   aes(reorder(tentative_annotation, -beta), y = beta, color = beta >0)) +
   scale_colour_manual(name = '', values = setNames(c('red','blue'),c(T, F)))+
   geom_rect(aes(fill = `SUB_PATHWAY`),xmin = -Inf,xmax = Inf,
             ymin = -Inf,ymax = Inf, alpha = 0, color = 'black', linejoin = "round") +
   facet_grid(`SUB_PATHWAY` ~ `SUPER_PATHWAY`, scales = "free_y", space = "free_y", labeller = label_wrap_gen()) +
   geom_errorbar(aes(ymin = lower95, ymax = upper95), width = 0.25) +
   geom_pointrange(aes(ymin = lower95, ymax = upper95), size = 0.25) +
   geom_hline(yintercept = 0, linetype = 2) +
   labs(
     x = "",
     y = ""
   ) +
   theme_minimal() +
   theme_bw() +
   theme(
     title = element_text(size = 16),
     strip.text.x = element_text(size = 12, face = "bold"),
     strip.text.y = element_text(size = 10, face = "bold"),
     axis.title = element_text(size = 12),
     axis.text = element_text(size = 10, face = "bold"),
     legend.position = "",
     legend.title = element_text(size = 12),
     legend.text = element_text(size = 12),
     strip.text.y.right = element_text(angle = 0),
     strip.background = element_rect(fill = "lightcyan", colour = "black")) +
   coord_flip() + scale_fill_viridis_d(option = "D", direction = -1, begin = 0, end = 1, alpha = 0.1)
   # scale_y_continuous(limits = c(-1.6, 4.1))

```


### 4.4.3 targeted str
```{r}
length(targeted_str_list) #16

res_4.4.3 <- linreg_res_fnc(outc = "FEV1_FVC_most_recent", 
                            mets_list = targeted_str_list, 
                            dat = "dat_tmp", 
                            add_covar = " + Age_at_Collection + bmi + Gender + race + Ethnicity + ics_trim_totnum_5y")

res_4.4.3 <- merge(res_4.4.3, str_mets_info, by = "feat_id", all.x = T)

## display results
datatable(res_4.4.3[pval <= 0.05, c("feat_id", "Component Name", "outc", "pval", "fdr", "beta", "lower95", "upper95", "SUB PATHWAY")], 
          filter = "top") %>% 
          formatSignif(columns = c("pval", "beta", "lower95", "upper95"), digits = 3)

res_4.4.3_sig <- res_4.4.3[pval <= 0.05,]

res_4.4.3_sig$Pval <- signif(res_4.4.3_sig$pval, 3)
res_4.4.3_sig <- res_4.4.3_sig[order(`SUB PATHWAY`, beta), ]

p_str_fev1_fvc <- ggplot(res_4.4.3_sig, 
                aes(reorder(tentative_annotation, -beta), y = beta, color = beta >0)) +
   scale_colour_manual(name = '', values = setNames(c('red','blue'),c(T, F)))+
   geom_rect(aes(fill = `SUB PATHWAY`),xmin = -Inf,xmax = Inf,
             ymin = -Inf,ymax = Inf, alpha = 0, color = 'black', linejoin = "round") +
   facet_grid(`SUB PATHWAY` ~ `SUPER PATHWAY`, scales = "free_y", space = "free_y", labeller = label_wrap_gen()) +
   geom_errorbar(aes(ymin = lower95, ymax = upper95), width = 0.25) +
   geom_pointrange(aes(ymin = lower95, ymax = upper95), size = 0.25) +
   geom_hline(yintercept = 0, linetype = 2) +
   labs(
     x = "",
     y = ""
   ) +
   theme_minimal() +
   theme_bw() +
   theme(
     title = element_text(size = 16),
     strip.text.x = element_text(size = 12, face = "bold"),
     strip.text.y = element_text(size = 10, face = "bold"),
     axis.title = element_text(size = 12),
     axis.text = element_text(size = 10, face = "bold"),
     legend.position = "",
     legend.title = element_text(size = 12),
     legend.text = element_text(size = 12),
     strip.text.y.right = element_text(angle = 0),
     strip.background = element_rect(fill = "lemonchiffon", colour = "black")) +
   coord_flip() + scale_fill_viridis_d(option = "D", direction = -1, begin = 0, end = 1, alpha = 0.1)
   # scale_y_continuous(limits = c(0.2, 1.6))

```

## 4.5 OCS use number within 5 years (exacerbation)
```{r}
sum(is.na(dat_tmp$ocs_trim_totnum_5y)) #12
ggplot(dat_tmp, aes(x=ocs_trim_totnum_5y)) + geom_histogram(color="darkblue", fill="lightblue")

```

### 4.5.1 targeted sph
```{r}
length(targeted_sph_list) #78

res_4.5.1 <- poiss_res_fnc(outc = "ocs_trim_totnum_5y", 
                           mets_list = targeted_sph_list, 
                           dat = "dat_tmp", 
                           add_covar = " + Age_at_Collection + bmi + Gender + race + Ethnicity + ics_trim_totnum_5y")

res_4.5.1 <- merge(res_4.5.1, sph_mets_info, by = "feat_id", all.x = T)

## display results
datatable(res_4.5.1[pval <= 0.05, c("feat_id", "metabolite", "outc", "pval", "fdr", "beta", "lower95", "upper95", "SUB PATHWAY")], 
          filter = "top") %>% 
          formatSignif(columns = c("pval", "beta", "lower95", "upper95"), digits = 3)

res_4.5.1_sig <- res_4.5.1[pval <= 0.05,]
res_4.5.1_sig$Pval <- signif(res_4.5.1_sig$pval, 3)
res_4.5.1_sig <- res_4.5.1_sig[order(`SUB PATHWAY`, beta), ]

p_sph_ocs <- ggplot(res_4.5.1_sig, 
                aes(reorder(tentative_annotation, -beta), y = beta, color = beta >0)) +
   scale_colour_manual(name = '', values = setNames(c('red','blue'),c(T, F)))+
   geom_rect(aes(fill = `SUB PATHWAY`),xmin = -Inf,xmax = Inf,
             ymin = -Inf,ymax = Inf, alpha = 0, color = 'black', linejoin = "round") +
   facet_grid(`SUB PATHWAY` ~ `SUPER PATHWAY`, scales = "free_y", space = "free_y", labeller = label_wrap_gen()) +
   geom_errorbar(aes(ymin = lower95, ymax = upper95), width = 0.25) +
   geom_pointrange(aes(ymin = lower95, ymax = upper95), size = 0.25) +
   geom_hline(yintercept = 0, linetype = 2) +
   labs(
     x = "",
     y = ""
   ) +
   theme_minimal() +
   theme_bw() +
   theme(
     title = element_text(size = 16),
     strip.text.x = element_text(size = 12, face = "bold"),
     strip.text.y = element_text(size = 10, face = "bold"),
     axis.title = element_text(size = 12),
     axis.text = element_text(size = 10, face = "bold"),
     axis.text.x = element_text(margin = margin(t = 10)),  # Adjust top margin of x-axis labels
     axis.text.y = element_text(margin = margin(r = 10)),   # Adjust right margin of y-axis labels
     legend.position = "",
     legend.title = element_text(size = 12),
     legend.text = element_text(size = 12),
     strip.text.y.right = element_text(angle = 0),
     strip.background = element_rect(fill = "mistyrose", colour = "black")) +
   coord_flip() + scale_fill_viridis_d(option = "D", direction = -1, begin = 0, end = 1, alpha = 0.1)
   # scale_y_continuous(limits = c(0.2, 1.6))

```


### 4.5.2 targeted microb
```{r}
length(targeted_microb_list) #51

res_4.5.2 <- poiss_res_fnc(outc = "ocs_trim_totnum_5y", 
                           mets_list = targeted_microb_list, 
                           dat = "dat_tmp", 
                           add_covar = " + Age_at_Collection + bmi + Gender + race + Ethnicity + ics_trim_totnum_5y")

res_4.5.2 <- merge(res_4.5.2, microb_mets_info, by = "feat_id", all.x = T)

## display results
datatable(res_4.5.2[pval <= 0.05, c("feat_id", "Component_Name", "outc", "pval", "fdr", "beta", "lower95", "upper95", "SUB_PATHWAY")], 
          filter = "top") %>% 
          formatSignif(columns = c("pval", "beta", "lower95", "upper95"), digits = 3)

res_4.5.2_sig <- res_4.5.2[pval <= 0.05,]
res_4.5.2_sig <- res_4.5.2_sig[feat_id != "trypkynu_ratio_microb",]
res_4.5.2_sig$Pval <- signif(res_4.5.2_sig$pval, 3)
res_4.5.2_sig <- res_4.5.2_sig[order(`SUB_PATHWAY`, beta), ]

p_microb_ocs <- ggplot(res_4.5.2_sig, 
                   aes(reorder(tentative_annotation, -beta), y = beta, color = beta >0)) +
   scale_colour_manual(name = '', values = setNames(c('red','blue'),c(T, F)))+
   geom_rect(aes(fill = `SUB_PATHWAY`),xmin = -Inf,xmax = Inf,
             ymin = -Inf,ymax = Inf, alpha = 0, color = 'black', linejoin = "round") +
   facet_grid(`SUB_PATHWAY` ~ `SUPER_PATHWAY`, scales = "free_y", space = "free_y", labeller = label_wrap_gen()) +
   geom_errorbar(aes(ymin = lower95, ymax = upper95), width = 0.25) +
   geom_pointrange(aes(ymin = lower95, ymax = upper95), size = 0.25) +
   geom_hline(yintercept = 0, linetype = 2) +
   labs(
     x = "",
     y = ""
   ) +
   theme_minimal() +
   theme_bw() +
   theme(
     title = element_text(size = 16),
     strip.text.x = element_text(size = 12, face = "bold"),
     strip.text.y = element_text(size = 10, face = "bold"),
     axis.title = element_text(size = 12),
     axis.text = element_text(size = 10, face = "bold"),
     legend.position = "",
     legend.title = element_text(size = 12),
     legend.text = element_text(size = 12),
     strip.text.y.right = element_text(angle = 0),
     strip.background = element_rect(fill = "lightcyan", colour = "black")) +
   coord_flip() + scale_fill_viridis_d(option = "D", direction = -1, begin = 0, end = 1, alpha = 0.1)
   # scale_y_continuous(limits = c(-1.6, 4.1))


```


### 4.5.3 targeted str
```{r}
length(targeted_str_list) #16

res_4.5.3 <- poiss_res_fnc(outc = "ocs_trim_totnum_5y", 
                           mets_list = targeted_str_list, 
                           dat = "dat_tmp", 
                           add_covar = " + Age_at_Collection + bmi + Gender + race + Ethnicity + ics_trim_totnum_5y")

res_4.5.3 <- merge(res_4.5.3, str_mets_info, by = "feat_id", all.x = T)

## display results
datatable(res_4.5.3[pval <= 0.05, c("feat_id", "Component Name", "outc", "pval", "fdr", "beta", "lower95", "upper95", "SUB PATHWAY")], 
          filter = "top") %>% 
          formatSignif(columns = c("pval", "beta", "lower95", "upper95"), digits = 3)

res_4.5.3_sig <- res_4.5.3[pval <= 0.05,]

res_4.5.3_sig$Pval <- signif(res_4.5.3_sig$pval, 3)
res_4.5.3_sig <- res_4.5.3_sig[order(`SUB PATHWAY`, beta), ]

p_str_ocs <- ggplot(res_4.5.3_sig, 
                aes(reorder(tentative_annotation, -beta), y = beta, color = beta >0)) +
   scale_colour_manual(name = '', values = setNames(c('red','blue'),c(T, F)))+
   geom_rect(aes(fill = `SUB PATHWAY`),xmin = -Inf,xmax = Inf,
             ymin = -Inf,ymax = Inf, alpha = 0, color = 'black', linejoin = "round") +
   facet_grid(`SUB PATHWAY` ~ `SUPER PATHWAY`, scales = "free_y", space = "free_y", labeller = label_wrap_gen()) +
   geom_errorbar(aes(ymin = lower95, ymax = upper95), width = 0.25) +
   geom_pointrange(aes(ymin = lower95, ymax = upper95), size = 0.25) +
   geom_hline(yintercept = 0, linetype = 2) +
   labs(
     x = "",
     y = ""
   ) +
   theme_minimal() +
   theme_bw() +
   theme(
     title = element_text(size = 16),
     strip.text.x = element_text(size = 12, face = "bold"),
     strip.text.y = element_text(size = 10, face = "bold"),
     axis.title = element_text(size = 12),
     axis.text = element_text(size = 10, face = "bold"),
     legend.position = "",
     legend.title = element_text(size = 12),
     legend.text = element_text(size = 12),
     strip.text.y.right = element_text(angle = 0),
     strip.background = element_rect(fill = "lemonchiffon", colour = "black")) +
   coord_flip() + scale_fill_viridis_d(option = "D", direction = -1, begin = 0, end = 1, alpha = 0.1)
   # scale_y_continuous(limits = c(0.2, 1.6))

```

# 5. Association with Asthma
## 5.1 targeted sph
```{r}
res_5.1 <- clogi_res_fnc(outc = "asthma_num", 
                         mets_list = targeted_sph_list, 
                         dat = "pheno_mets_trim", 
                         mets_info = "sph_mets_info", 
                         add_covar = " + bmi")

## display results
datatable(res_5.1[pval <= 0.05, c("feat_id", "tentative_annotation", "outc", "pval", "fdr", "beta", "lower95", "upper95", "SUB PATHWAY")], 
          filter = "top") %>% 
          formatSignif(columns = c("pval", "beta", "lower95", "upper95"), digits = 3)

res_5.1_sig <- res_5.1[pval <= 0.05,]

res_5.1_sig$Pval <- signif(res_5.1_sig$pval, 3)
res_5.1_sig <- res_5.1_sig[order(`SUB PATHWAY`, beta), ]

p_sph_asthma <- ggplot(res_5.1_sig, 
                aes(reorder(tentative_annotation, -or), y = or, color = or >1)) +
   scale_colour_manual(name = '', values = setNames(c('red','blue'),c(T, F)))+
   geom_rect(aes(fill = `SUB PATHWAY`),xmin = -Inf,xmax = Inf,
             ymin = -Inf,ymax = Inf, alpha = 0, color = 'black', linejoin = "round") +
   facet_grid(`SUB PATHWAY` ~ `SUPER PATHWAY`, scales = "free_y", space = "free_y", labeller = label_wrap_gen()) +
   geom_errorbar(aes(ymin = lower95, ymax = upper95), width = 0.25) +
   geom_pointrange(aes(ymin = lower95, ymax = upper95), size = 0.25) +
   geom_hline(yintercept = 1, linetype = 2) +
   labs(
     x = "",
     y = ""
   ) +
   theme_minimal() +
   theme_bw() +
   theme(
     title = element_text(size = 16),
     strip.text.x = element_text(size = 12, face = "bold"),
     strip.text.y = element_text(size = 10, face = "bold"),
     axis.title = element_text(size = 12),
     axis.text = element_text(size = 10, face = "bold"),
     axis.text.x = element_text(margin = margin(t = 10)),  # Adjust top margin of x-axis labels
     axis.text.y = element_text(margin = margin(r = 10)),   # Adjust right margin of y-axis labels
     legend.position = "",
     legend.title = element_text(size = 12),
     legend.text = element_text(size = 12),
     strip.text.y.right = element_text(angle = 0),
     strip.background = element_rect(fill = "mistyrose", colour = "black")) +
   coord_flip() + scale_fill_viridis_d(option = "D", direction = -1, begin = 0, end = 1, alpha = 0.1) +
   scale_y_continuous(limits = c(0.2, 1.6))

```

## 5.2 targeted microb
```{r}
res_5.2 <- clogi_res_fnc(outc = "asthma_num", 
                         mets_list = targeted_microb_list, 
                         dat = "pheno_mets_trim", 
                         mets_info = "microb_mets_info", 
                         add_covar = " + bmi")

## display results
datatable(res_5.2[pval <= 0.05, c("feat_id", "tentative_annotation", "outc", "pval", "fdr", "beta", "lower95", "upper95", "SUB_PATHWAY")], 
          filter = "top") %>% 
          formatSignif(columns = c("pval", "beta", "lower95", "upper95"), digits = 3)

res_5.2_sig <- res_5.2[pval <= 0.05,]

res_5.2_sig$SUPER_PATHWAY <- "Microbiome Metabolism"
res_5.2_sig$Pval <- signif(res_5.2_sig$pval, 3)
res_5.2_sig <- res_5.2_sig[order(`SUB_PATHWAY`, beta), ]


p_microb_asthma <- ggplot(res_5.2_sig, 
                   aes(reorder(tentative_annotation, -or), y = or, color = or >1)) +
   scale_colour_manual(name = '', values = setNames(c('red','blue'),c(T, F)))+
   geom_rect(aes(fill = `SUB_PATHWAY`),xmin = -Inf,xmax = Inf,
             ymin = -Inf,ymax = Inf, alpha = 0, color = 'black', linejoin = "round") +
   facet_grid(`SUB_PATHWAY` ~ `SUPER_PATHWAY`, scales = "free_y", space = "free_y", labeller = label_wrap_gen()) +
   geom_errorbar(aes(ymin = lower95, ymax = upper95), width = 0.25) +
   geom_pointrange(aes(ymin = lower95, ymax = upper95), size = 0.25) +
   geom_hline(yintercept = 1, linetype = 2) +
   labs(
     x = "",
     y = ""
   ) +
   theme_minimal() +
   theme_bw() +
   theme(
     title = element_text(size = 16),
     strip.text.x = element_text(size = 12, face = "bold"),
     strip.text.y = element_text(size = 10, face = "bold"),
     axis.title = element_text(size = 12),
     axis.text = element_text(size = 10, face = "bold"),
     legend.position = "",
     legend.title = element_text(size = 12),
     legend.text = element_text(size = 12),
     strip.text.y.right = element_text(angle = 0),
     strip.background = element_rect(fill = "lightcyan", colour = "black")) +
   coord_flip() + scale_fill_viridis_d(option = "D", direction = -1, begin = 0, end = 1, alpha = 0.1) +
   scale_y_continuous(limits = c(0.2, 1.6))

```

## 5.3 targeted str
```{r}
res_5.3 <- clogi_res_fnc(outc = "asthma_num", 
                         mets_list = targeted_str_list, 
                         dat = "pheno_mets_trim", 
                         mets_info = "str_mets_info", 
                         add_covar = " + bmi")

## display results
datatable(res_5.3[pval <= 0.05, c("feat_id", "tentative_annotation", "outc", "pval", "fdr", "beta", "lower95", "upper95", "SUB PATHWAY")], 
          filter = "top") %>% 
          formatSignif(columns = c("pval", "beta", "lower95", "upper95"), digits = 3)

res_5.3_sig <- res_5.3[pval <= 0.05,]

res_5.3_sig$`SUPER PATHWAY` <- "Steroid Hormones"
res_5.3_sig$Pval <- signif(res_5.3_sig$pval, 3)
res_5.3_sig <- res_5.3_sig[order(`SUB PATHWAY`, beta), ]

p_str_asthma <- ggplot(res_5.3_sig, 
                aes(reorder(tentative_annotation, -or), y = or, color = or >1)) +
   scale_colour_manual(name = '', values = setNames(c('red','blue'),c(T, F)))+
   geom_rect(aes(fill = `SUB PATHWAY`),xmin = -Inf,xmax = Inf,
             ymin = -Inf,ymax = Inf, alpha = 0, color = 'black', linejoin = "round") +
   facet_grid(`SUB PATHWAY` ~ `SUPER PATHWAY`, scales = "free_y", space = "free_y", labeller = label_wrap_gen()) +
   geom_errorbar(aes(ymin = lower95, ymax = upper95), width = 0.25) +
   geom_pointrange(aes(ymin = lower95, ymax = upper95), size = 0.25) +
   geom_hline(yintercept = 1, linetype = 2) +
   labs(
     x = "",
     y = ""
   ) +
   theme_minimal() +
   theme_bw() +
   theme(
     title = element_text(size = 16),
     strip.text.x = element_text(size = 12, face = "bold"),
     strip.text.y = element_text(size = 10, face = "bold"),
     axis.title = element_text(size = 12),
     axis.text = element_text(size = 10, face = "bold"),
     legend.position = "",
     legend.title = element_text(size = 12),
     legend.text = element_text(size = 12),
     strip.text.y.right = element_text(angle = 0),
     strip.background = element_rect(fill = "lemonchiffon", colour = "black")) +
   coord_flip() + scale_fill_viridis_d(option = "D", direction = -1, begin = 0, end = 1, alpha = 0.1) +
   scale_y_continuous(limits = c(0.2, 1.6))

```

# 6. Merge figures
```{r}
## targeted microb
png(file=str_c(fig_dir, "PEGASUS_targeted_microb_phenotypes.png"), width = 36, height = 6, units = "in", res = 300)
grid.arrange(p_microb_asthma, p_microb_fev1, p_microb_fvc, p_microb_fev1_fvc, p_microb_log_ige, p_microb_ocs, nrow = 1)
dev.off()

## targeted sph
png(file=str_c(fig_dir, "PEGASUS_targeted_sph_phenotypes.png"), width = 36, height = 9.5, units = "in", res = 300)
grid.arrange(p_sph_asthma, p_sph_fev1, p_sph_fvc, p_sph_fvc, p_sph_log_ige, p_sph_ocs, nrow = 1)
dev.off()

## targeted str
png(file=str_c(fig_dir, "PEGASUS_targeted_str_phenotypes.png"), width = 36, height = 3.5, units = "in", res = 300)
grid.arrange(p_str_asthma, p_str_fev1, p_str_fvc, p_str_fev1_fvc, p_str_log_ige, p_str_ocs, nrow = 1)
dev.off()

## targeted all
png(file=str_c(fig_dir, "PEGASUS_targeted_all_phenotypes.png"), width = 36, height = 20, units = "in", res = 300)
grid.arrange(p_str_asthma, p_str_fev1, p_str_fvc, p_str_fev1_fvc, p_str_log_ige, p_str_ocs,
             p_microb_asthma, p_microb_fev1, p_microb_fvc, p_microb_fev1_fvc, p_microb_log_ige, p_microb_ocs,
             p_sph_asthma, p_sph_fev1, p_sph_fvc, p_sph_fvc, p_sph_log_ige, p_sph_ocs,
             nrow = 3)
dev.off()

```

# 7. Save Data
```{r}

write.csv(microb_mets_info, file = str_c(res_dir, "PEGASUS_microb_mets_info.csv"), row.names = FALSE)
write.csv(sph_mets_info, file = str_c(res_dir, "PEGASUS_sph_mets_info.csv"), row.names = FALSE)
write.csv(str_mets_info, file = str_c(res_dir, "PEGASUS_str_mets_info.csv"), row.names = FALSE)

write.csv(res_4.1.1_sig, file = str_c(res_dir, "PEGASUS_sig_targeted_sph_log_ige.csv"), row.names = FALSE)
write.csv(res_4.1.2_sig, file = str_c(res_dir, "PEGASUS_sig_targeted_microb_log_ige.csv"), row.names = FALSE)
write.csv(res_4.1.3_sig, file = str_c(res_dir, "PEGASUS_sig_targeted_str_log_ige.csv"), row.names = FALSE)

write.csv(res_4.2.1_sig, file = str_c(res_dir, "PEGASUS_sig_targeted_sph_FEV1.csv"), row.names = FALSE)
write.csv(res_4.2.2_sig, file = str_c(res_dir, "PEGASUS_sig_targeted_microb_FEV1.csv"), row.names = FALSE)
write.csv(res_4.2.3_sig, file = str_c(res_dir, "PEGASUS_sig_targeted_str_FEV1.csv"), row.names = FALSE)

write.csv(res_4.3.1_sig, file = str_c(res_dir, "PEGASUS_sig_targeted_sph_FVC.csv"), row.names = FALSE)
write.csv(res_4.3.2_sig, file = str_c(res_dir, "PEGASUS_sig_targeted_microb_FVC.csv"), row.names = FALSE)
write.csv(res_4.3.3_sig, file = str_c(res_dir, "PEGASUS_sig_targeted_str_FVC.csv"), row.names = FALSE)

write.csv(res_4.4.1_sig, file = str_c(res_dir, "PEGASUS_sig_targeted_sph_FEV1_FVC.csv"), row.names = FALSE)
write.csv(res_4.4.2_sig, file = str_c(res_dir, "PEGASUS_sig_targeted_microb_FEV1_FVC.csv"), row.names = FALSE)
write.csv(res_4.4.3_sig, file = str_c(res_dir, "PEGASUS_sig_targeted_str_FEV1_FVC.csv"), row.names = FALSE)

write.csv(res_4.5.1_sig, file = str_c(res_dir, "PEGASUS_sig_targeted_sph_ocs_use.csv"), row.names = FALSE)
write.csv(res_4.5.2_sig, file = str_c(res_dir, "PEGASUS_sig_targeted_microb_ocs_use.csv"), row.names = FALSE)
write.csv(res_4.5.3_sig, file = str_c(res_dir, "PEGASUS_sig_targeted_str_ocs_use.csv"), row.names = FALSE)

write.csv(res_5.1_sig, file = str_c(res_dir, "PEGASUS_sig_targeted_sph_asthma.csv"), row.names = FALSE)
write.csv(res_5.2_sig, file = str_c(res_dir, "PEGASUS_sig_targeted_microb_asthma.csv"), row.names = FALSE)
write.csv(res_5.3_sig, file = str_c(res_dir, "PEGASUS_sig_targeted_str_asthma.csv"), row.names = FALSE)
```

# Session info
```{r}
sessionInfo()

```
