==> 6_MGB_KAS_targeted_ratio_association.Rmd <==
---
title: "Regression analysis of targeted ratios in PEGASUS"
author: "Yulu Chen"
date: "10/24/2023"
output: 
  html_document: 
    toc: true
    toc_float: 
      collapsed: false
      smooth_scroll: false
    code_folding: hide
editor_options: 
  chunk_output_type: console
---

```{r, include = FALSE}
##### 1. Setup #####
## Packages

pkg <- c("tidyverse", "data.table", "here", "survival", "gtools", "tableone", "DT", "sjPlot", 
         "sjmisc", "grid", "gridExtra", "ggpubr", "table1", "readxl", "ggrepel", "pheatmap",
         "Hmisc", "glmnet", "mltools", "cluster", "factoextra", "dendextend", 
         "survminer", "pROC", "caret", "randomForest", "e1071")
for (p in pkg) {
        if (require(p, character.only = T)) {
                print(paste0(p, " loaded successfully"))
        } else {
                install.packages(p)
                require(p, character.only = T)
                print(paste0(p, " downloaded and loaded successfully"))
        }
}

## Paths
pegasus_dir <- "/udd/reyul/mets_ratio_CodeReview/data/PEGASUS/"

dat_dir <- str_c(pegasus_dir, "2_assembly_global_targeted/")
fig_dir <- str_c("/udd/reyul/mets_ratio_CodeReview/", "figures/6_PEGASUS_targeted_ratio_association/")
res_dir <- str_c("/udd/reyul/mets_ratio_CodeReview/", "results/6_PEGASUS_targeted_ratio_association/")

## Hard-coded numbers

missing_thld <- 0.30
alpha_thld <- 0.05
z_95 <- qnorm(p = 0.975)

## Functions
pvalue <- function(x, ...) {
    # Construct vectors of data y, and groups (strata) g
    y <- unlist(x)
    g <- factor(rep(1:length(x), times=sapply(x, length)))
    if (is.numeric(y)) {
        # For numeric variables, perform a standard 2-sample t-test
        p <- as.data.frame(summary(aov(y ~ g))[[1]])[1,5]
    } else {
        # For categorical variables, perform a chi-squared test of independence
        p <- chisq.test(table(y, g))$p.value
    }
    # Format the p-value, using an HTML entity for the less-than sign.
    # The initial empty string places the output on the line below the variable label.
    c("", sub("<", "&lt;", format.pval(p, digits=6, eps=0.000001)))
}

```


# 1. Load data
## 1.1 all platform
Using the data generated by Yulu and Mengna, including the phenotype/EMR data, global metabolomics data (HILIC+, Lipid+, and Lipid-), targeted sphingolipids data, targeted microbial data

global metabolomics data:
- remove metatolites with sample CV >= 25%
- remove metabolites with >= 30% missing in mets
- impute with half min
- log10 transformation & pareto-scaling

targeted sphingolipids:  remove metabolites with >= 30% missing in mets & Standardize

targeted microbial: remove metabolites with >= 30% missing in mets & Standardize

targeted steroids: remove metabolites with >= 30% missing in mets & Standardize

```{r}

load(here(dat_dir, "all_platforms_data_updated.RData"), verbose = T)

targeted_sph_list <- sph_mets_info[pct_na_met <= 0.30, feat_id]
targeted_microb_list <- microb_mets_info[pct_na_met <= 0.30, feat_id]
targeted_str_list <- str_mets_info[pct_na_met <= 0.30, feat_id]

```

## 1.2 targeted panels
### 1.2.1 sphingolipid
unit: nM
```{r}
sph_dat <- read_xlsx(str_c(pegasus_dir, "targeted_sphingolipids/2020-02-12_PEGASuS_SLs.xlsx"), sheet = "Samples", skip = 2) %>% as.data.table()
setnames(sph_dat, gsub(" ", "_", colnames(sph_dat)))
setnames(sph_dat, "3_KS", "X3_KS")

which(colnames(sph_dat) == "Comments_sample") # 17
which(colnames(sph_dat) == "SubjectID") # 6
sph_dat <- sph_dat[,-c(1:5,7:17)]
colnames(sph_dat)[1] <- "Biobank_Subject_ID"

sph_mets_info$`SUPER PATHWAY` <- "Sphingolipid Metabolism"
sph_mets_info[, `SUB PATHWAY` := case_when(grepl("^HexCer|LacCer", feat_id) ~ "Hex/LacCeramide", 
                                    grepl("^SM", feat_id) ~ "Sphingomyelin", 
                                    grepl("^Sph|Sphd|S1P", feat_id) ~ "Sphingosine/S1P", 
                                    grepl("^Cer1P", feat_id) ~ "Ceramide-1-phosphate", 
                                    grepl("^Cer[1-2][0-9]", feat_id) ~ "Ceramide", 
                                    grepl("^DhCer", feat_id) ~ "Dihydroceramide", 
                                    grepl("^Spa|Spa1P", feat_id) ~ "Sphinganine/Spa1P", 
                                    feat_id == "Glucosph" ~ "Hexosylsphingosine", 
                                    feat_id == "X3_KS" ~ "X3_Ketosphinganine")]

sph_mets_info$tentative_annotation <- c("Cer(d18:1/10:0)", "Cer(d18:1/12:0)", "Cer(d18:1/14:0)", 
                                        "Cer(d16:1/16:0)", "Cer(d18:1/16:0)", "Cer(d18:2/16:0)", 
                                        "Cer(d18:1/17:0)", "Cer(d16:1/18:0)", "Cer(d18:1/18:0)", 
                                        "Cer(d18:2/18:0)", "Cer(d18:1/18:1)", "Cer-1P(d18:1/16:0)", 
                                        "Cer-1P(d18:1/24:0)", "Cer-1P(d18:1/24:1)", "Cer(d18:1/20:0)",
                                        "Cer(d18:1/20:1)", "Cer(d18:1/22:0)", "Cer(d16:1/24:0)", 
                                        "Cer(d18:1/24:0)", "Cer(d18:2/24:0)", "Cer(d16:1/24:1)", 
                                        "Cer(d18:1/24:1)", "Cer(d18:2/24:1)", "Cer(d18:1/26:0)",
                                        "Cer(d18:1/26:1)", "DhCer(d18:0/16:0)", "DhCer(d18:0/18:0)",
                                        "DhCer(d18:0/24:0)", "DhCer(d18:0/24:1)", "Glucosph", 
                                        "HexCer(d18:1/12:0)", "HexCer(d18:1/14:0)", "HexCer(d18:1/16:0)", 
                                        "HexCer(d18:1/18:0)", "HexCer(d18:1/18:1)", "HexCer(d18:1/20:0)",
                                        "HexCer(d18:1/22:0)", "HexCer(d18:1/24:0)", "HexCer(d18:1/24:1)",
                                        "LacCer(d18:1/12:0)", "LacCer(d18:1/14:0)", "LacCer(d18:1/16:0)",
                                        "LacCer(d18:1/17:0)", "LacCer(d18:1/18:0)", "LacCer(d18:1/18:1)", 
                                        "LacCer(d18:1/20:0)", "LacCer(d18:1/22:0)", "LacCer(d18:1/24:0)",
                                        "LacCer(d18:1/24:1)", "Sphingosine-1P(d16:1)", 
                                        "Sphingosine-1P(d17:1)", "Sphingosine-1P", "Sphingosine-1P(d18:2)", 
                                        "SM(d18:1/12:0)",  "SM(d18:1/14:0)", "SM(d16:1/16:0)", 
                                        "SM(d18:1/16:0)", "SM(d18:2/16:0)", "SM(d18:1/17:0)", 
                                        "SM(d16:1/18:0)", "SM(d18:1/18:0)", "SM(d18:2/18:0)", 
                                        "SM(d18:1/18:1)", "SM(d18:1/20:0)", "SM(d18:2/22:0)",
                                        "SM(d18:1/22:0)", "SM(d16:1/24:0)", "SM(d18:1/24:0)", 
                                        "SM(d18:2/24:0)", "SM(d16:1/24:1)",  "SM(d18:1/24:1)", 
                                        "SM(d18:2/24:1)", "Sphinganine-1P", "Sphinganine", 
                                        "Sphingosine(d16:1)", "Sphingosine", "Sphingosine(d18:2)",
                                        "3-Dehydrosphinganine")
```

### 1.2.2 microbial metabolite
unit: ug/ml
```{r}
link_id <- fread(str_c(pegasus_dir, "targeted_microb/list_of_BIOBANK_samples.csv"))
setnames(link_id, "SUBJ_ALIASID", "Biobank_Subject_ID")
# link_id[, .N, Biobank_Subject_ID][N == 2] 
# Remvoe duplicated records for this subject
link_id <- link_id[!duplicated(Biobank_Subject_ID), ]

microb_mets <- read_xlsx(str_c(pegasus_dir, "targeted_microb/PR003-21_AsthmaCohort1125_MicrobiomePanel_CDT_Test.xlsx")) %>% as.data.table()
setnames(microb_mets, gsub("[ ()/]", "_", colnames(microb_mets)))
microb_mets[, feat_id := str_c("X", seq(1, .N), "_microb")]
setcolorder(microb_mets, "feat_id")
microb_mets_cols <- microb_mets_info$feat_id
microb_mets_cols <- microb_mets_cols[!microb_mets_cols %in% "trypkynu_ratio_microb"]

subj_cols <- colnames(microb_mets)[-c(2:which(colnames(microb_mets) == "ALOQ__ug_mL_"))]
microb_mets_dat <- microb_mets[, ..subj_cols] %>% as.data.frame() 
rownames(microb_mets_dat) <- microb_mets_dat$feat_id
microb_mets_dat$feat_id <- NULL
microb_mets_dat <- t(microb_mets_dat) %>% as.data.frame() %>% rownames_to_column() %>% as.data.table()
setnames(microb_mets_dat, "rowname", "S_SAMPLEID")

microb_mets_dat[, (microb_mets_cols) := lapply(.SD, as.character), .SDcols = microb_mets_cols]
suppressWarnings(microb_mets_dat[, (microb_mets_cols) := lapply(.SD, as.numeric), .SDcols = microb_mets_cols])
microb_mets_dat <- merge(microb_mets_dat, link_id[, .(Biobank_Subject_ID, S_SAMPLEID)], by = "S_SAMPLEID")

microb_mets_dat %>% setcolorder(., "Biobank_Subject_ID")
microb_mets_dat <- microb_mets_dat[,-2]

```

### 1.2.3 steroid
unit: ng/ml
```{r}
# link id
link_id <- read_xlsx(str_c(pegasus_dir, "targeted_steroids/BIOBANK_Serum_manifest.xlsx"), sheet = 1) %>% as.data.table()
link_id <- link_id[,c(3,5)]
colnames(link_id) <- c("sampleID_biobank", "Biobank_Subject_ID")
link_id$Biobank_Subject_ID <- as.integer(as.character(link_id$Biobank_Subject_ID))

steroid_df <- read_xlsx(str_c(pegasus_dir, "targeted_steroids/PR005-21_AsthmaCohort1125_Steroids_CDT_062821.xlsx"), sheet = 1) %>% as.data.table()

# add a feat id and feature
steroid_df[, ':=' (feature = str_c("S", seq(1, .N)),
                   feat_id = str_c("S", seq(1, .N), "_steroid"))]

# transpose the steroid data
steroid_dat <- steroid_df[, c(11:1135)]
steroid_dat <- as.data.frame(t(steroid_dat))
colnames(steroid_dat) <- steroid_df$feat_id

# change NQ to be NA
steroid_dat[steroid_dat == "NQ"] <- NA

#
steroid_dat[] <- lapply(steroid_dat, function(x) as.numeric(as.character(x)))
steroid_dat <- as.data.frame(cbind(rownames(steroid_dat), steroid_dat))
colnames(steroid_dat)[1] <- "sampleID_biobank"
steroid_dat <- as.data.table(steroid_dat)
steroid_dat <- merge(link_id, steroid_dat, by = "sampleID_biobank", all.x = T)

steroid_dat <- steroid_dat[,-1]

## subject 10009114 has duplicated steroid measure so I remove this subjects
steroid_dat <- steroid_dat[!(Biobank_Subject_ID == 10009114),]

```

# 2. Create targeted ratio
## 2.1 targeted sph/targeted microb
```{r}
df <- merge(sph_dat, microb_mets_dat, by = "Biobank_Subject_ID", all = T) %>% as.data.frame()
rownames(df) <- df$Biobank_Subject_ID
df <- df[,-1]

sph_list <- colnames(sph_dat)[-1]
microb_list <- colnames(microb_mets_dat)[-1]

df_ratios <- df

for (i in 1:length(sph_list)) {
  for (j in 1:length(microb_list)) {
    sph <- sph_list[i]
    microb <- microb_list[j]
    eval(parse(text = str_c("df_ratios$", sph, ".", microb ," <- df_ratios$", sph, "/", "df_ratios$", microb)))
  }
}

df_ratios <- df_ratios[,-which(colnames(df_ratios) %in% colnames(df))]
ncol(df_ratios) #5460

df_sph_microb_ratio <- df_ratios
df_sph_microb_ratio$Biobank_Subject_ID <- rownames(df_sph_microb_ratio)
df_sph_microb_ratio %>% setcolorder(., "Biobank_Subject_ID")
df_sph_microb_ratio <- as.data.table(df_sph_microb_ratio)
df_sph_microb_ratio_log <- df_sph_microb_ratio
df_sph_microb_ratio_log[,c(2:5461)] <- log2(df_sph_microb_ratio_log[,c(2:5461)])

df_sph_microb_ratio_log <- df_sph_microb_ratio_log %>% mutate_if(is.numeric, ~ replace(., !is.finite(.), NA))

# missing
mets_missing_sph_microb <- df_sph_microb_ratio_log[, lapply(.SD, function(x) mean(is.na(x))), .SDcols = colnames(df_sph_microb_ratio_log)[2:ncol(df_sph_microb_ratio_log)]] %>% 
        t() %>% as.data.frame() %>% rownames_to_column()
setnames(mets_missing_sph_microb, c("feat_id", "pct_na_met"))
summary(mets_missing_sph_microb$pct_na_met)
ggplot(mets_missing_sph_microb) + 
        geom_histogram(aes(pct_na_met), bins = 50) + 
        labs(title = "targeted sph/targeted microb ratio missingness")


ggplot(df_sph_microb_ratio, aes(x=Cer24_0d16_1.X41_microb)) + geom_histogram(color="darkblue", fill="lightblue")
ggplot(df_sph_microb_ratio_log, aes(x=Cer24_0d16_1.X41_microb)) + geom_histogram(color="darkblue", fill="lightblue")


```

## 2.2 targeted sph/targeted str
```{r}
df <- merge(sph_dat, steroid_dat, by = "Biobank_Subject_ID", all = T) %>% as.data.frame()
rownames(df) <- df$Biobank_Subject_ID
df <- df[,-1]

steroid_list <- colnames(steroid_dat)[-1]

df_ratios <- df

for (i in 1:length(sph_list)) {
  for (j in 1:length(steroid_list)) {
    sph <- sph_list[i]
    steroid <- steroid_list[j]
    eval(parse(text = str_c("df_ratios$", sph, ".", steroid ," <- df_ratios$", sph, "/", "df_ratios$", steroid)))
  }
}

df_ratios <- df_ratios[,-which(colnames(df_ratios) %in% colnames(df))]
ncol(df_ratios) #1404

df_sph_steroid_ratio <- df_ratios
df_sph_steroid_ratio$Biobank_Subject_ID <- rownames(df_sph_steroid_ratio)
df_sph_steroid_ratio %>% setcolorder(., "Biobank_Subject_ID")
df_sph_steroid_ratio <- as.data.table(df_sph_steroid_ratio)
df_sph_steroid_ratio_log <- df_sph_steroid_ratio
df_sph_steroid_ratio_log[,c(2:1404)] <- log2(df_sph_steroid_ratio_log[,c(2:1404)])

df_sph_steroid_ratio_log <- df_sph_steroid_ratio_log %>% mutate_if(is.numeric, ~ replace(., !is.finite(.), NA))

# missing
mets_missing_sph_str <- df_sph_steroid_ratio_log[, lapply(.SD, function(x) mean(is.na(x))), .SDcols = colnames(df_sph_steroid_ratio_log)[2:ncol(df_sph_steroid_ratio_log)]] %>% 
        t() %>% as.data.frame() %>% rownames_to_column()
setnames(mets_missing_sph_str, c("feat_id", "pct_na_met"))
summary(mets_missing_sph_str$pct_na_met)
ggplot(mets_missing_sph_str) + 
        geom_histogram(aes(pct_na_met), bins = 50) + 
        labs(title = "targeted sph/targeted str ratio missingness")

ggplot(df_sph_steroid_ratio, aes(x=LacCer14_0.S10_steroid	)) + geom_histogram(color="darkblue", fill="lightblue")
ggplot(df_sph_steroid_ratio_log, aes(x=LacCer14_0.S10_steroid	)) + geom_histogram(color="darkblue", fill="lightblue")

```

## 2.3 targeted microb/targeted str
```{r}
df <- merge(microb_mets_dat, steroid_dat, by = "Biobank_Subject_ID", all = T) %>% as.data.frame()
rownames(df) <- df$Biobank_Subject_ID
df <- df[,-1]

df_ratios <- df

for (i in 1:length(microb_list)) {
  for (j in 1:length(steroid_list)) {
    microb <- microb_list[i]
    steroid <- steroid_list[j]
    eval(parse(text = str_c("df_ratios$", microb, ".", steroid ," <- df_ratios$", microb, "/", "df_ratios$", steroid)))
  }
}

df_ratios <- df_ratios[,-which(colnames(df_ratios) %in% colnames(df))]
ncol(df_ratios) #1260

df_microb_steroid_ratio <- df_ratios
df_microb_steroid_ratio$Biobank_Subject_ID <- rownames(df_microb_steroid_ratio)
df_microb_steroid_ratio %>% setcolorder(., "Biobank_Subject_ID")
df_microb_steroid_ratio <- as.data.table(df_microb_steroid_ratio)
df_microb_steroid_ratio_log <- df_microb_steroid_ratio
df_microb_steroid_ratio_log[,c(2:1261)] <- log2(df_microb_steroid_ratio_log[,c(2:1261)])

df_microb_steroid_ratio_log <- df_microb_steroid_ratio_log %>% mutate_if(is.numeric, ~ replace(., !is.finite(.), NA))

# missing
mets_missing_microb_str <- df_microb_steroid_ratio_log[, lapply(.SD, function(x) mean(is.na(x))), .SDcols = colnames(df_microb_steroid_ratio_log)[2:ncol(df_microb_steroid_ratio_log)]] %>% 
        t() %>% as.data.frame() %>% rownames_to_column()
setnames(mets_missing_microb_str, c("feat_id", "pct_na_met"))
summary(mets_missing_microb_str$pct_na_met)
ggplot(mets_missing_microb_str) + 
        geom_histogram(aes(pct_na_met), bins = 50) + 
        labs(title = "targeted microb/targeted str ratio missingness")

```

# 3 Trim data
- _trim: subjects missing bmi & matched subjects removed  

```{r}
nrow(pheno_mets) #1132

stratum_to_rm <- pheno_mets[is.na(bmi), unique(Stratum)]
stratum_to_rm <- c(stratum_to_rm, pheno_mets[, .N, Stratum][N == 1, Stratum])
print(str_c("Number of strata removed due to missing BMI: ", 
            length(stratum_to_rm)))
pheno_mets_trim <- pheno_mets[!(Stratum %in% stratum_to_rm), ]

```

# 4. Regression model functions
## 4.1 linear model
```{r}
linreg_res_fnc <- function(outc, mets_list, dat, mets_info, add_covar = "", 
                           first_cols = c("feat_id", "outc", "pval", "beta", "lower95", "upper95", "fdr", "fdr_bh"), 
                           na_thld = missing_thld) {
        
        linear_mdl_tmp <- function(met) {
                
                fit <- eval(parse(text = str_c("lm(", outc, " ~ ", met, 
                                               add_covar, ", data = ", dat, ")")))
                
                coef <- summary(fit)$coefficients %>% as.data.frame() %>% rownames_to_column() %>% as.data.table()
                coef <- coef[rowname == met, ]

                output <- cbind(outc, coef)
                colnames(output) <- c("outc", "feat_id", "beta", "se", "tval", "pval")
                output[, (colnames(output)[-c(1:2)]) := lapply(.SD, as.numeric), .SDcols = colnames(output)[-c(1:2)]]
                
                return(output)
        }
        
        res <- lapply(mets_list, function(x) {linear_mdl_tmp(met = x)}) %>% rbindlist()
        
        # eval(parse(text = str_c("mets_info <- copy(", mets_info, ")")))
        # res <- merge(res, mets_info, by = "feat_id")
        res[, ':='(fdr = p.adjust(pval, method = "fdr"), 
                   fdr_bh = p.adjust(pval, method = "BH"))]
        res[, totaln := eval(parse(text = str_c(dat, "[!is.na(", outc, "), .N]")))]
        res[, ':='(lower95 = beta - z_95 * se, 
                   upper95 = beta + z_95 * se)]
        setcolorder(res, first_cols)
        res <- res[order(pval)]
        return(res)
}

```

## 4.2 conditional logistic
```{r}
clogi_res_fnc <- function(outc, mets_list, dat, mets_info, add_covar = "", 
                          first_cols = c("feat_id", 
                                         "beta", "pval", "or", "fdr", "fdr_bh", "lower95", "upper95")) {
        
        clogi_mdl_tmp <- function(met) {
                
                fit <- eval(parse(text = str_c("clogit(", outc, " ~ ", met, " + strata(Stratum)", 
                                               add_covar, ", data = ", dat, ")")))
                
                coef <- summary(fit)$coefficients
                coef <- coef[rownames(coef) == met, ]

                totaln <- summary(fit)$n
                casen <- summary(fit)$nevent
                
                output <- t(c(outc, met, coef, totaln, casen)) %>% as.data.table()
                colnames(output) <- c("outc", "feat_id", "beta", "or", "se", "zval", "pval", 
                                      "totaln", "casen")
                output[, (colnames(output)[-c(1:2)]) := lapply(.SD, as.numeric), .SDcols = colnames(output)[-c(1:2)]]
                
                return(output)
        }
        
        res <- lapply(mets_list, function(x) {clogi_mdl_tmp(met = x)}) %>% rbindlist()
        
        # eval(parse(text = str_c("mets_info <- copy(", mets_info, ")")))
        # res <- merge(res, mets_info, by = "feat_id")
        res[, ':='(fdr = p.adjust(pval, method = "fdr"), 
                   fdr_bh = p.adjust(pval, method = "BH"))]
        res[, ':='(lower95 = exp(beta - z_95 * se), 
                   upper95 = exp(beta + z_95 * se))] 
        
        setcolorder(res, first_cols)
        res <- res[order(pval)]
        return(res)
}

```

## 4.3 poisson regression
```{r}
poiss_res_fnc <- function(outc, mets_list, dat, add_covar = "",
                         first_cols = c("feat_id", 
                                        "outc", "beta", "pval", "fdr")) {
        
        logi_mdl_tmp <- function(met) {
                
                fit <- eval(parse(text = str_c("glm(", outc, " ~ ", met, add_covar, ", data = ", dat, ", family = 'quasipoisson')")))
                
                coef <- summary(fit)$coefficients %>% as.data.frame() %>% rownames_to_column() %>% as.data.table()
                coef <- coef[rowname == met, ]

                output <- cbind(outc, coef)
                colnames(output) <- c("outc", "feat_id", "beta", "se", "zval", "pval")
                output[, (colnames(output)[-c(1:2)]) := lapply(.SD, as.numeric), .SDcols = colnames(output)[-c(1:2)]]
                
                return(output)
        }
        
        res <- lapply(mets_list, function(x) {logi_mdl_tmp(met = x)}) %>% rbindlist()
        
        # eval(parse(text = str_c("mets_info <- copy(", mets_info, ")")))
        # res <- merge(res, mets_info, by = "feat_id", all.x = T)
        res[, ':='(neg_log10_pval = -log10(pval), 
                   fdr = p.adjust(pval, method = "fdr"))]
        res[, ':='(lower95 = beta - z_95 * se, 
                   upper95 = beta + z_95 * se)]
        res[, totaln := eval(parse(text = str_c(dat, "[!is.na(", outc, "), .N]")))]
        
        setcolorder(res, first_cols)
        res <- res[order(pval)]
        return(res)
}

```

## 4.4 Manhattan Plot
```{r}
save_manh_plot <- function(dat, title_infig, fig_name, wdth = 20, ht = 12, pointsize = 3, pointstroke = 1.5,
                           transparency = 1, if_lbl = c(T, F), lbl_thld = 2e-04) {
  
        p <- ggplot(dat, aes(x = Name, y = -log10(pval))) + 
                    facet_grid(. ~ denominator_sub_pathway, scales = "free_x") +  
                    geom_point(aes(color = numerator_sub_pathway , shape = direction_of_effect), 
                               size = pointsize, stroke = pointstroke, alpha = transparency) + 
                    scale_shape_manual(values = c("-" = 6, "+" = 2)) + 
                    labs(title = title_infig, 
                         y = expression('-log'[10]*'('*italic(P)*'-value)')) + 
                    theme(legend.position = "bottom", 
                          legend.text = element_text(size = 16), 
                          plot.title = element_text(hjust = 0.5, size = 20), 
                          strip.text.x = element_text(size = 20), 
                          axis.title.x = element_blank(),
                          axis.text.x = element_blank(), 
                          axis.ticks.x = element_blank(), 
                          axis.title.y = element_text(size = 20, face = "bold"), 
                          axis.text.y = element_text(size = 24, face = "bold"),
                          panel.border = element_blank(),
                          panel.grid.major = element_blank(),
                          panel.grid.minor = element_blank())
        if (if_lbl == T) {
                p <- p + geom_label_repel(aes(label = ifelse(pval < lbl_thld, CHEMICAL_NAME, "")),
                                          box.padding   = 0.25,
                                          point.padding = 0.5,
                                          segment.color = 'grey50',
                                          max.overlaps = 80)
        }
        ggsave(file = here::here(fig_dir, fig_name), plot = p, width = wdth, height = ht)
}

```


# 5. Conduct association analysis
## 5.1 Asthma
### 5.1.1 targeted sph/microb
```{r}
list_analysis <- mets_missing_sph_microb[mets_missing_sph_microb$pct_na_met <= 0.3,"feat_id"]
length(list_analysis) #3898

list_tmp <- append("Biobank_Subject_ID", list_analysis)
df_sph_microb_ratio_log$Biobank_Subject_ID <- as.integer(df_sph_microb_ratio_log$Biobank_Subject_ID)
pheno_tmp <- merge(pheno_mets_trim, df_sph_microb_ratio_log[,..list_tmp], by = "Biobank_Subject_ID", all.x = T)

res_5.1.1 <- clogi_res_fnc(outc = "asthma_num", 
                           mets_list = list_analysis, 
                           dat = "pheno_tmp", 
                           add_covar = " + bmi")

## Save results
res_asthma_sph_microb_ratio <- res_5.1.1
res_asthma_sph_microb_ratio <- res_asthma_sph_microb_ratio %>% separate(feat_id, sep = "\\.", c("numerator", "denominator"), remove = F)

dict <- sph_mets_info[,c("feat_id", "tentative_annotation", "SUB PATHWAY", "SUPER PATHWAY")]
res_asthma_sph_microb_ratio$numerator <- dict$tentative_annotation[match(res_asthma_sph_microb_ratio$numerator, dict$feat_id)]
res_asthma_sph_microb_ratio$numerator_sub_pathway <- dict$`SUB PATHWAY`[match(res_asthma_sph_microb_ratio$numerator, dict$tentative_annotation)]

dict <- microb_mets_info[,c("feat_id", "Component_Name", "SUB_PATHWAY", "SUPER_PATHWAY")]
res_asthma_sph_microb_ratio$denominator <- dict$Component_Name[match(res_asthma_sph_microb_ratio$denominator, dict$feat_id)]
res_asthma_sph_microb_ratio$denominator_sub_pathway <- dict$`SUB_PATHWAY`[match(res_asthma_sph_microb_ratio$denominator, dict$`Component_Name`)]

setcolorder(res_asthma_sph_microb_ratio, c("feat_id", "numerator", "numerator_sub_pathway", "denominator", "denominator_sub_pathway"))

saveRDS(res_asthma_sph_microb_ratio, file = here(res_dir, "res_asthma_sph_microb_ratio.rds"))

## display results
### min p-value = 1.85e-7
### #sig = 411/3898
datatable(res_asthma_sph_microb_ratio[res_asthma_sph_microb_ratio$fdr <= 0.05, ], 
          filter = "top") %>% 
          formatSignif(columns = c("beta", "pval", "or", "fdr", "fdr_bh", "lower95", "upper95", "se", "zval"), digits = 3)

res_5.1.1_sig <- res_asthma_sph_microb_ratio[res_asthma_sph_microb_ratio$fdr <= 0.05, ]

res_5.1.1_sig_top20 <- res_5.1.1_sig[c(1:20),] %>% as.data.table()
res_5.1.1_sig_top20$annotation <- paste(res_5.1.1_sig_top20$numerator, res_5.1.1_sig_top20$denominator, sep = " / ")
res_5.1.1_sig_top20$class <- "sph/microb"
res_5.1.1_sig_top20$Pval <- signif(res_5.1.1_sig_top20$pval, 3)
res_5.1.1_sig_top20 <- res_5.1.1_sig_top20[order(denominator_sub_pathway, or), ]


p_sph_microb_asthma <- ggplot(res_5.1.1_sig_top20, 
                aes(reorder(annotation, -or), y = or, color = or >1)) +
   scale_colour_manual(name = '', values = setNames(c('red','blue'),c(T, F)))+
   geom_rect(aes(fill = denominator_sub_pathway),xmin = -Inf,xmax = Inf,
             ymin = -Inf,ymax = Inf, alpha = 0, color = 'black', linejoin = "round") +
   facet_grid(denominator_sub_pathway ~ class, scales = "free_y", space = "free_y", labeller = label_wrap_gen()) +
   geom_errorbar(aes(ymin = lower95, ymax = upper95), width = 0.25) +
   geom_pointrange(aes(ymin = lower95, ymax = upper95), size = 0.25) +
   geom_hline(yintercept = 1, linetype = 2) +
   labs(
     x = "",
     y = ""
   ) +
   theme_minimal() +
   theme_bw() +
   theme(
     title = element_text(size = 16),
     strip.text.x = element_text(size = 12, face = "bold"),
     strip.text.y = element_text(size = 10, face = "bold"),
     axis.title = element_text(size = 12),
     axis.text = element_text(size = 10, face = "bold"),
     legend.position = "",
     legend.title = element_text(size = 12),
     legend.text = element_text(size = 12),
     strip.text.y.right = element_text(angle = 0),
     strip.background = element_rect(fill = "#B2ABD2", colour = "black")) +
   coord_flip() + scale_fill_viridis_d(option = "D", direction = -1, begin = 0, end = 1, alpha = 0.1)
   # scale_y_continuous(limits = c(0.2, 1.6))

```

### 5.1.2 targeted sph/str
```{r}
list_analysis <- mets_missing_sph_str[mets_missing_sph_str$pct_na_met <= 0.3,"feat_id"]
length(list_analysis) #1248

list_tmp <- append("Biobank_Subject_ID", list_analysis)
df_sph_steroid_ratio_log$Biobank_Subject_ID <- as.integer(df_sph_steroid_ratio_log$Biobank_Subject_ID)
pheno_tmp <- merge(pheno_mets_trim, df_sph_steroid_ratio_log[,..list_tmp], by = "Biobank_Subject_ID", all.x = T)

res_5.1.2 <- clogi_res_fnc(outc = "asthma_num", 
                           mets_list = list_analysis, 
                           dat = "pheno_tmp", 
                           add_covar = " + bmi")

## Save results
res_asthma_sph_str_ratio <- res_5.1.2
res_asthma_sph_str_ratio <- res_asthma_sph_str_ratio %>% separate(feat_id, sep = "\\.", c("numerator", "denominator"), remove = F)

dict <- sph_mets_info[,c("feat_id", "tentative_annotation", "SUB PATHWAY", "SUPER PATHWAY")]
res_asthma_sph_str_ratio$numerator <- dict$tentative_annotation[match(res_asthma_sph_str_ratio$numerator, dict$feat_id)]
res_asthma_sph_str_ratio$numerator_sub_pathway <- dict$`SUB PATHWAY`[match(res_asthma_sph_str_ratio$numerator, dict$tentative_annotation)]

dict <- str_mets_info[,c("feat_id", "Component Name", "SUB PATHWAY", "SUPER PATHWAY")]
res_asthma_sph_str_ratio$denominator <- dict$`Component Name`[match(res_asthma_sph_str_ratio$denominator, dict$feat_id)]
res_asthma_sph_str_ratio$denominator_sub_pathway <- dict$`SUB PATHWAY`[match(res_asthma_sph_str_ratio$denominator, dict$`Component Name`)]

setcolorder(res_asthma_sph_str_ratio, c("feat_id", "numerator", "numerator_sub_pathway", "denominator", "denominator_sub_pathway"))


saveRDS(res_asthma_sph_str_ratio, file = here(res_dir, "res_asthma_sph_str_ratio.rds"))

## display results
### min p-value = 3.86e-16
### #sig = 592/1248
datatable(res_asthma_sph_str_ratio[res_asthma_sph_str_ratio$fdr <= 0.05,], 
          filter = "top") %>% 
          formatSignif(columns = c("beta", "pval", "or", "fdr", "fdr_bh", "lower95", "upper95", "se", "zval"), digits = 3)

res_5.1.2_sig <- res_asthma_sph_str_ratio[res_asthma_sph_str_ratio$fdr <= 0.05,]

res_5.1.2_sig_top20 <- res_5.1.2_sig[c(1:20),] %>% as.data.table()
res_5.1.2_sig_top20$denominator[res_5.1.2_sig_top20$denominator == "dehydroepiandrosterone sulfate"] <- "DHEAS"
res_5.1.2_sig_top20$annotation <- paste(res_5.1.2_sig_top20$numerator, res_5.1.2_sig_top20$denominator, sep = " / ")
res_5.1.2_sig_top20$class <- "sph/str"
res_5.1.2_sig_top20$Pval <- signif(res_5.1.2_sig_top20$pval, 3)
res_5.1.2_sig_top20 <- res_5.1.2_sig_top20[order(denominator_sub_pathway, or), ]


p_sph_str_asthma <- ggplot(res_5.1.2_sig_top20, 
                aes(reorder(annotation, -or), y = or, color = or >1)) +
   scale_colour_manual(name = '', values = setNames(c('red','blue'),c(T, F)))+
   geom_rect(aes(fill = denominator_sub_pathway),xmin = -Inf,xmax = Inf,
             ymin = -Inf,ymax = Inf, alpha = 0, color = 'black', linejoin = "round") +
   facet_grid(denominator_sub_pathway ~ class, scales = "free_y", space = "free_y", labeller = label_wrap_gen()) +
   geom_errorbar(aes(ymin = lower95, ymax = upper95), width = 0.25) +
   geom_pointrange(aes(ymin = lower95, ymax = upper95), size = 0.25) +
   geom_hline(yintercept = 1, linetype = 2) +
   labs(
     x = "",
     y = ""
   ) +
   theme_minimal() +
   theme_bw() +
   theme(
     title = element_text(size = 16),
     strip.text.x = element_text(size = 12, face = "bold"),
     strip.text.y = element_text(size = 10, face = "bold"),
     axis.title = element_text(size = 12),
     axis.text = element_text(size = 10, face = "bold"),
     legend.position = "",
     legend.title = element_text(size = 12),
     legend.text = element_text(size = 12),
     strip.text.y.right = element_text(angle = 0),
     strip.background = element_rect(fill = "#FDB863", colour = "black")) +
   coord_flip() + scale_fill_viridis_d(option = "D", direction = -1, begin = 0, end = 1, alpha = 0.1)
   # scale_y_continuous(limits = c(0.2, 1.6))

```


### 5.1.3 targeted microb/str
```{r}
list_analysis <- mets_missing_microb_str[mets_missing_microb_str$pct_na_met <= 0.3,"feat_id"]
length(list_analysis) #794

list_tmp <- append("Biobank_Subject_ID", list_analysis)
df_microb_steroid_ratio_log$Biobank_Subject_ID <- as.integer(df_microb_steroid_ratio_log$Biobank_Subject_ID)
pheno_tmp <- merge(pheno_mets_trim, df_microb_steroid_ratio_log[,..list_tmp], by = "Biobank_Subject_ID", all.x = T)

res_5.1.3 <- clogi_res_fnc(outc = "asthma_num",
                           mets_list = list_analysis, 
                           dat = "pheno_tmp",
                           add_covar = " + bmi")

## Save results
res_asthma_microb_str_ratio <- res_5.1.3
res_asthma_microb_str_ratio <- res_asthma_microb_str_ratio %>% separate(feat_id, sep = "\\.", c("numerator", "denominator"), remove = F)

dict <- microb_mets_info[,c("feat_id", "Component_Name", "SUB_PATHWAY", "SUPER_PATHWAY")]
res_asthma_microb_str_ratio$numerator <- dict$Component_Name[match(res_asthma_microb_str_ratio$numerator, dict$feat_id)]
res_asthma_microb_str_ratio$numerator_sub_pathway <- dict$SUB_PATHWAY[match(res_asthma_microb_str_ratio$numerator, dict$Component_Name)]

dict <- str_mets_info[,c("feat_id", "Component Name", "SUB PATHWAY", "SUPER PATHWAY")]
res_asthma_microb_str_ratio$denominator <- dict$`Component Name`[match(res_asthma_microb_str_ratio$denominator, dict$feat_id)]
res_asthma_microb_str_ratio$denominator_sub_pathway <- dict$`SUB PATHWAY`[match(res_asthma_microb_str_ratio$denominator, dict$`Component Name`)]

setcolorder(res_asthma_microb_str_ratio, c("feat_id", "numerator", "numerator_sub_pathway", "denominator", "denominator_sub_pathway"))


saveRDS(res_asthma_microb_str_ratio, file = here(res_dir, "res_asthma_microb_str_ratio.rds"))

## display results
### min p-value = 2.37e-13
### #sig = 281/794
datatable((res_asthma_microb_str_ratio[res_asthma_microb_str_ratio$fdr <= 0.05, ]), 
          filter = "top") %>% 
          formatSignif(columns = c("beta", "pval", "or", "fdr", "fdr_bh", "lower95", "upper95", "se", "zval"), digits = 3)

res_5.1.3_sig <- res_asthma_microb_str_ratio[res_asthma_microb_str_ratio$fdr <= 0.05, ]

res_5.1.3_sig_top20 <- res_5.1.3_sig[c(1:20),] %>% as.data.table()
res_5.1.3_sig_top20$denominator[res_5.1.3_sig_top20$denominator == "dehydroepiandrosterone sulfate"] <- "DHEAS"
res_5.1.3_sig_top20$annotation <- paste(res_5.1.3_sig_top20$numerator, res_5.1.3_sig_top20$denominator, sep = " / ")
res_5.1.3_sig_top20$class <- "microb/str"
res_5.1.3_sig_top20$Pval <- signif(res_5.1.3_sig_top20$pval, 3)
res_5.1.3_sig_top20 <- res_5.1.3_sig_top20[order(denominator_sub_pathway, or), ]

p_microb_str_asthma <- ggplot(res_5.1.3_sig_top20, 
                aes(reorder(annotation, -or), y = or, color = or >1)) +
   scale_colour_manual(name = '', values = setNames(c('red','blue'),c(T, F)))+
   geom_rect(aes(fill = denominator_sub_pathway),xmin = -Inf,xmax = Inf,
             ymin = -Inf,ymax = Inf, alpha = 0, color = 'black', linejoin = "round") +
   facet_grid(denominator_sub_pathway ~ class, scales = "free_y", space = "free_y", labeller = label_wrap_gen()) +
   geom_errorbar(aes(ymin = lower95, ymax = upper95), width = 0.25) +
   geom_pointrange(aes(ymin = lower95, ymax = upper95), size = 0.25) +
   geom_hline(yintercept = 1, linetype = 2) +
   labs(
     x = "",
     y = ""
   ) +
   theme_minimal() +
   theme_bw() +
   theme(
     title = element_text(size = 16),
     strip.text.x = element_text(size = 12, face = "bold"),
     strip.text.y = element_text(size = 10, face = "bold"),
     axis.title = element_text(size = 12),
     axis.text = element_text(size = 10, face = "bold"),
     legend.position = "",
     legend.title = element_text(size = 12),
     legend.text = element_text(size = 12),
     strip.text.y.right = element_text(angle = 0),
     strip.background = element_rect(fill = "#B8E186", colour = "black")) +
   coord_flip() + scale_fill_viridis_d(option = "D", direction = -1, begin = 0, end = 1, alpha = 0.1)
   # scale_y_continuous(limits = c(0.2, 1.6))

```

## 5.2 Within Asthma
```{r}
dat_tmp <- pheno_mets_trim[Asthma_Status == TRUE,]
# dat_tmp$Biobank_Subject_ID <- as.character(dat_tmp$Biobank_Subject_ID)
```

### 5.2.1 sph/microb: H-cluster + heatmap 

### 5.2.2 sph/str: H-cluster + heatmap
```{r}
# ratio with missing percentatge <= 30%
list_analysis <- mets_missing_sph_str[mets_missing_sph_str$pct_na_met <= 0.3,"feat_id"]
length(list_analysis) #1248

list_tmp <- append("Biobank_Subject_ID", list_analysis)
pheno_tmp <- merge(dat_tmp, df_sph_steroid_ratio_log[,..list_tmp], by = "Biobank_Subject_ID", all.x = T)

# get the ratio data
ratio_column <- c("Biobank_Subject_ID", list_analysis)
ratio_feat_data = pheno_tmp[, ..ratio_column] # n = 540
ratio_feat_data <- ratio_feat_data[complete_cases(ratio_feat_data),] # n = 383
ratio_subject_ID <- ratio_feat_data$Biobank_Subject_ID

# Standardize data for clustering
ratio_feat_data[, (list_analysis) := lapply(.SD, function(x) scale(x, center = T, scale = T)), .SDcols = list_analysis]

clust_df <- as.data.frame(ratio_feat_data)
rownames(clust_df) <- clust_df$Biobank_Subject_ID
clust_df <- clust_df[,-1]
clust_df <- transpose(clust_df)

colnames(clust_df) <- ratio_subject_ID
rownames(clust_df) <- list_analysis

# Compare linkage methods for agglomerative clustering using agnes
methods <- c( "average", "single", "complete", "ward")
names(methods) <- c( "average", "single", "complete", "ward")

# function to compute coefficient
ac <- function(x) {
  agnes(clust_df, diss = FALSE, method = x)$ac
}
map_dbl(methods, ac) # Ward gives best coefficient


# Elbow method
fviz_nbclust(clust_df, FUNcluster = hcut, method = "wss", k.max = 50) +
             labs(subtitle = "Elbow method")
#ggsave("figures/elbow_method.png")

# Silhouette method
fviz_nbclust(clust_df, FUNcluster = hcut, method = "silhouette", k.max = 30) +
  geom_vline(xintercept = 3, linetype = 2, colour = "steelblue") +
             labs(subtitle = "Silhouette method")
#ggsave("figures/silhouette_method.png")

# Gap stat method
fviz_nbclust(clust_df, FUNcluster = hcut, method = "gap_stat", k.max = 30, nboot = 10) +
  geom_vline(xintercept = 3, linetype = 2, colour = "steelblue") +
             labs(subtitle = "Gap Statistic method")
#ggsave("figures/gapstat_method.png")

# See how many in each group
hclust_ac <- agnes(clust_df, metric = "euclidean", method = "ward")
clust_ac <- cutree(hclust_ac, k = 15) # change k to optimal
table(clust_ac)

# Enhanced hierarchical clustering
res.hc <- eclust(clust_df, "hclust", k = 15,
                 method = "ward", graph = FALSE)
fviz_dend(res.hc, rect = TRUE, show_labels = FALSE,
          k_colors = c("#00BA38", "#619CFF", "#F8766D")
          ) # this graph takes a long time, manually order colors to match scatterplot
#ggsave("figures/dendrogram_with_labels.png")

# Visualize scatterplot using factoextra
fviz_cluster(list(data = clust_df, cluster = clust_ac), geom = "point")
#ggsave("figures/cluster_scatterplot.png")

clust_df_mem <- as.data.frame(clust_ac)
clust_df_mem <- as.data.table(cbind(rownames(clust_df_mem), clust_df_mem))
colnames(clust_df_mem) <- c("feat_id", "H_cluster_mem")

## add steroid information
clust_df_mem <- clust_df_mem %>% separate(feat_id, sep = "\\.", c("numerator", "denominator"), remove = F)

dict <- sph_mets_info[,c("feat_id", "SUB PATHWAY", "SUPER PATHWAY")]
clust_df_mem$numerator_sub_pathway <- dict$`SUB PATHWAY`[match(clust_df_mem$numerator, dict$feat_id)]

dict <- str_mets_info[,c("feat_id", "Component Name", "SUB PATHWAY", "SUPER PATHWAY")]
clust_df_mem$denominator <- dict$`Component Name`[match(clust_df_mem$denominator, dict$feat_id)]
clust_df_mem$denominator_sub_pathway <- dict$`SUB PATHWAY`[match(clust_df_mem$denominator, dict$`Component Name`)]

setcolorder(clust_df_mem, c("feat_id", "numerator", "numerator_sub_pathway", "denominator", "denominator_sub_pathway"))

## display results
clust_df_mem$H_cluster_mem <- as.character(as.integer(clust_df_mem$H_cluster_mem))
datatable(clust_df_mem, 
          filter = "top")

## heatmap
ratio_cor_df <- as.data.frame(t(clust_df))
ratio_cor_df <- ratio_cor_df %>% 
                as.matrix %>%
                cor

# Perform hierarchical (**thus NOT kmeans; see comment below**) clustering 'outside' ComplexHeatmap using hclust

hr <- hclust(dist(ratio_cor_df, method = "euclidean"), method="ward.D2") 
hc <- hclust(dist(t(ratio_cor_df), method = "euclidean"), method="ward.D2")

# define 15 row clusters using cutree outside ComplexHeatmap based on previous study
mycl.r <- cutree(hr, k=15)

annotation_row = data.frame(Cluster = factor(mycl.r))

png(file=str_c(fig_dir, "sph_str_cor_pheatmap.png"), width = 18, height = 18, units = "in", res = 300)
pheatmap(ratio_cor_df, scale = "none", show_rownames=FALSE, show_colnames=FALSE,
	cluster_rows=hr, cluster_cols=hc,
	annotation_row=annotation_row, main = "Ratio of sphingolipid/Steroid correlation",
	treeheight_row = 0)
dev.off()

```


### 5.2.3 microb/str: H-cluster + heatmap


## 5.3 Log(IgE)
```{r}
dat_tmp$log_ige <- log(dat_tmp$ige)
sum(is.na(dat_tmp$log_ige)) #363
ggplot(dat_tmp, aes(x=log_ige)) + geom_histogram(color="darkblue", fill="lightblue")

```

### 5.3.1 targeted sph/microb
```{r}
list_analysis <- mets_missing_sph_microb[mets_missing_sph_microb$pct_na_met <= 0.3,"feat_id"]
length(list_analysis)

list_tmp <- append("Biobank_Subject_ID", list_analysis)
pheno_tmp <- merge(dat_tmp, df_sph_microb_ratio_log[,..list_tmp], by = "Biobank_Subject_ID", all.x = T)

res_5.3.1 <- linreg_res_fnc(outc = "log_ige", 
                            mets_list = list_analysis, 
                            dat = "pheno_tmp", 
                            add_covar = " + Age_at_Collection + bmi + Gender + race + Ethnicity + ics_trim_totnum_5y")

## Save results
res_ige_sph_microb_ratio <- res_5.3.1
res_ige_sph_microb_ratio <- res_ige_sph_microb_ratio %>% separate(feat_id, sep = "\\.", c("numerator", "denominator"), remove = F)

dict <- sph_mets_info[,c("feat_id", "tentative_annotation", "SUB PATHWAY", "SUPER PATHWAY")]
res_ige_sph_microb_ratio$numerator <- dict$tentative_annotation[match(res_ige_sph_microb_ratio$numerator, dict$feat_id)]
res_ige_sph_microb_ratio$numerator_sub_pathway <- dict$`SUB PATHWAY`[match(res_ige_sph_microb_ratio$numerator, dict$tentative_annotation)]

dict <- microb_mets_info[,c("feat_id", "Component_Name", "SUB_PATHWAY", "SUPER_PATHWAY")]
res_ige_sph_microb_ratio$denominator <- dict$Component_Name[match(res_ige_sph_microb_ratio$denominator, dict$feat_id)]
res_ige_sph_microb_ratio$denominator_sub_pathway <- dict$`SUB_PATHWAY`[match(res_ige_sph_microb_ratio$denominator, dict$`Component_Name`)]

setcolorder(res_ige_sph_microb_ratio, c("feat_id", "numerator", "numerator_sub_pathway", "denominator", "denominator_sub_pathway"))

saveRDS(res_ige_sph_microb_ratio, file = here(res_dir, "res_ige_sph_microb_ratio.rds"))

## display results
### min p-value = 2.889224e-05
### #sig = 0/3898
datatable(res_ige_sph_microb_ratio[res_ige_sph_microb_ratio$fdr <= 0.05,], 
          filter = "top") %>% 
          formatSignif(columns = c("beta", "pval", "fdr", "fdr_bh", "se", "tval", "lower95", "upper95"), digits = 3)

res_5.3.1_sig <- res_ige_sph_microb_ratio[res_ige_sph_microb_ratio$fdr <= 0.05,]

```

### 5.3.2 targeted sph/str
```{r}
list_analysis <- mets_missing_sph_str[mets_missing_sph_str$pct_na_met <= 0.3,"feat_id"]
length(list_analysis) #1248

list_tmp <- append("Biobank_Subject_ID", list_analysis)
pheno_tmp <- merge(dat_tmp, df_sph_steroid_ratio_log[,..list_tmp], by = "Biobank_Subject_ID", all.x = T)

res_5.3.2 <- linreg_res_fnc(outc = "log_ige", 
                            mets_list = list_analysis, 
                            dat = "pheno_tmp", 
                            add_covar = " + Age_at_Collection + bmi + Gender + race + Ethnicity + ics_trim_totnum_5y")

## Save results
res_ige_sph_str_ratio <- res_5.3.2
res_ige_sph_str_ratio <- res_ige_sph_str_ratio %>% separate(feat_id, sep = "\\.", c("numerator", "denominator"), remove = F)

dict <- sph_mets_info[,c("feat_id", "tentative_annotation", "SUB PATHWAY", "SUPER PATHWAY")]
res_ige_sph_str_ratio$numerator <- dict$tentative_annotation[match(res_ige_sph_str_ratio$numerator, dict$feat_id)]
res_ige_sph_str_ratio$numerator_sub_pathway <- dict$`SUB PATHWAY`[match(res_ige_sph_str_ratio$numerator, dict$tentative_annotation)]

dict <- sph_mets_info[,c("feat_id", "tentative_annotation", "SUB PATHWAY", "SUPER PATHWAY")]
res_asthma_sph_str_ratio$numerator <- dict$tentative_annotation[match(res_asthma_sph_str_ratio$numerator, dict$feat_id)]
res_asthma_sph_str_ratio$numerator_sub_pathway <- dict$`SUB PATHWAY`[match(res_asthma_sph_str_ratio$numerator, dict$tentative_annotation)]


dict <- str_mets_info[,c("feat_id", "Component Name", "SUB PATHWAY", "SUPER PATHWAY")]
res_ige_sph_str_ratio$denominator <- dict$`Component Name`[match(res_ige_sph_str_ratio$denominator, dict$feat_id)]
res_ige_sph_str_ratio$denominator_sub_pathway <- dict$`SUB PATHWAY`[match(res_ige_sph_str_ratio$denominator, dict$`Component Name`)]

setcolorder(res_ige_sph_str_ratio, c("feat_id", "numerator", "numerator_sub_pathway", "denominator", "denominator_sub_pathway"))


saveRDS(res_ige_sph_str_ratio, file = here(res_dir, "res_ige_sph_str_ratio.rds"))

## display results
### min p-value = 0.0000399
### #sig = 202/1248
datatable(res_ige_sph_str_ratio[res_ige_sph_str_ratio$fdr <= 0.05, ], 
          filter = "top") %>% 
          formatSignif(columns = c("beta", "pval", "fdr", "fdr_bh", "se", "tval"), digits = 3)

res_5.3.2_sig <- res_ige_sph_str_ratio[res_ige_sph_str_ratio$fdr <= 0.05, ]

res_5.3.2_sig_top20 <- res_5.3.2_sig[c(1:20),] %>% as.data.table()
res_5.3.2_sig_top20$denominator[res_5.3.2_sig_top20$denominator == "dehydroepiandrosterone sulfate"] <- "DHEAS"
res_5.3.2_sig_top20$annotation <- paste(res_5.3.2_sig_top20$numerator, res_5.3.2_sig_top20$denominator, sep = " / ")
res_5.3.2_sig_top20$class <- "sph/str"
res_5.3.2_sig_top20$Pval <- signif(res_5.3.2_sig_top20$pval, 3)
res_5.3.2_sig_top20 <- res_5.3.2_sig_top20[order(denominator_sub_pathway, beta), ]


p_sph_str_log_ige <- ggplot(res_5.3.2_sig_top20, 
                aes(reorder(annotation, -beta), y = beta, color = beta >0)) +
   scale_colour_manual(name = '', values = setNames(c('red','blue'),c(T, F)))+
   geom_rect(aes(fill = denominator_sub_pathway),xmin = -Inf,xmax = Inf,
             ymin = -Inf,ymax = Inf, alpha = 0, color = 'black', linejoin = "round") +
   facet_grid(denominator_sub_pathway ~ class, scales = "free_y", space = "free_y", labeller = label_wrap_gen()) +
   geom_errorbar(aes(ymin = lower95, ymax = upper95), width = 0.25) +
   geom_pointrange(aes(ymin = lower95, ymax = upper95), size = 0.25) +
   geom_hline(yintercept = 0, linetype = 2) +
   labs(
     x = "",
     y = ""
   ) +
   theme_minimal() +
   theme_bw() +
   theme(
     title = element_text(size = 16),
     strip.text.x = element_text(size = 12, face = "bold"),
     strip.text.y = element_text(size = 10, face = "bold"),
     axis.title = element_text(size = 12),
     axis.text = element_text(size = 10, face = "bold"),
     legend.position = "",
     legend.title = element_text(size = 12),
     legend.text = element_text(size = 12),
     strip.text.y.right = element_text(angle = 0),
     strip.background = element_rect(fill = "#FDB863", colour = "black")) +
   coord_flip() + scale_fill_viridis_d(option = "D", direction = -1, begin = 0, end = 1, alpha = 0.1)
   # scale_y_continuous(limits = c(0.2, 1.6))

```

### 5.3.3 targeted microb/str
```{r}
list_analysis <- mets_missing_microb_str[mets_missing_microb_str$pct_na_met <= 0.3,"feat_id"]
length(list_analysis) #794

list_tmp <- append("Biobank_Subject_ID", list_analysis)
pheno_tmp <- merge(dat_tmp, df_microb_steroid_ratio_log[,..list_tmp], by = "Biobank_Subject_ID", all.x = T)

res_5.3.3 <- linreg_res_fnc(outc = "log_ige", 
                            mets_list = list_analysis, 
                            dat = "pheno_tmp", 
                            add_covar = " + Age_at_Collection + bmi + Gender + race + Ethnicity + ics_trim_totnum_5y")

## Save results
res_ige_microb_str_ratio <- res_5.3.3
res_ige_microb_str_ratio <- res_ige_microb_str_ratio %>% separate(feat_id, sep = "\\.", c("numerator", "denominator"), remove = F)

dict <- microb_mets_info[,c("feat_id", "Component_Name", "SUB_PATHWAY", "SUPER_PATHWAY")]
res_ige_microb_str_ratio$numerator <- dict$Component_Name[match(res_ige_microb_str_ratio$numerator, dict$feat_id)]
res_ige_microb_str_ratio$numerator_sub_pathway <- dict$SUB_PATHWAY[match(res_ige_microb_str_ratio$numerator, dict$Component_Name)]

dict <- str_mets_info[,c("feat_id", "Component Name", "SUB PATHWAY", "SUPER PATHWAY")]
res_ige_microb_str_ratio$denominator <- dict$`Component Name`[match(res_ige_microb_str_ratio$denominator, dict$feat_id)]
res_ige_microb_str_ratio$denominator_sub_pathway <- dict$`SUB PATHWAY`[match(res_ige_microb_str_ratio$denominator, dict$`Component Name`)]

setcolorder(res_ige_microb_str_ratio, c("feat_id", "numerator", "numerator_sub_pathway", "denominator", "denominator_sub_pathway"))


saveRDS(res_ige_microb_str_ratio, file = here(res_dir, "res_ige_microb_str_ratio.rds"))

## display results
### min p-value = 0.0000399
### #sig = 49/794
datatable(res_ige_microb_str_ratio[res_ige_microb_str_ratio$fdr <= 0.05,], 
          filter = "top") %>% 
          formatSignif(columns = c("beta", "pval", "fdr", "fdr_bh", "se", "tval"), digits = 3)

res_5.3.3_sig <- res_ige_microb_str_ratio[res_ige_microb_str_ratio$fdr <= 0.05,]

res_5.3.3_sig_top20 <- res_5.3.3_sig[c(1:20),] %>% as.data.table
res_5.3.3_sig_top20$denominator[res_5.3.3_sig_top20$denominator == "dehydroepiandrosterone sulfate"] <- "DHEAS"
res_5.3.3_sig_top20$annotation <- paste(res_5.3.3_sig_top20$numerator, res_5.3.3_sig_top20$denominator, sep = " / ")
res_5.3.3_sig_top20$class <- "microb/str"
res_5.3.3_sig_top20$Pval <- signif(res_5.3.3_sig_top20$pval, 3)
res_5.3.3_sig_top20 <- res_5.3.3_sig_top20[order(denominator_sub_pathway, beta), ]


p_microb_str_log_ige <- ggplot(res_5.3.3_sig_top20, 
                aes(reorder(annotation, -beta), y = beta, color = beta >0)) +
   scale_colour_manual(name = '', values = setNames(c('red','blue'),c(T, F)))+
   geom_rect(aes(fill = denominator_sub_pathway),xmin = -Inf,xmax = Inf,
             ymin = -Inf,ymax = Inf, alpha = 0, color = 'black', linejoin = "round") +
   facet_grid(denominator_sub_pathway ~ class, scales = "free_y", space = "free_y", labeller = label_wrap_gen()) +
   geom_errorbar(aes(ymin = lower95, ymax = upper95), width = 0.25) +
   geom_pointrange(aes(ymin = lower95, ymax = upper95), size = 0.25) +
   geom_hline(yintercept = 0, linetype = 2) +
   labs(
     x = "",
     y = ""
   ) +
   theme_minimal() +
   theme_bw() +
   theme(
     title = element_text(size = 16),
     strip.text.x = element_text(size = 12, face = "bold"),
     strip.text.y = element_text(size = 10, face = "bold"),
     axis.title = element_text(size = 12),
     axis.text = element_text(size = 10, face = "bold"),
     legend.position = "",
     legend.title = element_text(size = 12),
     legend.text = element_text(size = 12),
     strip.text.y.right = element_text(angle = 0),
     strip.background = element_rect(fill = "#B8E186", colour = "black")) +
   coord_flip() + scale_fill_viridis_d(option = "D", direction = -1, begin = 0, end = 1, alpha = 0.1)
   # scale_y_continuous(limits = c(0.2, 1.6))

```


## 5.4 FEV1
```{r}
sum(is.na(dat_tmp$FEV1_most_recent)) #266
ggplot(dat_tmp, aes(x=FEV1_most_recent)) + geom_histogram(color="darkblue", fill="lightblue")

```

### 5.4.1 targeted sph/microb
```{r}
list_analysis <- mets_missing_sph_microb[mets_missing_sph_microb$pct_na_met <= 0.3,"feat_id"]
length(list_analysis) #3898

list_tmp <- append("Biobank_Subject_ID", list_analysis)
pheno_tmp <- merge(dat_tmp, df_sph_microb_ratio_log[,..list_tmp], by = "Biobank_Subject_ID", all.x = T)

res_5.4.1 <- linreg_res_fnc(outc = "FEV1_most_recent", 
                            mets_list = list_analysis, 
                            dat = "pheno_tmp", 
                            add_covar = " + Age_at_Collection + bmi + Gender + race + Ethnicity + Height + ics_trim_totnum_5y")

## Save results
res_FEV1_sph_microb_ratio <- res_5.4.1
res_FEV1_sph_microb_ratio <- res_FEV1_sph_microb_ratio %>% separate(feat_id, sep = "\\.", c("numerator", "denominator"), remove = F)

dict <- sph_mets_info[,c("feat_id", "tentative_annotation", "SUB PATHWAY", "SUPER PATHWAY")]
res_FEV1_sph_microb_ratio$numerator <- dict$tentative_annotation[match(res_FEV1_sph_microb_ratio$numerator, dict$feat_id)]
res_FEV1_sph_microb_ratio$numerator_sub_pathway <- dict$`SUB PATHWAY`[match(res_FEV1_sph_microb_ratio$numerator, dict$tentative_annotation)]

dict <- microb_mets_info[,c("feat_id", "Component_Name", "SUB_PATHWAY", "SUPER_PATHWAY")]
res_FEV1_sph_microb_ratio$denominator <- dict$Component_Name[match(res_FEV1_sph_microb_ratio$denominator, dict$feat_id)]
res_FEV1_sph_microb_ratio$denominator_sub_pathway <- dict$`SUB_PATHWAY`[match(res_FEV1_sph_microb_ratio$denominator, dict$`Component_Name`)]

setcolorder(res_FEV1_sph_microb_ratio, c("feat_id", "numerator", "numerator_sub_pathway", "denominator", "denominator_sub_pathway"))


saveRDS(res_FEV1_sph_microb_ratio, file = here(res_dir, "res_FEV1_sph_microb_ratio.rds"))

## display results
### min p-value = 4.75e-7
### #sig = 321/3989
datatable(res_FEV1_sph_microb_ratio[res_FEV1_sph_microb_ratio$fdr <= 0.05,], 
          filter = "top") %>% 
          formatSignif(columns = c("beta", "pval", "fdr", "fdr_bh", "se", "tval"), digits = 3)

res_5.4.1_sig <- res_FEV1_sph_microb_ratio[res_FEV1_sph_microb_ratio$fdr <= 0.05,]

res_5.4.1_sig_top20 <- res_5.4.1_sig[c(1:20),] %>% as.data.table()
res_5.4.1_sig_top20$annotation <- paste(res_5.4.1_sig_top20$numerator, res_5.4.1_sig_top20$denominator, sep = " / ")
res_5.4.1_sig_top20$class <- "sph/microb"
res_5.4.1_sig_top20$Pval <- signif(res_5.4.1_sig_top20$pval, 3)
res_5.4.1_sig_top20 <- res_5.4.1_sig_top20[order(denominator_sub_pathway, beta), ]

p_sph_microb_FEV1 <- ggplot(res_5.4.1_sig_top20, 
                aes(reorder(annotation, -beta), y = beta, color = beta >0)) +
   scale_colour_manual(name = '', values = setNames(c('red','blue'),c(T, F)))+
   geom_rect(aes(fill = denominator_sub_pathway),xmin = -Inf,xmax = Inf,
             ymin = -Inf,ymax = Inf, alpha = 0, color = 'black', linejoin = "round") +
   facet_grid(denominator_sub_pathway ~ class, scales = "free_y", space = "free_y", labeller = label_wrap_gen()) +
   geom_errorbar(aes(ymin = lower95, ymax = upper95), width = 0.25) +
   geom_pointrange(aes(ymin = lower95, ymax = upper95), size = 0.25) +
   geom_hline(yintercept = 0, linetype = 2) +
   labs(
     x = "",
     y = ""
   ) +
   theme_minimal() +
   theme_bw() +
   theme(
     title = element_text(size = 16),
     strip.text.x = element_text(size = 12, face = "bold"),
     strip.text.y = element_text(size = 10, face = "bold"),
     axis.title = element_text(size = 12),
     axis.text = element_text(size = 10, face = "bold"),
     legend.position = "",
     legend.title = element_text(size = 12),
     legend.text = element_text(size = 12),
     strip.text.y.right = element_text(angle = 0),
     strip.background = element_rect(fill = "#B2ABD2", colour = "black")) +
   coord_flip() + scale_fill_viridis_d(option = "D", direction = -1, begin = 0, end = 1, alpha = 0.1)
   # scale_y_continuous(limits = c(0.2, 1.6))

```


### 5.4.2 targeted sph/str
```{r}
list_analysis <- mets_missing_sph_str[mets_missing_sph_str$pct_na_met <= 0.3,"feat_id"]
length(list_analysis) #1248

list_tmp <- append("Biobank_Subject_ID", list_analysis)
pheno_tmp <- merge(dat_tmp, df_sph_steroid_ratio_log[,..list_tmp], by = "Biobank_Subject_ID", all.x = T)

res_5.4.2 <- linreg_res_fnc(outc = "FEV1_most_recent", 
                            mets_list = list_analysis, 
                            dat = "pheno_tmp", 
                            add_covar = " + Age_at_Collection + bmi + Gender + race + Ethnicity + Height + ics_trim_totnum_5y")

## Save results
res_FEV1_sph_str_ratio <- res_5.4.2
res_FEV1_sph_str_ratio <- res_FEV1_sph_str_ratio %>% separate(feat_id, sep = "\\.", c("numerator", "denominator"), remove = F)

dict <- sph_mets_info[,c("feat_id", "tentative_annotation", "SUB PATHWAY", "SUPER PATHWAY")]
res_FEV1_sph_str_ratio$numerator <- dict$tentative_annotation[match(res_FEV1_sph_str_ratio$numerator, dict$feat_id)]
res_FEV1_sph_str_ratio$numerator_sub_pathway <- dict$`SUB PATHWAY`[match(res_FEV1_sph_str_ratio$numerator, dict$tentative_annotation)]

dict <- str_mets_info[,c("feat_id", "Component Name", "SUB PATHWAY", "SUPER PATHWAY")]
res_FEV1_sph_str_ratio$denominator <- dict$`Component Name`[match(res_FEV1_sph_str_ratio$denominator, dict$feat_id)]
res_FEV1_sph_str_ratio$denominator_sub_pathway <- dict$`SUB PATHWAY`[match(res_FEV1_sph_str_ratio$denominator, dict$`Component Name`)]

setcolorder(res_FEV1_sph_str_ratio, c("feat_id", "numerator", "numerator_sub_pathway", "denominator", "denominator_sub_pathway"))

saveRDS(res_FEV1_sph_str_ratio, file = here(res_dir, "res_FEV1_sph_str_ratio.rds"))

## display results
### min p-value = NA
### #sig = 0/1248
datatable(res_FEV1_sph_str_ratio[res_FEV1_sph_str_ratio$fdr <= 0.05, ], 
          filter = "top") %>% 
          formatSignif(columns = c("beta", "pval", "fdr", "fdr_bh", "se", "tval"), digits = 3)

res_5.4.2_sig <- res_FEV1_sph_str_ratio[res_FEV1_sph_str_ratio$fdr <= 0.05, ]

```


### 5.4.3 targeted microb/str
```{r}
list_analysis <- mets_missing_microb_str[mets_missing_microb_str$pct_na_met <= 0.3,"feat_id"]
length(list_analysis) #794

list_tmp <- append("Biobank_Subject_ID", list_analysis)
pheno_tmp <- merge(dat_tmp, df_microb_steroid_ratio_log[,..list_tmp], by = "Biobank_Subject_ID", all.x = T)

res_5.4.3 <- linreg_res_fnc(outc = "FEV1_most_recent", 
                            mets_list = list_analysis, 
                            dat = "pheno_tmp", 
                            add_covar = " + Age_at_Collection + bmi + Gender + race + Ethnicity + Height + ics_trim_totnum_5y")

## Save results
res_FEV1_microb_str_ratio <- res_5.4.3
res_FEV1_microb_str_ratio <- res_FEV1_microb_str_ratio %>% separate(feat_id, sep = "\\.", c("numerator", "denominator"), remove = F)

dict <- microb_mets_info[,c("feat_id", "Component_Name", "SUB_PATHWAY", "SUPER_PATHWAY")]
res_FEV1_microb_str_ratio$numerator <- dict$Component_Name[match(res_FEV1_microb_str_ratio$numerator, dict$feat_id)]
res_FEV1_microb_str_ratio$numerator_sub_pathway <- dict$SUB_PATHWAY[match(res_FEV1_microb_str_ratio$numerator, dict$Component_Name)]

dict <- str_mets_info[,c("feat_id", "Component Name", "SUB PATHWAY", "SUPER PATHWAY")]
res_FEV1_microb_str_ratio$denominator <- dict$`Component Name`[match(res_FEV1_microb_str_ratio$denominator, dict$feat_id)]
res_FEV1_microb_str_ratio$denominator_sub_pathway <- dict$`SUB PATHWAY`[match(res_FEV1_microb_str_ratio$denominator, dict$`Component Name`)]

setcolorder(res_FEV1_microb_str_ratio, c("feat_id", "numerator", "numerator_sub_pathway", "denominator", "denominator_sub_pathway"))


saveRDS(res_FEV1_microb_str_ratio, file = here(res_dir, "res_FEV1_microb_str_ratio.rds"))

## display results
### min p-value = 0.0000642
### #sig = 4/794
datatable((res_FEV1_microb_str_ratio[res_FEV1_microb_str_ratio$fdr <= 0.05, ]), 
          filter = "top") %>% 
          formatSignif(columns = c("beta", "pval", "fdr", "fdr_bh", "se", "tval"), digits = 3)

res_5.4.3_sig <- res_FEV1_microb_str_ratio[res_FEV1_microb_str_ratio$fdr <= 0.05, ]

res_5.4.3_sig_top20 <- res_5.4.3_sig %>% as.data.table
res_5.4.3_sig_top20$denominator[res_5.4.3_sig_top20$denominator == "dehydroepiandrosterone sulfate"] <- "DHEAS"
res_5.4.3_sig_top20$annotation <- paste(res_5.4.3_sig_top20$numerator, res_5.4.3_sig_top20$denominator, sep = " / ")
res_5.4.3_sig_top20$class <- "microb/str"
res_5.4.3_sig_top20$Pval <- signif(res_5.4.3_sig_top20$pval, 3)
res_5.4.3_sig_top20 <- res_5.4.3_sig_top20[order(denominator_sub_pathway, beta), ]

p_microb_str_FEV1 <- ggplot(res_5.4.3_sig_top20, 
                aes(reorder(annotation, -beta), y = beta, color = beta >0)) +
   scale_colour_manual(name = '', values = setNames(c('red','blue'),c(T, F)))+
   geom_rect(aes(fill = denominator_sub_pathway),xmin = -Inf,xmax = Inf,
             ymin = -Inf,ymax = Inf, alpha = 0, color = 'black', linejoin = "round") +
   facet_grid(denominator_sub_pathway ~ class, scales = "free_y", space = "free_y", labeller = label_wrap_gen()) +
   geom_errorbar(aes(ymin = lower95, ymax = upper95), width = 0.25) +
   geom_pointrange(aes(ymin = lower95, ymax = upper95), size = 0.25) +
   geom_hline(yintercept = 0, linetype = 2) +
   labs(
     x = "",
     y = ""
   ) +
   theme_minimal() +
   theme_bw() +
   theme(
     title = element_text(size = 16),
     strip.text.x = element_text(size = 12, face = "bold"),
     strip.text.y = element_text(size = 10, face = "bold"),
     axis.title = element_text(size = 12),
     axis.text = element_text(size = 10, face = "bold"),
     legend.position = "",
     legend.title = element_text(size = 12),
     legend.text = element_text(size = 12),
     strip.text.y.right = element_text(angle = 0),
     strip.background = element_rect(fill = "#B8E186", colour = "black")) +
   coord_flip() + scale_fill_viridis_d(option = "D", direction = -1, begin = 0, end = 1, alpha = 0.1)
   # scale_y_continuous(limits = c(0.2, 1.6))

```

## 5.5 FVC
```{r}
sum(is.na(dat_tmp$FVC_most_recent)) #268
ggplot(dat_tmp, aes(x=FVC_most_recent)) + geom_histogram(color="darkblue", fill="lightblue")

```

### 5.5.1 targeted sph/microb
```{r}
list_analysis <- mets_missing_sph_microb[mets_missing_sph_microb$pct_na_met <= 0.3,"feat_id"]
length(list_analysis) #3898

list_tmp <- append("Biobank_Subject_ID", list_analysis)
pheno_tmp <- merge(dat_tmp, df_sph_microb_ratio_log[,..list_tmp], by = "Biobank_Subject_ID", all.x = T)

res_5.5.1 <- linreg_res_fnc(outc = "FVC_most_recent", 
                            mets_list = list_analysis, 
                            dat = "pheno_tmp", 
                            add_covar = " + Age_at_Collection + bmi + Gender + race + Ethnicity + Height + ics_trim_totnum_5y")

## Save results
res_FVC_sph_microb_ratio <- res_5.5.1
res_FVC_sph_microb_ratio <- res_FVC_sph_microb_ratio %>% separate(feat_id, sep = "\\.", c("numerator", "denominator"), remove = F)

dict <- sph_mets_info[,c("feat_id", "tentative_annotation", "SUB PATHWAY", "SUPER PATHWAY")]
res_FVC_sph_microb_ratio$numerator <- dict$tentative_annotation[match(res_FVC_sph_microb_ratio$numerator, dict$feat_id)]
res_FVC_sph_microb_ratio$numerator_sub_pathway <- dict$`SUB PATHWAY`[match(res_FVC_sph_microb_ratio$numerator, dict$tentative_annotation)]

dict <- microb_mets_info[,c("feat_id", "Component_Name", "SUB_PATHWAY", "SUPER_PATHWAY")]
res_FVC_sph_microb_ratio$denominator <- dict$Component_Name[match(res_FVC_sph_microb_ratio$denominator, dict$feat_id)]
res_FVC_sph_microb_ratio$denominator_sub_pathway <- dict$`SUB_PATHWAY`[match(res_FVC_sph_microb_ratio$denominator, dict$`Component_Name`)]

setcolorder(res_FVC_sph_microb_ratio, c("feat_id", "numerator", "numerator_sub_pathway", "denominator", "denominator_sub_pathway"))


saveRDS(res_FVC_sph_microb_ratio, file = here(res_dir, "res_FVC_sph_microb_ratio.rds"))

## display results
### min p-value = 9.80e-8	
### #sig = 479/3898
datatable(res_FVC_sph_microb_ratio[res_FVC_sph_microb_ratio$fdr <= 0.05,], 
          filter = "top") %>% 
          formatSignif(columns = c("beta", "pval", "fdr", "fdr_bh", "se", "tval"), digits = 3)

res_5.5.1_sig <- res_FVC_sph_microb_ratio[res_FVC_sph_microb_ratio$fdr <= 0.05,]

res_5.5.1_sig_top20 <- res_5.5.1_sig[c(1:20),] %>% as.data.table()
res_5.5.1_sig_top20$annotation <- paste(res_5.5.1_sig_top20$numerator, res_5.5.1_sig_top20$denominator, sep = " / ")
res_5.5.1_sig_top20$class <- "sph/microb"
res_5.5.1_sig_top20$Pval <- signif(res_5.5.1_sig_top20$pval, 3)
res_5.5.1_sig_top20 <- res_5.5.1_sig_top20[order(denominator_sub_pathway, beta), ]

p_sph_microb_FVC <- ggplot(res_5.5.1_sig_top20, 
                aes(reorder(annotation, -beta), y = beta, color = beta >0)) +
   scale_colour_manual(name = '', values = setNames(c('red','blue'),c(T, F)))+
   geom_rect(aes(fill = denominator_sub_pathway),xmin = -Inf,xmax = Inf,
             ymin = -Inf,ymax = Inf, alpha = 0, color = 'black', linejoin = "round") +
   facet_grid(denominator_sub_pathway ~ class, scales = "free_y", space = "free_y", labeller = label_wrap_gen()) +
   geom_errorbar(aes(ymin = lower95, ymax = upper95), width = 0.25) +
   geom_pointrange(aes(ymin = lower95, ymax = upper95), size = 0.25) +
   geom_hline(yintercept = 0, linetype = 2) +
   labs(
     x = "",
     y = ""
   ) +
   theme_minimal() +
   theme_bw() +
   theme(
     title = element_text(size = 16),
     strip.text.x = element_text(size = 12, face = "bold"),
     strip.text.y = element_text(size = 10, face = "bold"),
     axis.title = element_text(size = 12),
     axis.text = element_text(size = 10, face = "bold"),
     legend.position = "",
     legend.title = element_text(size = 12),
     legend.text = element_text(size = 12),
     strip.text.y.right = element_text(angle = 0),
     strip.background = element_rect(fill = "#B2ABD2", colour = "black")) +
   coord_flip() + scale_fill_viridis_d(option = "D", direction = -1, begin = 0, end = 1, alpha = 0.1)
   # scale_y_continuous(limits = c(0.2, 1.6))

```

### 5.5.2 targeted sph/str
```{r}
list_analysis <- mets_missing_sph_str[mets_missing_sph_str$pct_na_met <= 0.3,"feat_id"]
length(list_analysis) #1248

list_tmp <- append("Biobank_Subject_ID", list_analysis)
pheno_tmp <- merge(dat_tmp, df_sph_steroid_ratio_log[,..list_tmp], by = "Biobank_Subject_ID", all.x = T)

res_5.5.2 <- linreg_res_fnc(outc = "FVC_most_recent", 
                            mets_list = list_analysis, 
                            dat = "pheno_tmp", 
                            add_covar = " + Age_at_Collection + bmi + Gender + race + Ethnicity + Height + ics_trim_totnum_5y")

## Save results
res_FVC_sph_str_ratio <- res_5.5.2
res_FVC_sph_str_ratio <- res_FVC_sph_str_ratio %>% separate(feat_id, sep = "\\.", c("numerator", "denominator"), remove = F)

dict <- sph_mets_info[,c("feat_id", "tentative_annotation", "SUB PATHWAY", "SUPER PATHWAY")]
res_FVC_sph_str_ratio$numerator <- dict$tentative_annotation[match(res_FVC_sph_str_ratio$numerator, dict$feat_id)]
res_FVC_sph_str_ratio$numerator_sub_pathway <- dict$`SUB PATHWAY`[match(res_FVC_sph_str_ratio$numerator, dict$tentative_annotation)]

dict <- str_mets_info[,c("feat_id", "Component Name", "SUB PATHWAY", "SUPER PATHWAY")]
res_FVC_sph_str_ratio$denominator <- dict$`Component Name`[match(res_FVC_sph_str_ratio$denominator, dict$feat_id)]
res_FVC_sph_str_ratio$denominator_sub_pathway <- dict$`SUB PATHWAY`[match(res_FVC_sph_str_ratio$denominator, dict$`Component Name`)]

setcolorder(res_FVC_sph_str_ratio, c("feat_id", "numerator", "numerator_sub_pathway", "denominator", "denominator_sub_pathway"))


saveRDS(res_FVC_sph_str_ratio, file = here(res_dir, "res_FVC_sph_str_ratio.rds"))

## display results
### min p-value = 9.80e-8	
### #sig = 0/1248
datatable(res_FVC_sph_str_ratio[res_FVC_sph_str_ratio$fdr <= 0.05,], 
          filter = "top") %>% 
          formatSignif(columns = c("beta", "pval", "fdr", "fdr_bh", "se", "tval"), digits = 3)

res_5.5.2_sig <- res_FVC_sph_str_ratio[res_FVC_sph_str_ratio$fdr <= 0.05,]

```

### 5.5.3 targeted microb/str
```{r}
list_analysis <- mets_missing_microb_str[mets_missing_microb_str$pct_na_met <= 0.3,"feat_id"]
length(list_analysis) #794

list_tmp <- append("Biobank_Subject_ID", list_analysis)
pheno_tmp <- merge(dat_tmp, df_microb_steroid_ratio_log[,..list_tmp], by = "Biobank_Subject_ID", all.x = T)

res_5.5.3 <- linreg_res_fnc(outc = "FVC_most_recent", 
                            mets_list = list_analysis, 
                            dat = "pheno_tmp", 
                            add_covar = " + Age_at_Collection + bmi + Gender + race + Ethnicity + Height + ics_trim_totnum_5y")

## Save results
res_FVC_microb_str_ratio <- res_5.5.3
res_FVC_microb_str_ratio <- res_FVC_microb_str_ratio %>% separate(feat_id, sep = "\\.", c("numerator", "denominator"), remove = F)

dict <- microb_mets_info[,c("feat_id", "Component_Name", "SUB_PATHWAY", "SUPER_PATHWAY")]
res_FVC_microb_str_ratio$numerator <- dict$Component_Name[match(res_FVC_microb_str_ratio$numerator, dict$feat_id)]
res_FVC_microb_str_ratio$numerator_sub_pathway <- dict$SUB_PATHWAY[match(res_FVC_microb_str_ratio$numerator, dict$Component_Name)]

dict <- str_mets_info[,c("feat_id", "Component Name", "SUB PATHWAY", "SUPER PATHWAY")]
res_FVC_microb_str_ratio$denominator <- dict$`Component Name`[match(res_FVC_microb_str_ratio$denominator, dict$feat_id)]
res_FVC_microb_str_ratio$denominator_sub_pathway <- dict$`SUB PATHWAY`[match(res_FVC_microb_str_ratio$denominator, dict$`Component Name`)]

setcolorder(res_FVC_microb_str_ratio, c("feat_id", "numerator", "numerator_sub_pathway", "denominator", "denominator_sub_pathway"))


saveRDS(res_FVC_microb_str_ratio, file = here(res_dir, "res_FVC_microb_str_ratio.rds"))

## display results
datatable(res_FVC_microb_str_ratio[res_FVC_microb_str_ratio$fdr <= 0.05,], 
          filter = "top") %>% 
          formatSignif(columns = c("beta", "pval", "fdr", "fdr_bh", "se", "tval"), digits = 3)

res_5.5.3_sig <- res_FVC_microb_str_ratio[res_FVC_microb_str_ratio$fdr <= 0.05,]

res_5.5.3_sig_top20 <- res_5.5.3_sig[c(1:20),] %>% as.data.table()
res_5.5.3_sig_top20$denominator[res_5.5.3_sig_top20$denominator == "dehydroepiandrosterone sulfate"] <- "DHEAS"
res_5.5.3_sig_top20$annotation <- paste(res_5.5.3_sig_top20$numerator, res_5.5.3_sig_top20$denominator, sep = " / ")
res_5.5.3_sig_top20$class <- "microb/str"
res_5.5.3_sig_top20$Pval <- signif(res_5.5.3_sig_top20$pval, 3)
res_5.5.3_sig_top20 <- res_5.5.3_sig_top20[order(denominator_sub_pathway, beta), ]

p_microb_str_FVC <- ggplot(res_5.5.3_sig_top20, 
                aes(reorder(annotation, -beta), y = beta, color = beta >0)) +
   scale_colour_manual(name = '', values = setNames(c('red','blue'),c(T, F)))+
   geom_rect(aes(fill = denominator_sub_pathway),xmin = -Inf,xmax = Inf,
             ymin = -Inf,ymax = Inf, alpha = 0, color = 'black', linejoin = "round") +
   facet_grid(denominator_sub_pathway ~ class, scales = "free_y", space = "free_y", labeller = label_wrap_gen()) +
   geom_errorbar(aes(ymin = lower95, ymax = upper95), width = 0.25) +
   geom_pointrange(aes(ymin = lower95, ymax = upper95), size = 0.25) +
   geom_hline(yintercept = 0, linetype = 2) +
   labs(
     x = "",
     y = ""
   ) +
   theme_minimal() +
   theme_bw() +
   theme(
     title = element_text(size = 16),
     strip.text.x = element_text(size = 12, face = "bold"),
     strip.text.y = element_text(size = 10, face = "bold"),
     axis.title = element_text(size = 12),
     axis.text = element_text(size = 10, face = "bold"),
     legend.position = "",
     legend.title = element_text(size = 12),
     legend.text = element_text(size = 12),
     strip.text.y.right = element_text(angle = 0),
     strip.background = element_rect(fill = "#B8E186", colour = "black")) +
   coord_flip() + scale_fill_viridis_d(option = "D", direction = -1, begin = 0, end = 1, alpha = 0.1)
   # scale_y_continuous(limits = c(0.2, 1.6))

```


## 5.6 FEV1/FVC
```{r}
sum(is.na(dat_tmp$FEV1_FVC_most_recent)) #268
ggplot(dat_tmp, aes(x=FEV1_FVC_most_recent)) + geom_histogram(color="darkblue", fill="lightblue")

```

### 5.6.1 targeted sph/microb
```{r}
list_analysis <- mets_missing_sph_microb[mets_missing_sph_microb$pct_na_met <= 0.3,"feat_id"]
length(list_analysis) #3898

list_tmp <- append("Biobank_Subject_ID", list_analysis)
pheno_tmp <- merge(dat_tmp, df_sph_microb_ratio_log[,..list_tmp], by = "Biobank_Subject_ID", all.x = T)

res_5.6.1 <- linreg_res_fnc(outc = "FEV1_FVC_most_recent", 
                            mets_list = list_analysis, 
                            dat = "pheno_tmp", 
                            add_covar = " + Age_at_Collection + bmi + Gender + race + Ethnicity + Height + ics_trim_totnum_5y")

## Save results
res_FEV1_FVC_sph_microb_ratio <- res_5.6.1
res_FEV1_FVC_sph_microb_ratio <- res_FEV1_FVC_sph_microb_ratio %>% separate(feat_id, sep = "\\.", c("numerator", "denominator"), remove = F)

dict <- sph_mets_info[,c("feat_id", "tentative_annotation", "SUB PATHWAY", "SUPER PATHWAY")]
res_FEV1_FVC_sph_microb_ratio$numerator <- dict$tentative_annotation[match(res_FEV1_FVC_sph_microb_ratio$numerator, dict$feat_id)]
res_FEV1_FVC_sph_microb_ratio$numerator_sub_pathway <- dict$`SUB PATHWAY`[match(res_FEV1_FVC_sph_microb_ratio$numerator, dict$tentative_annotation)]

dict <- microb_mets_info[,c("feat_id", "Component_Name", "SUB_PATHWAY", "SUPER_PATHWAY")]
res_FEV1_FVC_sph_microb_ratio$denominator <- dict$Component_Name[match(res_FEV1_FVC_sph_microb_ratio$denominator, dict$feat_id)]
res_FEV1_FVC_sph_microb_ratio$denominator_sub_pathway <- dict$`SUB_PATHWAY`[match(res_FEV1_FVC_sph_microb_ratio$denominator, dict$`Component_Name`)]

setcolorder(res_FEV1_FVC_sph_microb_ratio, c("feat_id", "numerator", "numerator_sub_pathway", "denominator", "denominator_sub_pathway"))


saveRDS(res_FEV1_FVC_sph_microb_ratio, file = here(res_dir, "res_FEV1_FVC_sph_microb_ratio.rds"))

## display results
### min p-value = 9.80e-8	
### #sig = 0/3898
datatable(res_FEV1_FVC_sph_microb_ratio[res_FEV1_FVC_sph_microb_ratio$fdr <= 0.05,], 
          filter = "top") %>% 
          formatSignif(columns = c("beta", "pval", "fdr", "fdr_bh", "se", "tval"), digits = 3)

res_5.6.1_sig <- res_FEV1_FVC_sph_microb_ratio[res_FEV1_FVC_sph_microb_ratio$fdr <= 0.05,]

```

### 5.6.2 targeted sph/str
```{r}
list_analysis <- mets_missing_sph_str[mets_missing_sph_str$pct_na_met <= 0.3,"feat_id"]
length(list_analysis) #1248

list_tmp <- append("Biobank_Subject_ID", list_analysis)
pheno_tmp <- merge(dat_tmp, df_sph_steroid_ratio_log[,..list_tmp], by = "Biobank_Subject_ID", all.x = T)

res_5.5.2 <- linreg_res_fnc(outc = "FEV1_FVC_most_recent", 
                            mets_list = list_analysis, 
                            dat = "pheno_tmp", 
                            add_covar = " + Age_at_Collection + bmi + Gender + race + Ethnicity + Height + ics_trim_totnum_5y")

## Save results
res_FEV1_FVC_sph_str_ratio <- res_5.5.2
res_FEV1_FVC_sph_str_ratio <- res_FEV1_FVC_sph_str_ratio %>% separate(feat_id, sep = "\\.", c("numerator", "denominator"), remove = F)

dict <- sph_mets_info[,c("feat_id", "tentative_annotation", "SUB PATHWAY", "SUPER PATHWAY")]
res_FEV1_FVC_sph_str_ratio$numerator <- dict$tentative_annotation[match(res_FEV1_FVC_sph_str_ratio$numerator, dict$feat_id)]
res_FEV1_FVC_sph_str_ratio$numerator_sub_pathway <- dict$`SUB PATHWAY`[match(res_FEV1_FVC_sph_str_ratio$numerator, dict$tentative_annotation)]

dict <- str_mets_info[,c("feat_id", "Component Name", "SUB PATHWAY", "SUPER PATHWAY")]
res_FEV1_FVC_sph_str_ratio$denominator <- dict$`Component Name`[match(res_FEV1_FVC_sph_str_ratio$denominator, dict$feat_id)]
res_FEV1_FVC_sph_str_ratio$denominator_sub_pathway <- dict$`SUB PATHWAY`[match(res_FEV1_FVC_sph_str_ratio$denominator, dict$`Component Name`)]

setcolorder(res_FEV1_FVC_sph_str_ratio, c("feat_id", "numerator", "numerator_sub_pathway", "denominator", "denominator_sub_pathway"))


saveRDS(res_FEV1_FVC_sph_str_ratio, file = here(res_dir, "res_FEV1_FVC_sph_str_ratio.rds"))

## display results
### min p-value = 9.80e-8	
### #sig = 0/1248
datatable(res_FEV1_FVC_sph_str_ratio[res_FEV1_FVC_sph_str_ratio$fdr <= 0.05,], 
          filter = "top") %>% 
          formatSignif(columns = c("beta", "pval", "fdr", "fdr_bh", "se", "tval"), digits = 3)

res_5.6.2_sig <- res_FEV1_FVC_sph_str_ratio[res_FEV1_FVC_sph_str_ratio$fdr <= 0.05,]

```

### 5.6.3 targeted microb/str
```{r}
list_analysis <- mets_missing_microb_str[mets_missing_microb_str$pct_na_met <= 0.3,"feat_id"]
length(list_analysis) #794

list_tmp <- append("Biobank_Subject_ID", list_analysis)
pheno_tmp <- merge(dat_tmp, df_microb_steroid_ratio_log[,..list_tmp], by = "Biobank_Subject_ID", all.x = T)

res_5.6.3 <- linreg_res_fnc(outc = "FEV1_FVC_most_recent", 
                            mets_list = list_analysis, 
                            dat = "pheno_tmp", 
                            add_covar = " + Age_at_Collection + bmi + Gender + race + Ethnicity + Height + ics_trim_totnum_5y")

## Save results
res_FEV1_FVC_microb_str_ratio <- res_5.6.3
res_FEV1_FVC_microb_str_ratio <- res_FEV1_FVC_microb_str_ratio %>% separate(feat_id, sep = "\\.", c("numerator", "denominator"), remove = F)

dict <- microb_mets_info[,c("feat_id", "Component_Name", "SUB_PATHWAY", "SUPER_PATHWAY")]
res_FEV1_FVC_microb_str_ratio$numerator <- dict$Component_Name[match(res_FEV1_FVC_microb_str_ratio$numerator, dict$feat_id)]
res_FEV1_FVC_microb_str_ratio$numerator_sub_pathway <- dict$SUB_PATHWAY[match(res_FEV1_FVC_microb_str_ratio$numerator, dict$Component_Name)]

dict <- str_mets_info[,c("feat_id", "Component Name", "SUB PATHWAY", "SUPER PATHWAY")]
res_FEV1_FVC_microb_str_ratio$denominator <- dict$`Component Name`[match(res_FEV1_FVC_microb_str_ratio$denominator, dict$feat_id)]
res_FEV1_FVC_microb_str_ratio$denominator_sub_pathway <- dict$`SUB PATHWAY`[match(res_FEV1_FVC_microb_str_ratio$denominator, dict$`Component Name`)]

setcolorder(res_FEV1_FVC_microb_str_ratio, c("feat_id", "numerator", "numerator_sub_pathway", "denominator", "denominator_sub_pathway"))


saveRDS(res_FEV1_FVC_microb_str_ratio, file = here(res_dir, "res_FEV1_FVC_microb_str_ratio.rds"))

## display results
datatable(res_FEV1_FVC_microb_str_ratio[res_FEV1_FVC_microb_str_ratio$fdr <= 0.05,], 
          filter = "top") %>% 
          formatSignif(columns = c("beta", "pval", "fdr", "fdr_bh", "se", "tval"), digits = 3)

res_5.6.3_sig <- res_FEV1_FVC_microb_str_ratio[res_FEV1_FVC_microb_str_ratio$fdr <= 0.05,]

```

## 5.7 OCS use number within 5 years (exacerbation)
```{r}
sum(is.na(dat_tmp$ocs_trim_totnum_5y)) #12
ggplot(dat_tmp, aes(x=ocs_trim_totnum_5y)) + geom_histogram(color="darkblue", fill="lightblue")

```

### 5.7.1 targeted sph/microb
```{r}
list_analysis <- mets_missing_sph_microb[mets_missing_sph_microb$pct_na_met <= 0.3,"feat_id"]
length(list_analysis) #3898

list_tmp <- append("Biobank_Subject_ID", list_analysis)
pheno_tmp <- merge(dat_tmp, df_sph_microb_ratio_log[,..list_tmp], by = "Biobank_Subject_ID", all.x = T)

res_5.7.1 <- poiss_res_fnc(outc = "ocs_trim_totnum_5y", 
                           mets_list = list_analysis, 
                           dat = "pheno_tmp", 
                           add_covar = " + Age_at_Collection + bmi + Gender + race + Ethnicity + ics_trim_totnum_5y")

## Save results
res_ocs_use_sph_microb_ratio <- res_5.7.1
res_ocs_use_sph_microb_ratio <- res_ocs_use_sph_microb_ratio %>% separate(feat_id, sep = "\\.", c("numerator", "denominator"), remove = F)

dict <- sph_mets_info[,c("feat_id", "tentative_annotation", "SUB PATHWAY", "SUPER PATHWAY")]
res_ocs_use_sph_microb_ratio$numerator <- dict$tentative_annotation[match(res_ocs_use_sph_microb_ratio$numerator, dict$feat_id)]
res_ocs_use_sph_microb_ratio$numerator_sub_pathway <- dict$`SUB PATHWAY`[match(res_ocs_use_sph_microb_ratio$numerator, dict$tentative_annotation)]

dict <- microb_mets_info[,c("feat_id", "Component_Name", "SUB_PATHWAY", "SUPER_PATHWAY")]
res_ocs_use_sph_microb_ratio$denominator <- dict$Component_Name[match(res_ocs_use_sph_microb_ratio$denominator, dict$feat_id)]
res_ocs_use_sph_microb_ratio$denominator_sub_pathway <- dict$`SUB_PATHWAY`[match(res_ocs_use_sph_microb_ratio$denominator, dict$`Component_Name`)]

setcolorder(res_ocs_use_sph_microb_ratio, c("feat_id", "numerator", "numerator_sub_pathway", "denominator", "denominator_sub_pathway"))


saveRDS(res_ocs_use_sph_microb_ratio, file = here(res_dir, "res_ocs_use_sph_microb_ratio.rds"))

## display results
### min p-value = 1.63e-9
### #sig = 392/3898
datatable(res_ocs_use_sph_microb_ratio[res_ocs_use_sph_microb_ratio$fdr <= 0.05,], 
          filter = "top") %>% 
          formatSignif(columns = c("beta", "pval", "fdr","se", "zval", "neg_log10_pval"), digits = 3)

res_5.7.1_sig <- res_ocs_use_sph_microb_ratio[res_ocs_use_sph_microb_ratio$fdr <= 0.05,]

res_5.7.1_sig_top20 <- res_5.7.1_sig[c(1:20),] %>% as.data.table()
res_5.7.1_sig_top20$annotation <- paste(res_5.7.1_sig_top20$numerator, res_5.7.1_sig_top20$denominator, sep = " / ")
res_5.7.1_sig_top20$class <- "sph/microb"
res_5.7.1_sig_top20$Pval <- signif(res_5.7.1_sig_top20$pval, 3)
res_5.7.1_sig_top20 <- res_5.7.1_sig_top20[order(denominator_sub_pathway, beta), ]

p_sph_microb_ocs_use <- ggplot(res_5.7.1_sig_top20, 
                aes(reorder(annotation, -beta), y = beta, color = beta >0)) +
   scale_colour_manual(name = '', values = setNames(c('red','blue'),c(T, F)))+
   geom_rect(aes(fill = denominator_sub_pathway),xmin = -Inf,xmax = Inf,
             ymin = -Inf,ymax = Inf, alpha = 0, color = 'black', linejoin = "round") +
   facet_grid(denominator_sub_pathway ~ class, scales = "free_y", space = "free_y", labeller = label_wrap_gen()) +
   geom_errorbar(aes(ymin = lower95, ymax = upper95), width = 0.25) +
   geom_pointrange(aes(ymin = lower95, ymax = upper95), size = 0.25) +
   geom_hline(yintercept = 0, linetype = 2) +
   labs(
     x = "",
     y = ""
   ) +
   theme_minimal() +
   theme_bw() +
   theme(
     title = element_text(size = 16),
     strip.text.x = element_text(size = 12, face = "bold"),
     strip.text.y = element_text(size = 10, face = "bold"),
     axis.title = element_text(size = 12),
     axis.text = element_text(size = 10, face = "bold"),
     legend.position = "",
     legend.title = element_text(size = 12),
     legend.text = element_text(size = 12),
     strip.text.y.right = element_text(angle = 0),
     strip.background = element_rect(fill = "#B2ABD2", colour = "black")) +
   coord_flip() + scale_fill_viridis_d(option = "D", direction = -1, begin = 0, end = 1, alpha = 0.1)
   # scale_y_continuous(limits = c(0.2, 1.6))
```

### 5.7.2 targeted sph/str
```{r}
list_analysis <- mets_missing_sph_str[mets_missing_sph_str$pct_na_met <= 0.3,"feat_id"]
length(list_analysis) #1248

list_tmp <- append("Biobank_Subject_ID", list_analysis)
pheno_tmp <- merge(dat_tmp, df_sph_steroid_ratio_log[,..list_tmp], by = "Biobank_Subject_ID", all.x = T)

res_5.7.2 <- poiss_res_fnc(outc = "ocs_trim_totnum_5y", 
                           mets_list = list_analysis, 
                           dat = "pheno_tmp", 
                           add_covar = " + Age_at_Collection + bmi + Gender + race + Ethnicity + ics_trim_totnum_5y")

## Save results
res_ocs_use_sph_str_ratio <- res_5.7.2
res_ocs_use_sph_str_ratio <- res_ocs_use_sph_str_ratio %>% separate(feat_id, sep = "\\.", c("numerator", "denominator"), remove = F)

dict <- sph_mets_info[,c("feat_id", "tentative_annotation", "SUB PATHWAY", "SUPER PATHWAY")]
res_ocs_use_sph_str_ratio$numerator <- dict$tentative_annotation[match(res_ocs_use_sph_str_ratio$numerator, dict$feat_id)]
res_ocs_use_sph_str_ratio$numerator_sub_pathway <- dict$`SUB PATHWAY`[match(res_ocs_use_sph_str_ratio$numerator, dict$tentative_annotation)]

dict <- str_mets_info[,c("feat_id", "Component Name", "SUB PATHWAY", "SUPER PATHWAY")]
res_ocs_use_sph_str_ratio$denominator <- dict$`Component Name`[match(res_ocs_use_sph_str_ratio$denominator, dict$feat_id)]
res_ocs_use_sph_str_ratio$denominator_sub_pathway <- dict$`SUB PATHWAY`[match(res_ocs_use_sph_str_ratio$denominator, dict$`Component Name`)]

setcolorder(res_ocs_use_sph_str_ratio, c("feat_id", "numerator", "numerator_sub_pathway", "denominator", "denominator_sub_pathway"))


saveRDS(res_ocs_use_sph_str_ratio, file = here(res_dir, "res_ocs_use_sph_str_ratio.rds"))

## display results
### min p-value = 1.63e-26	
### #sig = 748/1248 = 59.94%
datatable(res_ocs_use_sph_str_ratio[res_ocs_use_sph_str_ratio$fdr <= 0.05,], 
          filter = "top") %>% 
          formatSignif(columns = c("beta", "pval", "fdr","se", "zval", "neg_log10_pval"), digits = 3)

res_5.7.2_sig <- res_ocs_use_sph_str_ratio[res_ocs_use_sph_str_ratio$fdr <= 0.05,]


## manhattan plot
res_ocs_use_sph_str_ratio <- as.data.table(res_ocs_use_sph_str_ratio)
res_ocs_use_sph_str_ratio$direction_of_effect <- with(res_ocs_use_sph_str_ratio, ifelse(beta < 0, "-", "+"))
res_ocs_use_sph_str_ratio$CHEMICAL_NAME <- paste(res_ocs_use_sph_str_ratio$numerator, res_ocs_use_sph_str_ratio$denominator, sep = "/")

res_ocs_use_sph_str_ratio <- res_ocs_use_sph_str_ratio[order(denominator_sub_pathway, numerator_sub_pathway, feat_id)]
res_ocs_use_sph_str_ratio$numerator_sub_pathway <- factor(res_ocs_use_sph_str_ratio$numerator_sub_pathway, 
                                              levels = as.vector(unique(res_ocs_use_sph_str_ratio$numerator_sub_pathway)))
res_ocs_use_sph_str_ratio$Name <- factor(res_ocs_use_sph_str_ratio$feat_id, levels = as.vector(res_ocs_use_sph_str_ratio$feat_id))

res_ocs_use_sph_str_ratio$denominator_sub_pathway <- factor(res_ocs_use_sph_str_ratio$denominator_sub_pathway, levels = c("androgens", "glucocorticoids", "progestogens", "estrogens", "mineralocorticoids"))

res_ocs_use_sph_str_ratio$numerator_sub_pathway[res_ocs_use_sph_str_ratio$numerator_sub_pathway == "X3_Ketosphinganine"] <- "3_Ketosphinganine"

save_manh_plot(dat = res_ocs_use_sph_str_ratio, 
               title_infig = "sphingolipid/steroid metabolite ratio association results (ocs use)", 
               fig_name = "/res_ocs_use_sph_str_ratio.png", 
               if_lbl = T, lbl_thld = 1e-30, wdth = 36)

res_5.7.2_sig_top20 <- res_5.7.2_sig[c(1:20),] %>% as.data.table()
res_5.7.2_sig_top20$denominator[res_5.7.2_sig_top20$denominator == "dehydroepiandrosterone sulfate"] <- "DHEAS"
res_5.7.2_sig_top20$annotation <- paste(res_5.7.2_sig_top20$numerator, res_5.7.2_sig_top20$denominator, sep = " / ")
res_5.7.2_sig_top20$class <- "sph/str"
res_5.7.2_sig_top20$Pval <- signif(res_5.7.2_sig_top20$pval, 3)
res_5.7.2_sig_top20 <- res_5.7.2_sig_top20[order(denominator_sub_pathway, beta), ]

p_sph_str_ocs_use <- ggplot(res_5.7.2_sig_top20, 
                aes(reorder(annotation, -beta), y = beta, color = beta >0)) +
   scale_colour_manual(name = '', values = setNames(c('red','blue'),c(T, F)))+
   geom_rect(aes(fill = denominator_sub_pathway),xmin = -Inf,xmax = Inf,
             ymin = -Inf,ymax = Inf, alpha = 0, color = 'black', linejoin = "round") +
   facet_grid(denominator_sub_pathway ~ class, scales = "free_y", space = "free_y", labeller = label_wrap_gen()) +
   geom_errorbar(aes(ymin = lower95, ymax = upper95), width = 0.25) +
   geom_pointrange(aes(ymin = lower95, ymax = upper95), size = 0.25) +
   geom_hline(yintercept = 0, linetype = 2) +
   labs(
     x = "",
     y = ""
   ) +
   theme_minimal() +
   theme_bw() +
   theme(
     title = element_text(size = 16),
     strip.text.x = element_text(size = 12, face = "bold"),
     strip.text.y = element_text(size = 10, face = "bold"),
     axis.title = element_text(size = 12),
     axis.text = element_text(size = 10, face = "bold"),
     legend.position = "",
     legend.title = element_text(size = 12),
     legend.text = element_text(size = 12),
     strip.text.y.right = element_text(angle = 0),
     strip.background = element_rect(fill = "#FDB863", colour = "black")) +
   coord_flip() + scale_fill_viridis_d(option = "D", direction = -1, begin = 0, end = 1, alpha = 0.1)
   # scale_y_continuous(limits = c(0.2, 1.6))

```

#### top ratio quantile plot
ratio
```{r}
## top 1
# quantile
sum(is.na(pheno_tmp$Cer20_1d18_1.S10_steroid))
pheno_tmp_new_1 <- pheno_tmp[!is.na(Cer20_1d18_1.S10_steroid),]

vQuant = quantile(pheno_tmp_new_1$Cer20_1d18_1.S10_steroid, c(0:4/4))
pheno_tmp_new_1$Cer20_1d18_1.S10_steroid_cat = with(pheno_tmp_new_1, 
                                                cut(Cer20_1d18_1.S10_steroid, 
                                                    vQuant, 
                                                    include.lowest = T, 
                                                    labels = c("Q1_ratio", "Q2_ratio", "Q3_ratio", "Q4_ratio")))

pheno_tmp_new_1[, .N, .(Cer20_1d18_1.S10_steroid_cat)][order(Cer20_1d18_1.S10_steroid_cat)][, pct := N/sum(N)*100] %>% print(digits = 3)

data_msd <- pheno_tmp_new_1 %>% 
  group_by(Cer20_1d18_1.S10_steroid_cat) %>%
  summarise_at(vars(ocs_trim_totnum_5y),
               list(mean = ~ mean(. , na.rm = TRUE),
                    sd = ~ sd(. , na.rm = TRUE))) %>% 
  as.data.frame()
data_msd 

## top 2
# quantile
sum(is.na(pheno_tmp$Cer24_1d18_1.S10_steroid))
pheno_tmp_new_2 <- pheno_tmp[!is.na(Cer24_1d18_1.S10_steroid),]

vQuant = quantile(pheno_tmp_new_2$Cer24_1d18_1.S10_steroid, c(0:4/4))
pheno_tmp_new_2$Cer24_1d18_1.S10_steroid_cat = with(pheno_tmp_new_2, 
                                                cut(Cer24_1d18_1.S10_steroid, 
                                                    vQuant, 
                                                    include.lowest = T, 
                                                    labels = c("Q1_ratio", "Q2_ratio", "Q3_ratio", "Q4_ratio")))

pheno_tmp_new_2[, .N, .(Cer24_1d18_1.S10_steroid_cat)][order(Cer24_1d18_1.S10_steroid_cat)][, pct := N/sum(N)*100] %>% print(digits = 3)

data_msd <- pheno_tmp_new_2 %>% 
  group_by(Cer24_1d18_1.S10_steroid_cat) %>%
  summarise_at(vars(ocs_trim_totnum_5y),
               list(mean = ~ mean(. , na.rm = TRUE),
                    sd = ~ sd(. , na.rm = TRUE))) %>% 
  as.data.frame()
data_msd 

## top 3
# quantile
sum(is.na(pheno_tmp$Cer20_0d18_1.S10_steroid))
pheno_tmp_new_3 <- pheno_tmp[!is.na(Cer20_0d18_1.S10_steroid),]

vQuant = quantile(pheno_tmp_new_3$Cer20_0d18_1.S10_steroid, c(0:4/4))
pheno_tmp_new_3$Cer20_0d18_1.S10_steroid_cat = with(pheno_tmp_new_3, 
                                                cut(Cer20_0d18_1.S10_steroid, 
                                                    vQuant, 
                                                    include.lowest = T, 
                                                    labels = c("Q1_ratio", "Q2_ratio", "Q3_ratio", "Q4_ratio")))

pheno_tmp_new_3[, .N, .(Cer20_0d18_1.S10_steroid_cat)][order(Cer20_0d18_1.S10_steroid_cat)][, pct := N/sum(N)*100] %>% print(digits = 3)

data_msd <- pheno_tmp_new_3 %>% 
  group_by(Cer20_0d18_1.S10_steroid_cat) %>%
  summarise_at(vars(ocs_trim_totnum_5y),
               list(mean = ~ mean(. , na.rm = TRUE),
                    sd = ~ sd(. , na.rm = TRUE))) %>% 
  as.data.frame()
data_msd 

## top 4
# quantile
sum(is.na(pheno_tmp$SM14_d18_1.S10_steroid))
pheno_tmp_new_4 <- pheno_tmp[!is.na(SM14_d18_1.S10_steroid),]

vQuant = quantile(pheno_tmp_new_4$SM14_d18_1.S10_steroid, c(0:4/4))
pheno_tmp_new_4$SM14_d18_1.S10_steroid_cat = with(pheno_tmp_new_4, 
                                                cut(SM14_d18_1.S10_steroid, 
                                                    vQuant, 
                                                    include.lowest = T, 
                                                    labels = c("Q1_ratio", "Q2_ratio", "Q3_ratio", "Q4_ratio")))

pheno_tmp_new_4[, .N, .(SM14_d18_1.S10_steroid_cat)][order(SM14_d18_1.S10_steroid_cat)][, pct := N/sum(N)*100] %>% print(digits = 3)

data_msd <- pheno_tmp_new_4 %>% 
  group_by(SM14_d18_1.S10_steroid_cat) %>%
  summarise_at(vars(ocs_trim_totnum_5y),
               list(mean = ~ mean(. , na.rm = TRUE),
                    sd = ~ sd(. , na.rm = TRUE))) %>% 
  as.data.frame()
data_msd 

## top 5
# quantile
sum(is.na(pheno_tmp$Cer18_0d18_1.S10_steroid))
pheno_tmp_new_5 <- pheno_tmp[!is.na(Cer18_0d18_1.S10_steroid),]

vQuant = quantile(pheno_tmp_new_5$Cer18_0d18_1.S10_steroid, c(0:4/4))
pheno_tmp_new_5$Cer18_0d18_1.S10_steroid_cat = with(pheno_tmp_new_5, 
                                                cut(Cer18_0d18_1.S10_steroid, 
                                                    vQuant, 
                                                    include.lowest = T, 
                                                    labels = c("Q1_ratio", "Q2_ratio", "Q3_ratio", "Q4_ratio")))

pheno_tmp_new_5[, .N, .(Cer18_0d18_1.S10_steroid_cat)][order(Cer18_0d18_1.S10_steroid_cat)][, pct := N/sum(N)*100] %>% print(digits = 3)

data_msd <- pheno_tmp_new_5 %>% 
  group_by(Cer18_0d18_1.S10_steroid_cat) %>%
  summarise_at(vars(ocs_trim_totnum_5y),
               list(mean = ~ mean(. , na.rm = TRUE),
                    sd = ~ sd(. , na.rm = TRUE))) %>% 
  as.data.frame()
data_msd 

##
png(here::here(fig_dir, "/boxplots_top_ratio_ocs_use.png"), width = 5000, height = 1000, res = 300)
par(mfrow=c(1,5))
boxplot(ocs_trim_totnum_5y ~ Cer20_1d18_1.S10_steroid_cat, 
        data = pheno_tmp_new_1, 
        xlab = "The level of Cer(d18:1/20:1)	/ DHEAS", 
        ylab = "asthma exacerbation (prevalent 5 years)",
        main = "Cer(d18:1/20:1)/DHEAS")

boxplot(ocs_trim_totnum_5y ~ Cer24_1d18_1.S10_steroid_cat, 
        data = pheno_tmp_new_2, 
        xlab = "The level of Cer(d18:1/24:1) / DHEAS", 
        ylab = "asthma exacerbation (prevalent 5 years)",
        main = "Cer(d18:1/24:1)/DHEAS")

boxplot(ocs_trim_totnum_5y ~ Cer20_0d18_1.S10_steroid_cat, 
        data = pheno_tmp_new_3, 
        xlab = "The level of Cer(d18:1/20:0) / DHEAS", 
        ylab = "asthma exacerbation (prevalent 5 years)",
        main = "Cer(d18:1/20:0)/DHEAS")

boxplot(ocs_trim_totnum_5y ~ SM14_d18_1.S10_steroid_cat, 
        data = pheno_tmp_new_4, 
        xlab = "The level of SM(d18:1/14:0) / DHEAS", 
        ylab = "asthma exacerbation (prevalent 5 years)",
        main = "SM(d18:1/14:0)/DHEAS")

boxplot(ocs_trim_totnum_5y ~ Cer18_0d18_1.S10_steroid_cat, 
        data = pheno_tmp_new_5, 
        xlab = "The level of Cer(d18:1/18:0) / DHEAS", 
        ylab = "asthma exacerbation (prevalent 5 years)",
        main = "Cer(d18:1/18:0)/DHEAS")
# Close the PNG device
dev.off()
```

#### strata denominator and numerator
```{r}
# top 1
vQuant = quantile(pheno_tmp_new_1$S10_steroid, c(0:4/4))
pheno_tmp_new_1$S10_steroid_cat = with(pheno_tmp_new_1, 
                                  cut(S10_steroid, 
                                      vQuant, 
                                      include.lowest = T, 
                                      labels = c("Q1_str", "Q2_str", "Q3_str", "Q4_str")))

pheno_tmp_new_1[, .N, .(S10_steroid_cat)][order(S10_steroid_cat)][, pct := N/sum(N)*100] %>% print(digits = 3)

vQuant = quantile(pheno_tmp_new_1$Cer20_1d18_1, c(0:4/4))
pheno_tmp_new_1$Cer20_1d18_1_cat = with(pheno_tmp_new_1, 
                                 cut(Cer20_1d18_1, 
                                     vQuant, 
                                     include.lowest = T, 
                                     labels = c("Q1_sph", "Q2_sph", "Q3_sph", "Q4_sph")))

pheno_tmp_new_1[, .N, .(Cer20_1d18_1_cat)][order(Cer20_1d18_1_cat)][, pct := N/sum(N)*100] %>% print(digits = 3)

# top 2
vQuant = quantile(pheno_tmp_new_2$S10_steroid, c(0:4/4))
pheno_tmp_new_2$S10_steroid_cat = with(pheno_tmp_new_2, 
                                  cut(S10_steroid, 
                                      vQuant, 
                                      include.lowest = T, 
                                      labels = c("Q1_str", "Q2_str", "Q3_str", "Q4_str")))

pheno_tmp_new_2[, .N, .(S10_steroid_cat)][order(S10_steroid_cat)][, pct := N/sum(N)*100] %>% print(digits = 3)

vQuant = quantile(pheno_tmp_new_2$Cer24_1d18_1, c(0:4/4))
pheno_tmp_new_2$Cer24_1d18_1_cat = with(pheno_tmp_new_2, 
                                 cut(Cer24_1d18_1, 
                                     vQuant, 
                                     include.lowest = T, 
                                     labels = c("Q1_sph", "Q2_sph", "Q3_sph", "Q4_sph")))

pheno_tmp_new_2[, .N, .(Cer24_1d18_1_cat)][order(Cer24_1d18_1_cat)][, pct := N/sum(N)*100] %>% print(digits = 3)

# top 3
vQuant = quantile(pheno_tmp_new_3$S10_steroid, c(0:4/4))
pheno_tmp_new_3$S10_steroid_cat = with(pheno_tmp_new_3, 
                                  cut(S10_steroid, 
                                      vQuant, 
                                      include.lowest = T, 
                                      labels = c("Q1_str", "Q2_str", "Q3_str", "Q4_str")))

pheno_tmp_new_3[, .N, .(S10_steroid_cat)][order(S10_steroid_cat)][, pct := N/sum(N)*100] %>% print(digits = 3)

vQuant = quantile(pheno_tmp_new_3$Cer20_0d18_1, c(0:4/4))
pheno_tmp_new_3$Cer20_0d18_1_cat = with(pheno_tmp_new_3, 
                                 cut(Cer20_0d18_1, 
                                     vQuant, 
                                     include.lowest = T, 
                                     labels = c("Q1_sph", "Q2_sph", "Q3_sph", "Q4_sph")))

pheno_tmp_new_3[, .N, .(Cer20_0d18_1_cat)][order(Cer20_0d18_1_cat)][, pct := N/sum(N)*100] %>% print(digits = 3)

# top 4
vQuant = quantile(pheno_tmp_new_4$S10_steroid, c(0:4/4))
pheno_tmp_new_4$S10_steroid_cat = with(pheno_tmp_new_4, 
                                  cut(S10_steroid, 
                                      vQuant, 
                                      include.lowest = T, 
                                      labels = c("Q1_str", "Q2_str", "Q3_str", "Q4_str")))

pheno_tmp_new_4[, .N, .(S10_steroid_cat)][order(S10_steroid_cat)][, pct := N/sum(N)*100] %>% print(digits = 3)

vQuant = quantile(pheno_tmp_new_4$SM14_d18_1, c(0:4/4))
pheno_tmp_new_4$SM14_d18_1_cat = with(pheno_tmp_new_4, 
                                 cut(SM14_d18_1, 
                                     vQuant, 
                                     include.lowest = T, 
                                     labels = c("Q1_sph", "Q2_sph", "Q3_sph", "Q4_sph")))

pheno_tmp_new_4[, .N, .(SM14_d18_1_cat)][order(SM14_d18_1_cat)][, pct := N/sum(N)*100] %>% print(digits = 3)

# top 5
vQuant = quantile(pheno_tmp_new_5$S10_steroid, c(0:4/4))
pheno_tmp_new_5$S10_steroid_cat = with(pheno_tmp_new_5, 
                                  cut(S10_steroid, 
                                      vQuant, 
                                      include.lowest = T, 
                                      labels = c("Q1_str", "Q2_str", "Q3_str", "Q4_str")))

pheno_tmp_new_5[, .N, .(S10_steroid_cat)][order(S10_steroid_cat)][, pct := N/sum(N)*100] %>% print(digits = 3)

vQuant = quantile(pheno_tmp_new_5$Cer18_0d18_1, c(0:4/4))
pheno_tmp_new_5$Cer18_0d18_1_cat = with(pheno_tmp_new_5, 
                                 cut(Cer18_0d18_1, 
                                     vQuant, 
                                     include.lowest = T, 
                                     labels = c("Q1_sph", "Q2_sph", "Q3_sph", "Q4_sph")))

pheno_tmp_new_5[, .N, .(Cer18_0d18_1_cat)][order(Cer18_0d18_1_cat)][, pct := N/sum(N)*100] %>% print(digits = 3)

## plot
plot1 <- ggplot(pheno_tmp_new_1, aes(x = S10_steroid_cat, y = ocs_trim_totnum_5y, fill = Cer20_1d18_1_cat)) + 
  geom_boxplot() + 
  labs(
    title = "Cer(d18:1/20:1)/DHEAS",
    x = "The quantile of denominator DHEAS",
    y = "Asthma Exacerbation (Prevalent 5 Years)",
    fill = "The quantile of numerator Cer(d18:1/20:1)"
  ) +
theme(
    legend.position = "bottom",
    plot.title = element_text(hjust = 0.5),
    legend.text = element_text(size = 8),  # Adjust the text size of the legend
    legend.title = element_text(size = 10)  # Adjust the text size of the legend title
  ) +
  guides(fill = guide_legend(nrow = 2))  # Set the number of rows in the legend

plot2 <- ggplot(pheno_tmp_new_2, aes(x = S10_steroid_cat, y = ocs_trim_totnum_5y, fill = Cer24_1d18_1_cat)) + 
  geom_boxplot() + 
  labs(
    title = "Cer(d18:1/24:1)/DHEAS",
    x = "The quantile of denominator DHEAS",
    y = "Asthma Exacerbation (Prevalent 5 Years)",
    fill = "The quantile of numerator Cer(d18:1/24:1)"
  ) +
theme(
    legend.position = "bottom",
    plot.title = element_text(hjust = 0.5),
    legend.text = element_text(size = 8),  # Adjust the text size of the legend
    legend.title = element_text(size = 10)  # Adjust the text size of the legend title
  ) +
  guides(fill = guide_legend(nrow = 2))  # Set the number of rows in the legend

plot3 <- ggplot(pheno_tmp_new_3, aes(x = S10_steroid_cat, y = ocs_trim_totnum_5y, fill = Cer20_0d18_1_cat)) + 
  geom_boxplot() + 
  labs(
    title = "Cer(d18:1/20:0)/DHEAS",
    x = "The quantile of denominator DHEAS",
    y = "Asthma Exacerbation (Prevalent 5 Years)",
    fill = "The quantile of numerator Cer(d18:1/20:0)"
  ) +
theme(
    legend.position = "bottom",
    plot.title = element_text(hjust = 0.5),
    legend.text = element_text(size = 8),  # Adjust the text size of the legend
    legend.title = element_text(size = 10)  # Adjust the text size of the legend title
  ) +
  guides(fill = guide_legend(nrow = 2))  # Set the number of rows in the legend


plot4 <- ggplot(pheno_tmp_new_4, aes(x = S10_steroid_cat, y = ocs_trim_totnum_5y, fill = SM14_d18_1_cat)) + 
  geom_boxplot() + 
  labs(
    title = "SM(d18:1/14:0)/DHEAS",
    x = "The quantile of denominator DHEAS",
    y = "Asthma Exacerbation (Prevalent 5 Years)",
    fill = "The quantile of numerator SM(d18:1/14:0)"
  ) +
theme(
    legend.position = "bottom",
    plot.title = element_text(hjust = 0.5),
    legend.text = element_text(size = 8),  # Adjust the text size of the legend
    legend.title = element_text(size = 10)  # Adjust the text size of the legend title
  ) +
  guides(fill = guide_legend(nrow = 2))  # Set the number of rows in the legend


plot5 <- ggplot(pheno_tmp_new_5, aes(x = S10_steroid_cat, y = ocs_trim_totnum_5y, fill = Cer18_0d18_1_cat)) + 
  geom_boxplot() + 
  labs(
    title = "Cer(d18:1/18:0)/DHEAS",
    x = "The quantile of denominator DHEAS",
    y = "Asthma Exacerbation (Prevalent 5 Years)",
    fill = "The quantile of numerator Cer(d18:1/18:0)"
  ) +
theme(
    legend.position = "bottom",
    plot.title = element_text(hjust = 0.5),
    legend.text = element_text(size = 8),  # Adjust the text size of the legend
    legend.title = element_text(size = 10)  # Adjust the text size of the legend title
  ) +
  guides(fill = guide_legend(nrow = 2))  # Set the number of rows in the legend

png(here::here(fig_dir, "/boxplots_top_den_num_ocs_use.png"), width = 6800, height = 1560, res = 300)
grid.arrange(plot1, plot2, plot3, plot4, plot5, ncol = 5)
# Close the PNG device
dev.off()
```

### 5.7.3 targeted microb/str
```{r}
list_analysis <- mets_missing_microb_str[mets_missing_microb_str$pct_na_met <= 0.3,"feat_id"]
length(list_analysis) #794

list_tmp <- append("Biobank_Subject_ID", list_analysis)
pheno_tmp <- merge(dat_tmp, df_microb_steroid_ratio_log[,..list_tmp], by = "Biobank_Subject_ID", all.x = T)

res_5.7.3 <- poiss_res_fnc(outc = "ocs_trim_totnum_5y", 
                           mets_list = list_analysis, 
                           dat = "pheno_tmp", 
                           add_covar = " + Age_at_Collection + bmi + Gender + race + Ethnicity + ics_trim_totnum_5y")

## Save results
res_ocs_use_microb_str_ratio <- res_5.7.3
res_ocs_use_microb_str_ratio <- res_ocs_use_microb_str_ratio %>% separate(feat_id, sep = "\\.", c("numerator", "denominator"), remove = F)

dict <- microb_mets_info[,c("feat_id", "Component_Name", "SUB_PATHWAY", "SUPER_PATHWAY")]
res_ocs_use_microb_str_ratio$numerator <- dict$Component_Name[match(res_ocs_use_microb_str_ratio$numerator, dict$feat_id)]
res_ocs_use_microb_str_ratio$numerator_sub_pathway <- dict$SUB_PATHWAY[match(res_ocs_use_microb_str_ratio$numerator, dict$Component_Name)]

dict <- str_mets_info[,c("feat_id", "Component Name", "SUB PATHWAY", "SUPER PATHWAY")]
res_ocs_use_microb_str_ratio$denominator <- dict$`Component Name`[match(res_ocs_use_microb_str_ratio$denominator, dict$feat_id)]
res_ocs_use_microb_str_ratio$denominator_sub_pathway <- dict$`SUB PATHWAY`[match(res_ocs_use_microb_str_ratio$denominator, dict$`Component Name`)]

setcolorder(res_ocs_use_microb_str_ratio, c("feat_id", "numerator", "numerator_sub_pathway", "denominator", "denominator_sub_pathway"))


saveRDS(res_ocs_use_microb_str_ratio, file = here(res_dir, "res_ocs_use_microb_str_ratio.rds"))

## display results
datatable(res_ocs_use_microb_str_ratio[res_ocs_use_microb_str_ratio$fdr <= 0.05,], 
          filter = "top") %>% 
          formatSignif(columns = c("beta", "pval", "fdr","se", "zval", "neg_log10_pval"), digits = 3)

res_5.7.3_sig <- res_ocs_use_microb_str_ratio[res_ocs_use_microb_str_ratio$fdr <= 0.05,]

res_5.7.3_sig_top20 <- res_5.7.3_sig[c(1:20),] %>% as.data.table()
res_5.7.3_sig_top20$denominator[res_5.7.3_sig_top20$denominator == "dehydroepiandrosterone sulfate"] <- "DHEAS"
res_5.7.3_sig_top20$annotation <- paste(res_5.7.3_sig_top20$numerator, res_5.7.3_sig_top20$denominator, sep = " / ")
res_5.7.3_sig_top20$class <- "microb/str"
res_5.7.3_sig_top20$Pval <- signif(res_5.7.3_sig_top20$pval, 3)
res_5.7.3_sig_top20 <- res_5.7.3_sig_top20[order(denominator_sub_pathway, beta), ]

p_microb_str_ocs_use <- ggplot(res_5.7.3_sig_top20, 
                aes(reorder(annotation, -beta), y = beta, color = beta >0)) +
   scale_colour_manual(name = '', values = setNames(c('red','blue'),c(T, F)))+
   geom_rect(aes(fill = denominator_sub_pathway),xmin = -Inf,xmax = Inf,
             ymin = -Inf,ymax = Inf, alpha = 0, color = 'black', linejoin = "round") +
   facet_grid(denominator_sub_pathway ~ class, scales = "free_y", space = "free_y", labeller = label_wrap_gen()) +
   geom_errorbar(aes(ymin = lower95, ymax = upper95), width = 0.25) +
   geom_pointrange(aes(ymin = lower95, ymax = upper95), size = 0.25) +
   geom_hline(yintercept = 0, linetype = 2) +
   labs(
     x = "",
     y = ""
   ) +
   theme_minimal() +
   theme_bw() +
   theme(
     title = element_text(size = 16),
     strip.text.x = element_text(size = 12, face = "bold"),
     strip.text.y = element_text(size = 10, face = "bold"),
     axis.title = element_text(size = 12),
     axis.text = element_text(size = 10, face = "bold"),
     legend.position = "",
     legend.title = element_text(size = 12),
     legend.text = element_text(size = 12),
     strip.text.y.right = element_text(angle = 0),
     strip.background = element_rect(fill = "#B8E186", colour = "black")) +
   coord_flip() + scale_fill_viridis_d(option = "D", direction = -1, begin = 0, end = 1, alpha = 0.1)
   # scale_y_continuous(limits = c(0.2, 1.6))

```


## 5.8 OCS use number after 5 years (exacerbation)
```{r}
sum(is.na(dat_tmp$ocs_trim_totnum_5y_after)) #56
ggplot(dat_tmp, aes(x=ocs_trim_totnum_5y_after)) + geom_histogram(color="darkblue", fill="lightblue")

```

### 5.8.1 targeted sph/microb
```{r}
list_analysis <- mets_missing_sph_microb[mets_missing_sph_microb$pct_na_met <= 0.3,"feat_id"]
length(list_analysis) #3898

list_tmp <- append("Biobank_Subject_ID", list_analysis)
pheno_tmp <- merge(dat_tmp, df_sph_microb_ratio_log[,..list_tmp], by = "Biobank_Subject_ID", all.x = T)

res_5.8.1 <- poiss_res_fnc(outc = "ocs_trim_totnum_5y_after", 
                           mets_list = list_analysis, 
                           dat = "pheno_tmp", 
                           add_covar = " + Age_at_Collection + bmi + Gender + race + Ethnicity + ics_trim_totnum_5y_after")

## Save results
res_ocs_use_after_sph_microb_ratio <- res_5.8.1
res_ocs_use_after_sph_microb_ratio <- res_ocs_use_after_sph_microb_ratio %>% separate(feat_id, sep = "\\.", c("numerator", "denominator"), remove = F)

dict <- sph_mets_info[,c("feat_id", "tentative_annotation", "SUB PATHWAY", "SUPER PATHWAY")]
res_ocs_use_after_sph_microb_ratio$numerator <- dict$tentative_annotation[match(res_ocs_use_after_sph_microb_ratio$numerator, dict$feat_id)]
res_ocs_use_after_sph_microb_ratio$numerator_sub_pathway <- dict$`SUB PATHWAY`[match(res_ocs_use_after_sph_microb_ratio$numerator, dict$tentative_annotation)]

dict <- microb_mets_info[,c("feat_id", "Component_Name", "SUB_PATHWAY", "SUPER_PATHWAY")]
res_ocs_use_after_sph_microb_ratio$denominator <- dict$Component_Name[match(res_ocs_use_after_sph_microb_ratio$denominator, dict$feat_id)]
res_ocs_use_after_sph_microb_ratio$denominator_sub_pathway <- dict$`SUB_PATHWAY`[match(res_ocs_use_after_sph_microb_ratio$denominator, dict$`Component_Name`)]

setcolorder(res_ocs_use_after_sph_microb_ratio, c("feat_id", "numerator", "numerator_sub_pathway", "denominator", "denominator_sub_pathway"))


saveRDS(res_ocs_use_after_sph_microb_ratio, file = here(res_dir, "res_ocs_use_after_sph_microb_ratio.rds"))

## display results
### min p-value = #	
### #sig = 88/3898 = 2.26%
datatable(res_ocs_use_after_sph_microb_ratio[res_ocs_use_after_sph_microb_ratio$fdr <= 0.05,], 
          filter = "top") %>% 
          formatSignif(columns = c("beta", "pval", "fdr","se", "zval", "neg_log10_pval"), digits = 3)

res_5.8.1_sig <- res_ocs_use_after_sph_microb_ratio[res_ocs_use_after_sph_microb_ratio$fdr <= 0.05,]

```

### 5.8.2 targeted sph/str
```{r}
list_analysis <- mets_missing_sph_str[mets_missing_sph_str$pct_na_met <= 0.3,"feat_id"]
length(list_analysis) #1248

list_tmp <- append("Biobank_Subject_ID", list_analysis)
pheno_tmp <- merge(dat_tmp, df_sph_steroid_ratio_log[,..list_tmp], by = "Biobank_Subject_ID", all.x = T)

res_5.8.2 <- poiss_res_fnc(outc = "ocs_trim_totnum_5y_after", 
                           mets_list = list_analysis, 
                           dat = "pheno_tmp", 
                           add_covar = " + Age_at_Collection + bmi + Gender + race + Ethnicity + ics_trim_totnum_5y_after")

## Save results
res_ocs_use_after_sph_str_ratio <- res_5.8.2
res_ocs_use_after_sph_str_ratio <- res_ocs_use_after_sph_str_ratio %>% separate(feat_id, sep = "\\.", c("numerator", "denominator"), remove = F)

dict <- sph_mets_info[,c("feat_id", "tentative_annotation", "SUB PATHWAY", "SUPER PATHWAY")]
res_ocs_use_after_sph_str_ratio$numerator <- dict$tentative_annotation[match(res_ocs_use_after_sph_str_ratio$numerator, dict$feat_id)]
res_ocs_use_after_sph_str_ratio$numerator_sub_pathway <- dict$`SUB PATHWAY`[match(res_ocs_use_after_sph_str_ratio$numerator, dict$tentative_annotation)]

dict <- str_mets_info[,c("feat_id", "Component Name", "SUB PATHWAY", "SUPER PATHWAY")]
res_ocs_use_after_sph_str_ratio$denominator <- dict$`Component Name`[match(res_ocs_use_after_sph_str_ratio$denominator, dict$feat_id)]
res_ocs_use_after_sph_str_ratio$denominator_sub_pathway <- dict$`SUB PATHWAY`[match(res_ocs_use_after_sph_str_ratio$denominator, dict$`Component Name`)]

setcolorder(res_ocs_use_after_sph_str_ratio, c("feat_id", "numerator", "numerator_sub_pathway", "denominator", "denominator_sub_pathway"))


saveRDS(res_ocs_use_after_sph_str_ratio, file = here(res_dir, "res_ocs_use_after_sph_str_ratio.rds"))

## display results
### min p-value = 6.31e-13	
### #sig = 442/1248
datatable(res_ocs_use_after_sph_str_ratio[res_ocs_use_after_sph_str_ratio$fdr <= 0.05,], 
          filter = "top") %>% 
          formatSignif(columns = c("beta", "pval", "fdr","se", "zval", "neg_log10_pval"), digits = 3)

res_5.8.2_sig <- res_ocs_use_after_sph_str_ratio[res_ocs_use_after_sph_str_ratio$fdr <= 0.05,]

## manhattan plot
res_ocs_use_after_sph_str_ratio <- as.data.table(res_ocs_use_after_sph_str_ratio)
res_ocs_use_after_sph_str_ratio$direction_of_effect <- with(res_ocs_use_after_sph_str_ratio, ifelse(beta < 0, "-", "+"))
res_ocs_use_after_sph_str_ratio$CHEMICAL_NAME <- paste(res_ocs_use_after_sph_str_ratio$numerator, res_ocs_use_after_sph_str_ratio$denominator, sep = "/")

res_ocs_use_after_sph_str_ratio <- res_ocs_use_after_sph_str_ratio[order(denominator_sub_pathway, numerator_sub_pathway, feat_id)]
res_ocs_use_after_sph_str_ratio$numerator_sub_pathway <- factor(res_ocs_use_after_sph_str_ratio$numerator_sub_pathway, 
                                              levels = as.vector(unique(res_ocs_use_after_sph_str_ratio$numerator_sub_pathway)))
res_ocs_use_after_sph_str_ratio$Name <- factor(res_ocs_use_after_sph_str_ratio$feat_id, levels = as.vector(res_ocs_use_after_sph_str_ratio$feat_id))

save_manh_plot(dat = res_ocs_use_after_sph_str_ratio, 
               title_infig = "sphingolipid/steroid metabolite ratio association results (ocs use after)", 
               fig_name = "/res_ocs_use_after_sph_str_ratio.png", 
               if_lbl = T, lbl_thld = 1e-10, wdth = 36)
```



### 5.8.3 targeted microb/str
```{r}
list_analysis <- mets_missing_microb_str[mets_missing_microb_str$pct_na_met <= 0.3,"feat_id"]
length(list_analysis) #794

list_tmp <- append("Biobank_Subject_ID", list_analysis)
pheno_tmp <- merge(dat_tmp, df_microb_steroid_ratio_log[,..list_tmp], by = "Biobank_Subject_ID", all.x = T)

res_5.8.3 <- poiss_res_fnc(outc = "ocs_trim_totnum_5y_after", 
                           mets_list = list_analysis, 
                           dat = "pheno_tmp", 
                           add_covar = " + Age_at_Collection + bmi + Gender + race + Ethnicity + ics_trim_totnum_5y_after")

## Save results
res_ocs_use_after_microb_str_ratio <- res_5.8.3
res_ocs_use_after_microb_str_ratio <- res_ocs_use_after_microb_str_ratio %>% separate(feat_id, sep = "\\.", c("numerator", "denominator"), remove = F)

dict <- microb_mets_info[,c("feat_id", "Component_Name", "SUB_PATHWAY", "SUPER_PATHWAY")]
res_ocs_use_after_microb_str_ratio$numerator <- dict$Component_Name[match(res_ocs_use_after_microb_str_ratio$numerator, dict$feat_id)]
res_ocs_use_after_microb_str_ratio$numerator_sub_pathway <- dict$SUB_PATHWAY[match(res_ocs_use_after_microb_str_ratio$numerator, dict$Component_Name)]

dict <- str_mets_info[,c("feat_id", "Component Name", "SUB PATHWAY", "SUPER PATHWAY")]
res_ocs_use_after_microb_str_ratio$denominator <- dict$`Component Name`[match(res_ocs_use_after_microb_str_ratio$denominator, dict$feat_id)]
res_ocs_use_after_microb_str_ratio$denominator_sub_pathway <- dict$`SUB PATHWAY`[match(res_ocs_use_after_microb_str_ratio$denominator, dict$`Component Name`)]

setcolorder(res_ocs_use_after_microb_str_ratio, c("feat_id", "numerator", "numerator_sub_pathway", "denominator", "denominator_sub_pathway"))


saveRDS(res_ocs_use_after_microb_str_ratio, file = here(res_dir, "res_ocs_use_after_microb_str_ratio.rds"))

## display results
datatable(res_ocs_use_after_microb_str_ratio[res_ocs_use_after_microb_str_ratio$fdr <= 0.05,], 
          filter = "top") %>% 
          formatSignif(columns = c("beta", "pval", "fdr","se", "zval", "neg_log10_pval"), digits = 3)

res_5.8.3_sig <- res_ocs_use_after_microb_str_ratio[res_ocs_use_after_microb_str_ratio$fdr <= 0.05,]

```



# 6. Save Data
```{r}

write.csv(res_5.1.1_sig, file = str_c(res_dir, "PEGASUS_sig_targeted_sph_microb_asthma.csv"), row.names = FALSE)
write.csv(res_5.1.2_sig, file = str_c(res_dir, "PEGASUS_sig_targeted_sph_str_asthma.csv"), row.names = FALSE)
write.csv(res_5.1.3_sig, file = str_c(res_dir, "PEGASUS_sig_targeted_microb_str_asthma.csv"), row.names = FALSE)

write.csv(res_5.3.1_sig, file = str_c(res_dir, "PEGASUS_sig_targeted_sph_microb_log_ige.csv"), row.names = FALSE)
write.csv(res_5.3.2_sig, file = str_c(res_dir, "PEGASUS_sig_targeted_sph_str_log_ige.csv"), row.names = FALSE)
write.csv(res_5.3.3_sig, file = str_c(res_dir, "PEGASUS_sig_targeted_microb_str_log_ige.csv"), row.names = FALSE)

write.csv(res_5.4.1_sig, file = str_c(res_dir, "PEGASUS_sig_targeted_sph_microb_FEV1.csv"), row.names = FALSE)
write.csv(res_5.4.2_sig, file = str_c(res_dir, "PEGASUS_sig_targeted_sph_str_FEV1.csv"), row.names = FALSE)
write.csv(res_5.4.3_sig, file = str_c(res_dir, "PEGASUS_sig_targeted_microb_str_FEV1.csv"), row.names = FALSE)

write.csv(res_5.5.1_sig, file = str_c(res_dir, "PEGASUS_sig_targeted_sph_microb_FVC.csv"), row.names = FALSE)
write.csv(res_5.5.2_sig, file = str_c(res_dir, "PEGASUS_sig_targeted_sph_str_FVC.csv"), row.names = FALSE)
write.csv(res_5.5.3_sig, file = str_c(res_dir, "PEGASUS_sig_targeted_microb_str_FVC.csv"), row.names = FALSE)

write.csv(res_5.6.1_sig, file = str_c(res_dir, "PEGASUS_sig_targeted_sph_microb_FEV1_FVC.csv"), row.names = FALSE)
write.csv(res_5.6.2_sig, file = str_c(res_dir, "PEGASUS_sig_targeted_sph_str_FEV1_FVC.csv"), row.names = FALSE)
write.csv(res_5.6.3_sig, file = str_c(res_dir, "PEGASUS_sig_targeted_microb_str_FEV1_FVC.csv"), row.names = FALSE)

write.csv(res_5.7.1_sig, file = str_c(res_dir, "PEGASUS_sig_targeted_sph_microb_ocs_use.csv"), row.names = FALSE)
write.csv(res_5.7.2_sig, file = str_c(res_dir, "PEGASUS_sig_targeted_sph_str_ocs_use.csv"), row.names = FALSE)
write.csv(res_5.7.3_sig, file = str_c(res_dir, "PEGASUS_sig_targeted_microb_str_ocs_use.csv"), row.names = FALSE)

write.csv(res_5.8.1_sig, file = str_c(res_dir, "PEGASUS_sig_targeted_sph_microb_ocs_use_after.csv"), row.names = FALSE)
write.csv(res_5.8.2_sig, file = str_c(res_dir, "PEGASUS_sig_targeted_sph_str_ocs_use_after.csv"), row.names = FALSE)
write.csv(res_5.8.3_sig, file = str_c(res_dir, "PEGASUS_sig_targeted_microb_str_ocs_use_after.csv"), row.names = FALSE)

## sph/microb
sph_microb_asthma_sig <- res_5.1.1_sig[c("feat_id", "outc", "numerator", "numerator_sub_pathway",
                                         "denominator", "denominator_sub_pathway",
                                         "pval", "fdr", "beta", "lower95", "upper95", "totaln",  "or")]
sph_microb_fev1_sig <- res_5.4.1_sig[c("feat_id", "outc", "numerator", "numerator_sub_pathway",
                                        "denominator", "denominator_sub_pathway",
                                         "pval", "fdr", "beta", "lower95", "upper95", "totaln")]
sph_microb_fev1_sig$or <- NA
sph_microb_fvc_sig <- res_5.5.1_sig[c("feat_id", "outc", "numerator", "numerator_sub_pathway",
                                        "denominator", "denominator_sub_pathway",
                                         "pval", "fdr", "beta", "lower95", "upper95", "totaln")]
sph_microb_fvc_sig$or <- NA
sph_microb_ocs_sig <- res_5.7.1_sig[c("feat_id", "outc", "numerator", "numerator_sub_pathway",
                                        "denominator", "denominator_sub_pathway",
                                         "pval", "fdr", "beta", "lower95", "upper95", "totaln")]
sph_microb_ocs_sig$or <- NA
sph_microb_sig <- rbind(sph_microb_asthma_sig, sph_microb_fev1_sig, sph_microb_fvc_sig, sph_microb_ocs_sig)
sph_microb_sig <- sph_microb_sig[c("feat_id", "outc", "numerator", "numerator_sub_pathway",
                                         "denominator", "denominator_sub_pathway",
                                         "pval", "fdr", "or", "beta", "lower95", "upper95", "totaln")]

write.csv(sph_microb_sig, file = str_c(res_dir, "PEGASUS_sig_targeted_sph_microb_all.csv"), row.names = FALSE)

## sph/str
sph_str_asthma_sig <- res_5.1.2_sig[c("feat_id", "outc", "numerator", "numerator_sub_pathway",
                                         "denominator", "denominator_sub_pathway",
                                         "pval", "fdr", "beta", "lower95", "upper95", "totaln",  "or")]
sph_str_ige_sig <- res_5.3.2_sig[c("feat_id", "outc", "numerator", "numerator_sub_pathway",
                                        "denominator", "denominator_sub_pathway",
                                         "pval", "fdr", "beta", "lower95", "upper95", "totaln")]
sph_str_ige_sig$or <- NA
sph_str_ocs_sig <- res_5.7.2_sig[c("feat_id", "outc", "numerator", "numerator_sub_pathway",
                                        "denominator", "denominator_sub_pathway",
                                         "pval", "fdr", "beta", "lower95", "upper95", "totaln")]
sph_str_ocs_sig$or <- NA
sph_str_sig <- rbind(sph_str_asthma_sig, sph_str_ige_sig, sph_str_ocs_sig)
sph_str_sig <- sph_str_sig[c("feat_id", "outc", "numerator", "numerator_sub_pathway",
                                         "denominator", "denominator_sub_pathway",
                                         "pval", "fdr", "or", "beta", "lower95", "upper95", "totaln")]

write.csv(sph_str_sig, file = str_c(res_dir, "PEGASUS_sig_targeted_sph_str_all.csv"), row.names = FALSE)

## microb/str
microb_str_asthma_sig <- res_5.1.3_sig[c("feat_id", "outc", "numerator", "numerator_sub_pathway",
                                         "denominator", "denominator_sub_pathway",
                                         "pval", "fdr", "beta", "lower95", "upper95", "totaln",  "or")]
microb_str_fev1_sig <- res_5.4.3_sig[c("feat_id", "outc", "numerator", "numerator_sub_pathway",
                                        "denominator", "denominator_sub_pathway",
                                         "pval", "fdr", "beta", "lower95", "upper95", "totaln")]
microb_str_fev1_sig$or <- NA
microb_str_fvc_sig <- res_5.5.3_sig[c("feat_id", "outc", "numerator", "numerator_sub_pathway",
                                        "denominator", "denominator_sub_pathway",
                                         "pval", "fdr", "beta", "lower95", "upper95", "totaln")]
microb_str_fvc_sig$or <- NA
microb_str_ige_sig <- res_5.3.3_sig[c("feat_id", "outc", "numerator", "numerator_sub_pathway",
                                        "denominator", "denominator_sub_pathway",
                                         "pval", "fdr", "beta", "lower95", "upper95", "totaln")]
microb_str_ige_sig$or <- NA
microb_str_ocs_sig <- res_5.7.3_sig[c("feat_id", "outc", "numerator", "numerator_sub_pathway",
                                        "denominator", "denominator_sub_pathway",
                                         "pval", "fdr", "beta", "lower95", "upper95", "totaln")]
microb_str_ocs_sig$or <- NA
microb_str_sig <- rbind(microb_str_asthma_sig, microb_str_fev1_sig, microb_str_fvc_sig, microb_str_ige_sig, microb_str_ocs_sig)
microb_str_sig <- microb_str_sig[c("feat_id", "outc", "numerator", "numerator_sub_pathway",
                                         "denominator", "denominator_sub_pathway",
                                         "pval", "fdr", "or", "beta", "lower95", "upper95", "totaln")]

write.csv(microb_str_sig, file = str_c(res_dir, "PEGASUS_sig_targeted_microb_str_all.csv"), row.names = FALSE)

```

# 7. Sig Percent Plot
```{r}
sig_pct_df <- read_xlsx(str_c(pegasus_dir, "PEGASUS_sig_pert.xlsx"), sheet = "Sheet1") %>% as.data.table()
colnames(sig_pct_df)[1] <- "Ratio_group"
sig_pct_long_df <-  melt(sig_pct_df, id.vars = "Ratio_group", variable.name = "Phenotypes", value.name = "Pct")

plot1 <- ggplot(sig_pct_long_df, aes(x = Phenotypes, y = Pct, fill = Ratio_group)) +
         geom_bar(stat = "identity", position = position_dodge(width = 0.7), width = 0.7) +
         # ggtitle("Bar Plot with Different Groups") +
         xlab("Asthma Phenotypes") +
         ylab("Significant Percentage") +
         theme(legend.position = "none") +
         theme(panel.background = element_rect(fill = "white"),
          plot.background = element_rect(fill = "white", color = "white"),
          panel.grid.major = element_blank(),
          panel.grid.minor = element_blank(),
          axis.title.y = element_text(face = "bold"),
          axis.text.y = element_text(face = "bold"))

ggsave(file = here::here(fig_dir, "/sig_mets_bar_plot.png"), plot = plot1, width = 10, height = 4)
```

# 8. Merge figures
```{r}
## targeted sph/microb
png(file=str_c(fig_dir, "PEGASUS_targeted_sph_microb_phenotypes.png"), width = 42, height = 7.5, units = "in", res = 300)
grid.arrange(p_sph_microb_asthma, p_sph_microb_FEV1, p_sph_microb_FVC, p_sph_microb_FVC, p_sph_microb_FVC,  p_sph_microb_ocs_use, nrow = 1)
dev.off()

## targeted sph/str
png(file=str_c(fig_dir, "PEGASUS_targeted_sph_str_phenotypes.png"), width = 42, height = 7.5, units = "in", res = 300)
grid.arrange(p_sph_str_asthma, p_sph_str_asthma, p_sph_str_asthma, p_sph_str_asthma, p_sph_str_log_ige, p_sph_str_ocs_use, nrow = 1)
dev.off()

## targeted microb/str
png(file=str_c(fig_dir, "PEGASUS_targeted_microb_str_phenotypes.png"), width = 42, height = 7.5, units = "in", res = 300)
grid.arrange(p_microb_str_asthma, p_microb_str_FEV1, p_microb_str_FVC, p_microb_str_FVC, p_microb_str_log_ige, p_microb_str_ocs_use, nrow = 1)
dev.off()

```

# Session info
```{r}
sessionInfo()

```
