==> 7_ODOLLFA_targeted_ratio_association.Rmd <==
---
title: "Regression analysis of targeted ratios in ODOLLFA"
author: "Yulu Chen"
date: "10/30/2023"
output: 
  html_document: 
    toc: true
    toc_float: 
      collapsed: false
      smooth_scroll: false
    code_folding: hide
editor_options: 
  chunk_output_type: console
---

```{r, include = FALSE}
##### 1. Setup #####
## Packages

pkg <- c("tidyverse", "data.table", "here", "survival", "gtools", "tableone", "DT", "sjPlot", 
         "sjmisc", "grid", "gridExtra", "ggpubr", "table1", "readxl")
for (p in pkg) {
        if (require(p, character.only = T)) {
                print(paste0(p, " loaded successfully"))
        } else {
                install.packages(p)
                require(p, character.only = T)
                print(paste0(p, " downloaded and loaded successfully"))
        }
}

## Paths
dat_dir <- "/udd/reyul/mets_ratio_CodeReview/data/ODOLLFA/"
fig_dir <- str_c("/udd/reyul/mets_ratio_CodeReview/", "figures/7_ODOLLFA_targeted_ratio_association/")
res_dir <- str_c("/udd/reyul/mets_ratio_CodeReview/", "results/7_ODOLLFA_targeted_ratio_association/")

## Hard-coded numbers

missing_thld <- 0.30
alpha_thld <- 0.05
z_95 <- qnorm(p = 0.975)

## Functions
pvalue <- function(x, ...) {
    # Construct vectors of data y, and groups (strata) g
    y <- unlist(x)
    g <- factor(rep(1:length(x), times=sapply(x, length)))
    if (is.numeric(y)) {
        # For numeric variables, perform a standard 2-sample t-test
        p <- as.data.frame(summary(aov(y ~ g))[[1]])[1,5]
    } else {
        # For categorical variables, perform a chi-squared test of independence
        p <- chisq.test(table(y, g))$p.value
    }
    # Format the p-value, using an HTML entity for the less-than sign.
    # The initial empty string places the output on the line below the variable label.
    c("", sub("<", "&lt;", format.pval(p, digits=6, eps=0.000001)))
}

```


# 1. Load data
## 1.1 phenotypes

```{r}
# ICS and OCS prior to sample collection
med_cs_smry_pre <- fread(here(dat_dir, "/medication/pre/med_corticosteroids_summary.csv")) %>% as.data.table()
length(unique(med_cs_smry_pre$Biobank_Subject_ID)) #1695

# ICS and OCS after sample collection
med_cs_smry_after <- fread(here(dat_dir, "/medication/after/med_corticosteroids_after_sample_collection_summary.csv")) %>% as.data.table()
length(unique(med_cs_smry_after$Biobank_Subject_ID)) #1521

med_cs_smry_after <- med_cs_smry_after[,c("EMPI", "Biobank_Subject_ID", "Collect_Date", 
                                          "ics_trim_totnum", "ocs_trim_totnum", "ics_trim_totnum_1y", "ocs_trim_totnum_1y",
                                          "ics_trim_totnum_2y", "ocs_trim_totnum_2y", "ics_trim_totnum_3y", "ocs_trim_totnum_3y",
                                          "ics_trim_totnum_5y", "ocs_trim_totnum_5y", "ics_trim_totnum_10y", "ocs_trim_totnum_10y")]
colnames(med_cs_smry_after)[4:15] <-paste(colnames(med_cs_smry_after)[4:15], "after",sep="_")

ics_after_df <- fread(here(dat_dir, "/medication/after/med_first_ics_after_sample_collection.csv"))
ics_after_df <- ics_after_df[,c("EMPI", "Medication_Date", "Collect_Date", "ics", "cs_timegap_d", "cs_timegap_y")]
colnames(ics_after_df)[2:6] <-paste(colnames(ics_after_df)[2:6], "ics_after",sep="_")

ocs_after_df <- fread(here(dat_dir, "/medication/after/med_first_ocs_after_sample_collection.csv"))
ocs_after_df <- ocs_after_df[,c("EMPI", "Medication_Date", "Collect_Date", "ocs", "cs_timegap_d", "cs_timegap_y")]
colnames(ocs_after_df)[2:6] <-paste(colnames(ocs_after_df)[2:6], "ocs_after",sep="_")

#eth data
eth_df <- fread(str_c(dat_dir, "/BiobankPortal_yc746_2022-01-10-205252_eth.csv")) %>% as.data.table()
eth_df <- eth_df[, -c(1,4)]
colnames(eth_df)[1] <- "Biobank_Subject_ID"


# merge data
pool <- merge(med_cs_smry_pre, eth_df, by = "Biobank_Subject_ID", all.x = T)
pool <- merge(pool, med_cs_smry_after, by = c("EMPI", "Biobank_Subject_ID", "Collect_Date"))
pool <- merge(pool, ics_after_df, by = "EMPI", all.x = T)
pool <- merge(pool, ocs_after_df, by = "EMPI", all.x = T)
```

## 1.2 sample data
```{r}
sample_df <- read_excel(here(dat_dir, "/OODOLFA Mets_QC.xlsx"), sheet = 1) %>% as.data.table()

which(colnames(sample_df) == "CLIENT_SAMPLE_ID")
colnames(sample_df)[9] <- "Biobank_Subject_ID"
sample_df$Biobank_Subject_ID <- as.integer(sample_df$Biobank_Subject_ID)

```

## 1.3 raw metabolon data
```{r}
mets_info <- read_excel(here(dat_dir, "/OODOLFA Mets_QC.xlsx"), sheet = 2) %>% as.data.table()

mets_raw_df <- read_excel(here(dat_dir, "/BRIG-02-22PHML+ DATA TABLES.XLSX"), sheet = "Batch-normalized Data") %>% data.table()

## add X to met ID
mets_info$Name <- paste("X", mets_info$CHEM_ID, sep = "")

colnames(mets_raw_df)[2:1729] <- paste("X", colnames(mets_raw_df)[2:1729], sep = "")

## add biobank subject ID
mets_raw_df <- merge(sample_df[,c("PARENT_SAMPLE_NAME", "Biobank_Subject_ID")], mets_raw_df, by = "PARENT_SAMPLE_NAME", all = T)

## QC mets list
mets_list <- mets_info[miss_flag == 0,]$Name #1459

```

## 1.4 Qingwen's RPDR data
```{r}
pheno_odollfa_rpdr <- fread(here(dat_dir, "/ODOLLFA_pheno_mets.csv"))
pheno_phasic_rpdr <- fread(here(dat_dir, "/PHASIC_pheno_mets.csv"))

pheno_rpdr <- rbind(pheno_odollfa_rpdr, pheno_phasic_rpdr) %>% as.data.table()

pheno_rpdr$`FEV1/FVC_median_6M` <- (pheno_rpdr$FEV1_median_6M)/(pheno_rpdr$FVC_median_6M)

pheno_rpdr$`FEV1/FVC_median_12M` <- (pheno_rpdr$FEV1_median_12M)/(pheno_rpdr$FVC_median_12M)

pheno_rpdr$`FEV1/FVC_median_24M` <- (pheno_rpdr$FEV1_median_24M)/(pheno_rpdr$FVC_median_24M)

which(colnames(pheno_rpdr) == "EMPI")
pheno_rpdr <- pheno_rpdr[,-c(2:1460)]

which(colnames(pheno_rpdr) == "Subject_Id")
colnames(pheno_rpdr)[7] <- "Biobank_Subject_ID"

```

## 1.5 Merge and trim
```{r}
# check the replicated samples
n_occur <- data.frame(table(sample_df$Biobank_Subject_ID))
table(n_occur$Freq)
# n_occur[n_occur$Freq > 1,]
dim(n_occur[n_occur$Freq > 1,])
sample_df[sample_df$Biobank_Subject_ID %in% n_occur$Var1[n_occur$Freq > 1],]

# remove the replicated sample
unique_subject_list <- n_occur[n_occur$Freq < 2, ]$Var1 %>% as.character()

mets_df_unique <- mets_raw_df[Biobank_Subject_ID %in% unique_subject_list,]

# merge to phenotype data
pool_df <- merge(pool, mets_df_unique, by = c("PARENT_SAMPLE_NAME", "Biobank_Subject_ID"), all.y = T)

# # merge to Qingwen's data
col.name_1 <- colnames(pool_df)
col.name_2 <- colnames(pheno_rpdr)
col.name_3 <- col.name_2[!(col.name_2 %in% col.name_1)]
col.name_3 <- append("Biobank_Subject_ID", col.name_3)

# pool_df <- merge(pool_df, pheno_rpdr, by = "Biobank_Subject_ID", all.x = T)

pool_df_new <- merge(pool_df, pheno_rpdr[,..col.name_3], by = "Biobank_Subject_ID", all.x = T)

# check the replicated samples
n_occur <- data.frame(table(pool_df_new$Biobank_Subject_ID))
table(n_occur$Freq)
# n_occur[n_occur$Freq > 1,]
dim(n_occur[n_occur$Freq > 1,])

# remove the replicated sample
# unique_subject_list <- n_occur[n_occur$Freq < 2, ]$Var1 %>% as.character()
# pool_df_new <- pool_df_new[Biobank_Subject_ID %in% unique_subject_list,]

# remove COPD and smokers
pool_df_new[, .N, .(copd_existence)][order(-N)][, pct := N/sum(N)*100] %>% print(digits = 3)
pool_df_new[, .N, .(Smoking_flag_Biobank)][order(-N)][, pct := N/sum(N)*100] %>% print(digits = 3)
pool_df_new[, .N, .(smk_ever_alltime)][order(-N)][, pct := N/sum(N)*100] %>% print(digits = 3)
pool_df_new[, .N, .(Smoking_RPDR_most_recent)][order(-N)][, pct := N/sum(N)*100] %>% print(digits = 3)
pool_df_new[, .N, .(Smoking_RPDR_most_recent, Smoking_flag_Biobank)][order(-N)][, pct := N/sum(N)*100] %>% print(digits = 3)


pool_df_new <- pool_df_new[pool_df_new$copd_existence != "Yes" & pool_df_new$Smoking_RPDR_most_recent != "Former-smoker" & pool_df_new$Smoking_RPDR_most_recent != "Current-smoker",]

nrow(pool_df_new)

pool_df_new$race <- with(pool_df_new, ifelse(race == "White", "White", 
                                      ifelse(race == "Black", "Black", "Other")))
pool_df_new$Ethnicity <- with(pool_df_new, ifelse(Ethnicity == "HISPANIC", "Hispanic", "Non Hispanic"))

pool_df_new[, .N, .(race)][order(-N)][, pct := N/sum(N)*100] %>% print(digits = 3)
pool_df_new[, .N, .(Ethnicity)][order(-N)][, pct := N/sum(N)*100] %>% print(digits = 3)

```

## 1.6 Table 1
```{r}
table1(~ Age_at_collect + BMI_most_recent + Gender + race + Ethnicity + Smoking_RPDR_most_recent | asthma_existence, data = pool_df_new, overall = F)

table1(~ EOScount_median_6M + EOScount_median_12M + EOScount_median_24M + 
         EOSpct_median_6M + EOSpct_median_12M + EOSpct_median_24M + 
         NEUcount_median_6M + NEUcount_median_12M + NEUcount_median_24M + 
         NEUpct_median_6M + NEUpct_median_12M + NEUpct_median_24M + 
         log10(ige_closest_lastpl) | asthma_existence, data = pool_df_new, overall = F)

log10_ige <- log10(pool_df_new$ige_closest_lastpl)
log10_ige <- log10_ige[log10_ige != -Inf]
mean(log10_ige, na.rm = T)
sd(log10_ige, na.rm = T)

table1(~ ics_trim_totnum_5y + ics_trim_totnum_5y_after + 
         ocs_trim_totnum_5y + ocs_trim_totnum_5y_after| asthma_existence, data = pool_df_new, overall = F)

table1( ~ FEV1_median_24M + FVC_median_24M + `FEV1/FVC_median_24M` | asthma_existence, data = pool_df_new, overall = F)
```

# 2. Create targeted ratio
## 2.1 summary the subpathay
```{r}
mets_info_updated <- mets_info[Name %in% mets_list,]

mets_info_subpathway <- mets_info_updated[, .N, .(SUB_PATHWAY)][order(-N)][, pct := N/sum(N)*100] %>% print(digits = 2, topn = 150)

```

## 2.2 sphingolipid list
```{r}
# sphingolipid
## Sphingomyelins 29
## Ceramides 10
## Hexosylceramides (HCER) 6
## Dihydrosphingomyelins 5
## Sphingosines 3
## Sphingolipid Synthesis 3
## Lactosylceramides (LCER) 3
## Dihydroceramides 2
## Ceramide PEs 1

sph_sub_pathway_list <- c("Sphingomyelins", "Ceramides", "Ceramide PEs", "Hexosylceramides (HCER)", "Dihydrosphingomyelins", "Sphingosines", "Sphingolipid Synthesis", "Lactosylceramides (LCER)", "Dihydroceramides")

sph_mets_list <- mets_info_updated[SUB_PATHWAY %in% sph_sub_pathway_list,]$Name
length(sph_mets_list)

sph_mets_info <- mets_info_updated[SUB_PATHWAY %in% sph_sub_pathway_list,]
setcolorder(sph_mets_info, c("SUPER_PATHWAY", "SUB_PATHWAY", "CHEMICAL_NAME"))

```

## 2.3 steroid list
```{r}
## Androgenic Steroids 24
## Corticosteroids 7
## Progestin Steroids 6
## Estrogenic Steroids 1
## Sterol 8

str_sub_pathway_list <- c("Androgenic Steroids", "Corticosteroids", "Progestin Steroids", "Estrogenic Steroids", "Sterol")

str_mets_list <- mets_info_updated[SUB_PATHWAY %in% str_sub_pathway_list,]$Name
length(str_mets_list)

str_mets_info <- mets_info_updated[SUB_PATHWAY %in% str_sub_pathway_list,]
setcolorder(str_mets_info, c("SUPER_PATHWAY", "SUB_PATHWAY", "CHEMICAL_NAME"))

```

## 2.4 microbial list
```{r}
## sub pathway
## Phenylalanine Metabolism 8
## Tyrosine Metabolism 24
## Tryptophan Metabolism 26
## Short Chain Fatty Acid 2
## Polyamine Metabolism 7
## Secondary Bile Acid Metabolism 29
## Carnitine Metabolism 2
## Histidine Metabolism 18


## chemical name
## betaine X799
## choline X1256

microb_sub_pathway_list <- c("Phenylalanine Metabolism", "Tyrosine Metabolism", "Tryptophan Metabolism", "Short Chain Fatty Acid", "Polyamine Metabolism", "Secondary Bile Acid Metabolism", "Carnitine Metabolism", "Histidine Metabolism")

microb_mets_list <- mets_info_updated[SUB_PATHWAY %in% microb_sub_pathway_list,]$Name
microb_mets_list <- c(microb_mets_list, "X799", "X1256")
length(microb_mets_list)

microb_mets_info <- mets_info_updated[SUB_PATHWAY %in% microb_sub_pathway_list,]
setcolorder(microb_mets_info, c("SUPER_PATHWAY", "SUB_PATHWAY", "CHEMICAL_NAME"))

```

## 2.5 sph/str
```{r}
df <- mets_df_unique %>% as.data.frame()
df <- df[,-c(1)]
rownames(df) <- df$Biobank_Subject_ID
df <- df[,-1]

df_ratios <- df

for (i in 1:length(sph_mets_list)) {
  for (j in 1:length(str_mets_list)) {
    sph <- sph_mets_list[i]
    steroid <- str_mets_list[j]
    eval(parse(text = str_c("df_ratios$", sph, ".", steroid ," <- df_ratios$", sph, "/", "df_ratios$", steroid)))
  }
}

df_ratios <- df_ratios[,-which(colnames(df_ratios) %in% colnames(df))]
ncol(df_ratios) #2852

df_sph_steroid_ratio <- df_ratios
df_sph_steroid_ratio$Biobank_Subject_ID <- rownames(df_sph_steroid_ratio)
df_sph_steroid_ratio %>% setcolorder(., "Biobank_Subject_ID")
df_sph_steroid_ratio <- as.data.table(df_sph_steroid_ratio)
df_sph_steroid_ratio_new <- df_sph_steroid_ratio[Biobank_Subject_ID %in% pool_df_new$Biobank_Subject_ID,]

# missing
mets_missing_sph_str <- df_sph_steroid_ratio_new[, lapply(.SD, function(x) mean(is.na(x))), .SDcols = colnames(df_sph_steroid_ratio_new)[2:ncol(df_sph_steroid_ratio_new)]] %>% 
        t() %>% as.data.frame() %>% rownames_to_column()
setnames(mets_missing_sph_str, c("Name", "pct_na_met"))
summary(mets_missing_sph_str$pct_na_met)

# histogram
ggplot(df_sph_steroid_ratio_new, aes(x=X100000626.X100000792)) + geom_histogram()
ggplot(df_sph_steroid_ratio_new, aes(x=X100000626.X100001987)) + geom_histogram()


# log transformed
df_sph_steroid_ratio_log <- df_sph_steroid_ratio_new
df_sph_steroid_ratio_log[,c(2:2853)] <- log2(df_sph_steroid_ratio_log[,c(2:2853)])

df_sph_steroid_ratio_log <- df_sph_steroid_ratio_log %>% mutate_if(is.numeric, ~ replace(., !is.finite(.), NA))

ggplot(df_sph_steroid_ratio_log, aes(x=X100000626.X100000792)) + geom_histogram()
ggplot(df_sph_steroid_ratio_log, aes(x=X100000626.X100001987)) + geom_histogram()

```

## 2.6 sph/microb
```{r}
df <- mets_df_unique %>% as.data.frame()
df <- df[,-c(1)]
rownames(df) <- df$Biobank_Subject_ID
df <- df[,-1]

df_ratios <- df

for (i in 1:length(sph_mets_list)) {
  for (j in 1:length(microb_mets_list)) {
    sph <- sph_mets_list[i]
    microb <- microb_mets_list[j]
    eval(parse(text = str_c("df_ratios$", sph, ".", microb ," <- df_ratios$", sph, "/", "df_ratios$", microb)))
  }
}

df_ratios <- df_ratios[,-which(colnames(df_ratios) %in% colnames(df))]
ncol(df_ratios) #7316

df_sph_microb_ratio <- df_ratios
df_sph_microb_ratio$Biobank_Subject_ID <- rownames(df_sph_microb_ratio)
df_sph_microb_ratio %>% setcolorder(., "Biobank_Subject_ID")
df_sph_microb_ratio <- as.data.table(df_sph_microb_ratio)
df_sph_microb_ratio_new <- df_sph_microb_ratio[Biobank_Subject_ID %in% pool_df_new$Biobank_Subject_ID,]

# missing
mets_missing_sph_microb <- df_sph_microb_ratio_new[, lapply(.SD, function(x) mean(is.na(x))), .SDcols = colnames(df_sph_microb_ratio_new)[2:ncol(df_sph_microb_ratio_new)]] %>% 
        t() %>% as.data.frame() %>% rownames_to_column()
setnames(mets_missing_sph_str, c("Name", "pct_na_met"))
summary(mets_missing_sph_str$pct_na_met)

# histogram
ggplot(df_sph_microb_ratio_new, aes(x=X100000626.X100000007)) + geom_histogram()
ggplot(df_sph_microb_ratio_new, aes(x=X100000626.X100000463)) + geom_histogram()


# log transformed
df_sph_microb_ratio_log <- df_sph_microb_ratio_new
df_sph_microb_ratio_log[,c(2:7317)] <- log2(df_sph_microb_ratio_log[,c(2:7317)])

df_sph_microb_ratio_log <- df_sph_microb_ratio_log %>% mutate_if(is.numeric, ~ replace(., !is.finite(.), NA))

ggplot(df_sph_microb_ratio_log, aes(x=X100000626.X100000007)) + geom_histogram()
ggplot(df_sph_microb_ratio_log, aes(x=X100000626.X100000463)) + geom_histogram()


```


## 2.7 microb/str
```{r}
df <- mets_df_unique %>% as.data.frame()
df <- df[,-c(1)]
rownames(df) <- df$Biobank_Subject_ID
df <- df[,-1]

df_ratios <- df

for (i in 1:length(microb_mets_list)) {
  for (j in 1:length(str_mets_list)) {
    microb <- microb_mets_list[i]
    str <- str_mets_list[j]
    eval(parse(text = str_c("df_ratios$", microb, ".", str ," <- df_ratios$", microb, "/", "df_ratios$", str)))
  }
}

df_ratios <- df_ratios[,-which(colnames(df_ratios) %in% colnames(df))]
ncol(df_ratios) #5428

df_microb_str_ratio <- df_ratios
df_microb_str_ratio$Biobank_Subject_ID <- rownames(df_microb_str_ratio)
df_microb_str_ratio %>% setcolorder(., "Biobank_Subject_ID")
df_microb_str_ratio <- as.data.table(df_microb_str_ratio)
df_microb_str_ratio_new <- df_microb_str_ratio[Biobank_Subject_ID %in% pool_df_new$Biobank_Subject_ID,]

# missing
mets_missing_microb_str <- df_microb_str_ratio_new[, lapply(.SD, function(x) mean(is.na(x))), .SDcols = colnames(df_microb_str_ratio_new)[2:ncol(df_microb_str_ratio_new)]] %>% 
        t() %>% as.data.frame() %>% rownames_to_column()
setnames(mets_missing_sph_str, c("Name", "pct_na_met"))
summary(mets_missing_sph_str$pct_na_met)

# histogram
ggplot(df_microb_str_ratio_new, aes(x=X100000007.X100000792)) + geom_histogram()
ggplot(df_microb_str_ratio_new, aes(x=X100000007.X100002014)) + geom_histogram()


# log transformed
df_microb_str_ratio_log <- df_microb_str_ratio_new
df_microb_str_ratio_log[,c(2:5429)] <- log2(df_microb_str_ratio_log[,c(2:5429)])

df_microb_str_ratio_log <- df_microb_str_ratio_log %>% mutate_if(is.numeric, ~ replace(., !is.finite(.), NA))

ggplot(df_microb_str_ratio_log, aes(x=X100000007.X100000792)) + geom_histogram()
ggplot(df_microb_str_ratio_log, aes(x=X100000007.X100002014)) + geom_histogram()


```


# 3. Regression model functions
## 3.1 linear regression
```{r}
linreg_res_fnc <- function(outc, mets_list, dat, add_covar = "", 
                           first_cols = c("Name", "outc", "beta", "pval", "fdr", "fdr_bh"), 
                           na_thld = missing_thld) {
        
        linear_mdl_tmp <- function(met) {
                
                fit <- eval(parse(text = str_c("lm(", outc, " ~ ", met, 
                                               add_covar, ", data = ", dat, ")")))
                
                coef <- summary(fit)$coefficients %>% as.data.frame() %>% rownames_to_column() %>% as.data.table()
                coef <- coef[rowname == met, ]

                output <- cbind(outc, coef)
                colnames(output) <- c("outc", "Name", "beta", "se", "tval", "pval")
                output[, (colnames(output)[-c(1:2)]) := lapply(.SD, as.numeric), .SDcols = colnames(output)[-c(1:2)]]
                
                return(output)
        }
        
        res <- lapply(mets_list, function(x) {linear_mdl_tmp(met = x)}) %>% rbindlist()
        
        # eval(parse(text = str_c("mets_info <- copy(", mets_info, ")")))
        # res <- merge(res, mets_info, by = "feat_id")
        res[, ':='(fdr = p.adjust(pval, method = "fdr"), 
                   fdr_bh = p.adjust(pval, method = "BH"))]
        res[, totaln := eval(parse(text = str_c(dat, "[!is.na(", outc, "), .N]")))]

        setcolorder(res, first_cols)
        res <- res[order(pval)]
        return(res)
}


```

## 3.2 logistic regression
```{r}
logi_res_fnc <- function(outc, mets_list, dat, add_covar = "",
                         first_cols = c("Name", 
                                        "outc", "beta", "pval", "fdr", "or", "lower95", "upper95")) {
        
        logi_mdl_tmp <- function(met) {
                
                fit <- eval(parse(text = str_c("glm(", outc, " ~ ", met, add_covar, ", data = ", dat, ", family = 'binomial')")))
                
                coef <- summary(fit)$coefficients %>% as.data.frame() %>% rownames_to_column() %>% as.data.table()
                coef <- coef[rowname == met, ]

                output <- cbind(outc, coef)
                colnames(output) <- c("outc", "Name", "beta", "se", "zval", "pval")
                output[, (colnames(output)[-c(1:2)]) := lapply(.SD, as.numeric), .SDcols = colnames(output)[-c(1:2)]]
                
                return(output)
        }
        
        res <- lapply(mets_list, function(x) {logi_mdl_tmp(met = x)}) %>% rbindlist()
        
        # eval(parse(text = str_c("mets_info <- copy(", mets_info, ")")))
        # res <- merge(res, mets_info, by = "feat_id", all.x = T)
        res[, ':='(neg_log10_pval = -log10(pval), 
                   fdr = p.adjust(pval, method = "fdr"), 
                   or = exp(beta), 
                   lower95 = exp(beta-z_95*se), 
                   upper95 = exp(beta+z_95*se))]
        res[, totaln := eval(parse(text = str_c(dat, "[!is.na(", outc, "), .N]")))]
        
        setcolorder(res, first_cols)
        res <- res[order(pval)]
        return(res)
}
```

## 3.3 poisson regression
```{r}
poiss_res_fnc <- function(outc, mets_list, dat, add_covar = "",
                         first_cols = c("Name", 
                                        "outc", "beta", "pval", "fdr")) {
        
        logi_mdl_tmp <- function(met) {
                
                fit <- eval(parse(text = str_c("glm(", outc, " ~ ", met, add_covar, ", data = ", dat, ", family = 'quasipoisson')")))
                
                coef <- summary(fit)$coefficients %>% as.data.frame() %>% rownames_to_column() %>% as.data.table()
                coef <- coef[rowname == met, ]

                output <- cbind(outc, coef)
                colnames(output) <- c("outc", "Name", "beta", "se", "zval", "pval")
                output[, (colnames(output)[-c(1:2)]) := lapply(.SD, as.numeric), .SDcols = colnames(output)[-c(1:2)]]
                
                return(output)
        }
        
        res <- lapply(mets_list, function(x) {logi_mdl_tmp(met = x)}) %>% rbindlist()
        
        # eval(parse(text = str_c("mets_info <- copy(", mets_info, ")")))
        # res <- merge(res, mets_info, by = "feat_id", all.x = T)
        res[, ':='(neg_log10_pval = -log10(pval), 
                   fdr = p.adjust(pval, method = "fdr"))]
        res[, totaln := eval(parse(text = str_c(dat, "[!is.na(", outc, "), .N]")))]
        
        setcolorder(res, first_cols)
        res <- res[order(pval)]
        return(res)
}

```

## 3.4 Manhattan Plot
```{r}
save_manh_plot <- function(dat, title_infig, fig_name, wdth = 20, ht = 12, pointsize = 3, pointstroke = 1.5,
                           transparency = 1, if_lbl = c(T, F), lbl_thld = 2e-04) {
  
        p <- ggplot(dat, aes(x = Name, y = -log10(pval))) + 
                    facet_grid(. ~ denominator_sub_pathway, scales = "free_x") +  
                    geom_point(aes(color = numerator_sub_pathway , shape = direction_of_effect), 
                               size = pointsize, stroke = pointstroke, alpha = transparency) + 
                    scale_shape_manual(values = c("-" = 6, "+" = 2)) + 
                    labs(title = title_infig, 
                         y = expression('-log'[10]*'('*italic(P)*'-value)')) + 
                    theme(legend.position = "bottom", 
                          legend.text = element_text(size = 16), 
                          plot.title = element_text(hjust = 0.5, size = 20), 
                          strip.text.x = element_text(size = 16), 
                          axis.title.x = element_blank(),
                          axis.text.x = element_blank(), 
                          axis.ticks.x = element_blank(), 
                          axis.title.y = element_text(size = 20, face = "bold"), 
                          axis.text.y = element_text(size = 20, face = "bold"),
                          panel.border = element_blank(),
                          panel.grid.major = element_blank(),
                          panel.grid.minor = element_blank())
        if (if_lbl == T) {
                p <- p + geom_label_repel(aes(label = ifelse(pval < lbl_thld, CHEMICAL_NAME, "")),
                                          box.padding   = 0.25,
                                          point.padding = 0.5,
                                          segment.color = 'grey50',
                                          max.overlaps = 80)
        }
        ggsave(file = here::here(fig_dir, fig_name), plot = p, width = wdth, height = ht)
}

```

# 4. Conduct association analysis
## 4.1 Log(IgE)
```{r}
dat_tmp <- pool_df_new
dat_tmp$log_ige <- log(dat_tmp$ige_closest_lastpl)
dat_tmp <- dat_tmp %>% mutate_if(is.numeric, ~ replace(., !is.finite(.), NA))
```

### 4.1.1 targeted sph/microb
```{r}
list_analysis <- mets_missing_sph_microb[mets_missing_sph_microb$V1 <= 0.3,"rowname"]
length(list_analysis) #5743

df_sph_microb_ratio_log$Biobank_Subject_ID <- as.integer(df_sph_microb_ratio_log$Biobank_Subject_ID)
pheno_tmp <- merge(dat_tmp, df_sph_microb_ratio_log, by = "Biobank_Subject_ID", all.y = T)

res_4.1.1 <- linreg_res_fnc(outc = "log_ige", 
                            mets_list = list_analysis, 
                            dat = "pheno_tmp", 
                            add_covar = " + Age_at_collect + BMI_most_recent + Gender + race + ics_trim_totnum_5y")

## Save results
res_log_ige_sph_microb_ratio <- res_4.1.1
res_log_ige_sph_microb_ratio <- res_log_ige_sph_microb_ratio %>% separate(Name, sep = "\\.", c("numerator", "denominator"), remove = F)

dict <- mets_info_updated[,c("Name", "CHEMICAL_NAME", "SUB_PATHWAY", "SUPER_PATHWAY")]
res_log_ige_sph_microb_ratio$numerator <- dict$CHEMICAL_NAME[match(res_log_ige_sph_microb_ratio$numerator, dict$Name)]
res_log_ige_sph_microb_ratio$numerator_sub_pathway <- dict$SUB_PATHWAY[match(res_log_ige_sph_microb_ratio$numerator, dict$CHEMICAL_NAME)]

res_log_ige_sph_microb_ratio$denominator <- dict$CHEMICAL_NAME[match(res_log_ige_sph_microb_ratio$denominator, dict$Name)]
res_log_ige_sph_microb_ratio$denominator_sub_pathway <- dict$SUB_PATHWAY[match(res_log_ige_sph_microb_ratio$denominator, dict$CHEMICAL_NAME)]

setcolorder(res_log_ige_sph_microb_ratio, c("Name", "numerator", "numerator_sub_pathway", "denominator", "denominator_sub_pathway"))


saveRDS(res_log_ige_sph_microb_ratio, file = here(res_dir, "res_log_ige_sph_microb_ratio.rds"))

## display results
### min p-value = #
### #sig = 0/5743
datatable(res_log_ige_sph_microb_ratio[res_log_ige_sph_microb_ratio$fdr <= 0.05,], 
          filter = "top") %>% 
          formatSignif(columns = c("beta", "pval", "fdr", "fdr", "se", "tval"), digits = 3)

res_4.1.1_sig <- res_log_ige_sph_microb_ratio[res_log_ige_sph_microb_ratio$fdr <= 0.05,]

```

### 4.1.2 targeted sph/str
```{r}
list_analysis <- mets_missing_sph_str[mets_missing_sph_str$pct_na_met <= 0.3,"Name"]
length(list_analysis) #1827

df_sph_steroid_ratio_log$Biobank_Subject_ID <- as.integer(df_sph_steroid_ratio_log$Biobank_Subject_ID)
pheno_tmp <- merge(dat_tmp, df_sph_steroid_ratio_log, by = "Biobank_Subject_ID", all.y = T)

res_4.1.2 <- linreg_res_fnc(outc = "log_ige", 
                            mets_list = list_analysis, 
                            dat = "pheno_tmp", 
                            add_covar = " + Age_at_collect + BMI_most_recent + Gender + race + ics_trim_totnum_5y")

## Save results
res_log_ige_sph_str_ratio <- res_4.1.2
res_log_ige_sph_str_ratio <- res_log_ige_sph_str_ratio %>% separate(Name, sep = "\\.", c("numerator", "denominator"), remove = F)

dict <- mets_info_updated[,c("Name", "CHEMICAL_NAME", "SUB_PATHWAY", "SUPER_PATHWAY")]
res_log_ige_sph_str_ratio$numerator <- dict$CHEMICAL_NAME[match(res_log_ige_sph_str_ratio$numerator, dict$Name)]
res_log_ige_sph_str_ratio$numerator_sub_pathway <- dict$SUB_PATHWAY[match(res_log_ige_sph_str_ratio$numerator, dict$CHEMICAL_NAME)]

res_log_ige_sph_str_ratio$denominator <- dict$CHEMICAL_NAME[match(res_log_ige_sph_str_ratio$denominator, dict$Name)]
res_log_ige_sph_str_ratio$denominator_sub_pathway <- dict$SUB_PATHWAY[match(res_log_ige_sph_str_ratio$denominator, dict$CHEMICAL_NAME)]

setcolorder(res_log_ige_sph_str_ratio, c("Name", "numerator", "numerator_sub_pathway", "denominator", "denominator_sub_pathway"))


saveRDS(res_log_ige_sph_str_ratio, file = here(res_dir, "res_log_ige_sph_str_ratio.rds"))

## display results
### min p-value = #
### #sig = 0/1827
datatable(res_log_ige_sph_str_ratio[res_log_ige_sph_str_ratio$fdr <= 0.05,], 
          filter = "top") %>% 
          formatSignif(columns = c("beta", "pval", "fdr", "fdr", "se", "tval"), digits = 3)

res_4.1.2_sig <- res_log_ige_sph_str_ratio[res_log_ige_sph_str_ratio$fdr <= 0.05,]

```


### 4.1.3 targeted microb/str
```{r}
list_analysis <- mets_missing_microb_str[mets_missing_microb_str$V1 <= 0.3,"rowname"]
length(list_analysis) #3085

df_microb_str_ratio_log$Biobank_Subject_ID <- as.integer(df_microb_str_ratio_log$Biobank_Subject_ID)
pheno_tmp <- merge(dat_tmp, df_microb_str_ratio_log, by = "Biobank_Subject_ID", all.y = T)

res_4.1.3 <- linreg_res_fnc(outc = "log_ige", 
                            mets_list = list_analysis, 
                            dat = "pheno_tmp", 
                            add_covar = " + Age_at_collect + BMI_most_recent + Gender + race +  ics_trim_totnum_5y")

## Save results
res_log_ige_microb_str_ratio <- res_4.1.3
res_log_ige_microb_str_ratio <- res_log_ige_microb_str_ratio %>% separate(Name, sep = "\\.", c("numerator", "denominator"), remove = F)

dict <- mets_info_updated[,c("Name", "CHEMICAL_NAME", "SUB_PATHWAY", "SUPER_PATHWAY")]
res_log_ige_microb_str_ratio$numerator <- dict$CHEMICAL_NAME[match(res_log_ige_microb_str_ratio$numerator, dict$Name)]
res_log_ige_microb_str_ratio$numerator_sub_pathway <- dict$SUB_PATHWAY[match(res_log_ige_microb_str_ratio$numerator, dict$CHEMICAL_NAME)]

res_log_ige_microb_str_ratio$denominator <- dict$CHEMICAL_NAME[match(res_log_ige_microb_str_ratio$denominator, dict$Name)]
res_log_ige_microb_str_ratio$denominator_sub_pathway <- dict$SUB_PATHWAY[match(res_log_ige_microb_str_ratio$denominator, dict$CHEMICAL_NAME)]

setcolorder(res_log_ige_microb_str_ratio, c("Name", "numerator", "numerator_sub_pathway", "denominator", "denominator_sub_pathway"))

saveRDS(res_log_ige_microb_str_ratio, file = here(res_dir, "res_log_ige_microb_str_ratio.rds"))

## display results
### min p-value = #
### #sig = 0/3085
datatable(res_log_ige_microb_str_ratio[res_log_ige_microb_str_ratio$fdr <= 0.05,], 
          filter = "top") %>% 
          formatSignif(columns = c("beta", "pval", "fdr", "fdr", "se", "tval"), digits = 3)

res_4.1.3_sig <- res_log_ige_microb_str_ratio[res_log_ige_microb_str_ratio$fdr <= 0.05,]

```


## 4.2 FEV1
```{r}
ggplot(dat_tmp, aes(x=FEV1_most_recent)) + geom_histogram(bins = 50)
```

### 4.2.1 targeted sph/microb
```{r}
list_analysis <- mets_missing_sph_microb[mets_missing_sph_microb$V1 <= 0.3,"rowname"]
length(list_analysis) #5743

df_sph_microb_ratio_log$Biobank_Subject_ID <- as.integer(df_sph_microb_ratio_log$Biobank_Subject_ID)
pheno_tmp <- merge(dat_tmp, df_sph_microb_ratio_log, by = "Biobank_Subject_ID", all.y = T)

res_4.2.1 <- linreg_res_fnc(outc = "FEV1_most_recent", 
                            mets_list = list_analysis, 
                            dat = "pheno_tmp", 
                            add_covar = " + Age_at_collect + BMI_most_recent + Gender + race + Height +  ics_trim_totnum_5y")

## Save results
res_FEV1_sph_microb_ratio <- res_4.2.1
res_FEV1_sph_microb_ratio <- res_FEV1_sph_microb_ratio %>% separate(Name, sep = "\\.", c("numerator", "denominator"), remove = F)

dict <- mets_info_updated[,c("Name", "CHEMICAL_NAME", "SUB_PATHWAY", "SUPER_PATHWAY")]
res_FEV1_sph_microb_ratio$numerator <- dict$CHEMICAL_NAME[match(res_FEV1_sph_microb_ratio$numerator, dict$Name)]
res_FEV1_sph_microb_ratio$numerator_sub_pathway <- dict$SUB_PATHWAY[match(res_FEV1_sph_microb_ratio$numerator, dict$CHEMICAL_NAME)]

res_FEV1_sph_microb_ratio$denominator <- dict$CHEMICAL_NAME[match(res_FEV1_sph_microb_ratio$denominator, dict$Name)]
res_FEV1_sph_microb_ratio$denominator_sub_pathway <- dict$SUB_PATHWAY[match(res_FEV1_sph_microb_ratio$denominator, dict$CHEMICAL_NAME)]

setcolorder(res_FEV1_sph_microb_ratio, c("Name", "numerator", "numerator_sub_pathway", "denominator", "denominator_sub_pathway"))


saveRDS(res_FEV1_sph_microb_ratio, file = here(res_dir, "res_FEV1_sph_microb_ratio.rds"))

## display results
### min p-value = 4.18e-10
### #sig = 242/5743
datatable(res_FEV1_sph_microb_ratio[res_FEV1_sph_microb_ratio$fdr <= 0.05,], 
          filter = "top") %>% 
          formatSignif(columns = c("beta", "pval", "fdr", "fdr", "se", "tval"), digits = 3)

res_4.2.1_sig <- res_FEV1_sph_microb_ratio[res_FEV1_sph_microb_ratio$fdr <= 0.05,]

```

### 4.2.2 targeted sph/str
```{r}
list_analysis <- mets_missing_sph_str[mets_missing_sph_str$pct_na_met <= 0.3,"Name"]
length(list_analysis) #1827

df_sph_steroid_ratio_log$Biobank_Subject_ID <- as.integer(df_sph_steroid_ratio_log$Biobank_Subject_ID)
pheno_tmp <- merge(dat_tmp, df_sph_steroid_ratio_log, by = "Biobank_Subject_ID", all.y = T)

res_4.2.2 <- linreg_res_fnc(outc = "FEV1_most_recent", 
                            mets_list = list_analysis, 
                            dat = "pheno_tmp", 
                            add_covar = " + Age_at_collect + BMI_most_recent + Gender + race + Height +  ics_trim_totnum_5y")

## Save results
res_FEV1_sph_str_ratio <- res_4.2.2
res_FEV1_sph_str_ratio <- res_FEV1_sph_str_ratio %>% separate(Name, sep = "\\.", c("numerator", "denominator"), remove = F)

dict <- mets_info_updated[,c("Name", "CHEMICAL_NAME", "SUB_PATHWAY", "SUPER_PATHWAY")]
res_FEV1_sph_str_ratio$numerator <- dict$CHEMICAL_NAME[match(res_FEV1_sph_str_ratio$numerator, dict$Name)]
res_FEV1_sph_str_ratio$numerator_sub_pathway <- dict$SUB_PATHWAY[match(res_FEV1_sph_str_ratio$numerator, dict$CHEMICAL_NAME)]

res_FEV1_sph_str_ratio$denominator <- dict$CHEMICAL_NAME[match(res_FEV1_sph_str_ratio$denominator, dict$Name)]
res_FEV1_sph_str_ratio$denominator_sub_pathway <- dict$SUB_PATHWAY[match(res_FEV1_sph_str_ratio$denominator, dict$CHEMICAL_NAME)]

setcolorder(res_FEV1_sph_str_ratio, c("Name", "numerator", "numerator_sub_pathway", "denominator", "denominator_sub_pathway"))


saveRDS(res_FEV1_sph_str_ratio, file = here(res_dir, "res_FEV1_sph_str_ratio.rds"))

## display results
### min p-value = 0.0000152
### #sig = 25/1827
datatable(res_FEV1_sph_str_ratio[res_FEV1_sph_str_ratio$fdr <= 0.05,], 
          filter = "top") %>% 
          formatSignif(columns = c("beta", "pval", "fdr", "fdr", "se", "tval"), digits = 3)

res_4.2.2_sig <- res_FEV1_sph_str_ratio[res_FEV1_sph_str_ratio$fdr <= 0.05,]

```


### 4.2.3 targeted microb/str
```{r}
list_analysis <- mets_missing_microb_str[mets_missing_microb_str$V1 <= 0.3,"rowname"]
length(list_analysis) #3085

df_microb_str_ratio_log$Biobank_Subject_ID <- as.integer(df_microb_str_ratio_log$Biobank_Subject_ID)
pheno_tmp <- merge(dat_tmp, df_microb_str_ratio_log, by = "Biobank_Subject_ID", all.y = T)

res_4.2.3 <- linreg_res_fnc(outc = "FEV1_most_recent", 
                            mets_list = list_analysis, 
                            dat = "pheno_tmp", 
                            add_covar = " + Age_at_collect + BMI_most_recent + Gender + race + Height +  ics_trim_totnum_5y")

## Save results
res_FEV1_microb_str_ratio <- res_4.2.3
res_FEV1_microb_str_ratio <- res_FEV1_microb_str_ratio %>% separate(Name, sep = "\\.", c("numerator", "denominator"), remove = F)

dict <- mets_info_updated[,c("Name", "CHEMICAL_NAME", "SUB_PATHWAY", "SUPER_PATHWAY")]
res_FEV1_microb_str_ratio$numerator <- dict$CHEMICAL_NAME[match(res_FEV1_microb_str_ratio$numerator, dict$Name)]
res_FEV1_microb_str_ratio$numerator_sub_pathway <- dict$SUB_PATHWAY[match(res_FEV1_microb_str_ratio$numerator, dict$CHEMICAL_NAME)]

res_FEV1_microb_str_ratio$denominator <- dict$CHEMICAL_NAME[match(res_FEV1_microb_str_ratio$denominator, dict$Name)]
res_FEV1_microb_str_ratio$denominator_sub_pathway <- dict$SUB_PATHWAY[match(res_FEV1_microb_str_ratio$denominator, dict$CHEMICAL_NAME)]

setcolorder(res_FEV1_microb_str_ratio, c("Name", "numerator", "numerator_sub_pathway", "denominator", "denominator_sub_pathway"))

saveRDS(res_FEV1_microb_str_ratio, file = here(res_dir, "res_FEV1_microb_str_ratio.rds"))

## display results
### min p-value = 1.11e-9	
### #sig = 150/3085
datatable(res_FEV1_microb_str_ratio[res_FEV1_microb_str_ratio$fdr <= 0.05,], 
          filter = "top") %>% 
          formatSignif(columns = c("beta", "pval", "fdr", "fdr", "se", "tval"), digits = 3)

res_4.2.3_sig <- res_FEV1_microb_str_ratio[res_FEV1_microb_str_ratio$fdr <= 0.05,]

```

## 4.3 FVC
```{r}
ggplot(dat_tmp, aes(x=FVC_most_recent)) + geom_histogram(bins = 50)
```

### 4.3.1 targeted sph/microb
```{r}
list_analysis <- mets_missing_sph_microb[mets_missing_sph_microb$V1 <= 0.3,"rowname"]
length(list_analysis) #5743

df_sph_microb_ratio_log$Biobank_Subject_ID <- as.integer(df_sph_microb_ratio_log$Biobank_Subject_ID)
pheno_tmp <- merge(dat_tmp, df_sph_microb_ratio_log, by = "Biobank_Subject_ID", all.y = T)

res_4.3.1 <- linreg_res_fnc(outc = "FVC_most_recent", 
                            mets_list = list_analysis, 
                            dat = "pheno_tmp", 
                            add_covar = " + Age_at_collect + BMI_most_recent + Gender + race + Height +  ics_trim_totnum_5y")

## Save results
res_FVC_sph_microb_ratio <- res_4.3.1
res_FVC_sph_microb_ratio <- res_FVC_sph_microb_ratio %>% separate(Name, sep = "\\.", c("numerator", "denominator"), remove = F)

dict <- mets_info_updated[,c("Name", "CHEMICAL_NAME", "SUB_PATHWAY", "SUPER_PATHWAY")]
res_FVC_sph_microb_ratio$numerator <- dict$CHEMICAL_NAME[match(res_FVC_sph_microb_ratio$numerator, dict$Name)]
res_FVC_sph_microb_ratio$numerator_sub_pathway <- dict$SUB_PATHWAY[match(res_FVC_sph_microb_ratio$numerator, dict$CHEMICAL_NAME)]

res_FVC_sph_microb_ratio$denominator <- dict$CHEMICAL_NAME[match(res_FVC_sph_microb_ratio$denominator, dict$Name)]
res_FVC_sph_microb_ratio$denominator_sub_pathway <- dict$SUB_PATHWAY[match(res_FVC_sph_microb_ratio$denominator, dict$CHEMICAL_NAME)]

setcolorder(res_FVC_sph_microb_ratio, c("Name", "numerator", "numerator_sub_pathway", "denominator", "denominator_sub_pathway"))


saveRDS(res_FVC_sph_microb_ratio, file = here(res_dir, "res_FVC_sph_microb_ratio.rds"))

## display results
### min p-value = 1.93e-11
### #sig = 928/5743
datatable(res_FVC_sph_microb_ratio[res_FVC_sph_microb_ratio$fdr <= 0.05,], 
          filter = "top") %>% 
          formatSignif(columns = c("beta", "pval", "fdr", "fdr", "se", "tval"), digits = 3)

res_4.3.1_sig <- res_FVC_sph_microb_ratio[res_FVC_sph_microb_ratio$fdr <= 0.05,]

```

### 4.3.2 targeted sph/str
```{r}
list_analysis <- mets_missing_sph_str[mets_missing_sph_str$pct_na_met <= 0.3,"Name"]
length(list_analysis) #1827

df_sph_steroid_ratio_log$Biobank_Subject_ID <- as.integer(df_sph_steroid_ratio_log$Biobank_Subject_ID)
pheno_tmp <- merge(dat_tmp, df_sph_steroid_ratio_log, by = "Biobank_Subject_ID", all.y = T)

res_4.3.2 <- linreg_res_fnc(outc = "FVC_most_recent", 
                            mets_list = list_analysis, 
                            dat = "pheno_tmp", 
                            add_covar = " + Age_at_collect + BMI_most_recent + Gender + race + Height +  ics_trim_totnum_5y")

## Save results
res_FVC_sph_str_ratio <- res_4.3.2
res_FVC_sph_str_ratio <- res_FVC_sph_str_ratio %>% separate(Name, sep = "\\.", c("numerator", "denominator"), remove = F)

dict <- mets_info_updated[,c("Name", "CHEMICAL_NAME", "SUB_PATHWAY", "SUPER_PATHWAY")]
res_FVC_sph_str_ratio$numerator <- dict$CHEMICAL_NAME[match(res_FVC_sph_str_ratio$numerator, dict$Name)]
res_FVC_sph_str_ratio$numerator_sub_pathway <- dict$SUB_PATHWAY[match(res_FVC_sph_str_ratio$numerator, dict$CHEMICAL_NAME)]

res_FVC_sph_str_ratio$denominator <- dict$CHEMICAL_NAME[match(res_FVC_sph_str_ratio$denominator, dict$Name)]
res_FVC_sph_str_ratio$denominator_sub_pathway <- dict$SUB_PATHWAY[match(res_FVC_sph_str_ratio$denominator, dict$CHEMICAL_NAME)]

setcolorder(res_FVC_sph_str_ratio, c("Name", "numerator", "numerator_sub_pathway", "denominator", "denominator_sub_pathway"))


saveRDS(res_FVC_sph_str_ratio, file = here(res_dir, "res_FVC_sph_str_ratio.rds"))

## display results
### min p-value = 0.00000161
### #sig = 439/1827
datatable(res_FVC_sph_str_ratio[res_FVC_sph_str_ratio$fdr <= 0.05,], 
          filter = "top") %>% 
          formatSignif(columns = c("beta", "pval", "fdr", "fdr", "se", "tval"), digits = 3)

res_4.3.2_sig <- res_FVC_sph_str_ratio[res_FVC_sph_str_ratio$fdr <= 0.05,]

```


### 4.3.3 targeted microb/str
```{r}
list_analysis <- mets_missing_microb_str[mets_missing_microb_str$V1 <= 0.3,"rowname"]
length(list_analysis) #3085

df_microb_str_ratio_log$Biobank_Subject_ID <- as.integer(df_microb_str_ratio_log$Biobank_Subject_ID)
pheno_tmp <- merge(dat_tmp, df_microb_str_ratio_log, by = "Biobank_Subject_ID", all.y = T)

res_4.3.3 <- linreg_res_fnc(outc = "FVC_most_recent", 
                            mets_list = list_analysis, 
                            dat = "pheno_tmp", 
                            add_covar = " + Age_at_collect + BMI_most_recent + Gender + race + Height +  ics_trim_totnum_5y")

## Save results
res_FVC_microb_str_ratio <- res_4.3.3
res_FVC_microb_str_ratio <- res_FVC_microb_str_ratio %>% separate(Name, sep = "\\.", c("numerator", "denominator"), remove = F)

dict <- mets_info_updated[,c("Name", "CHEMICAL_NAME", "SUB_PATHWAY", "SUPER_PATHWAY")]
res_FVC_microb_str_ratio$numerator <- dict$CHEMICAL_NAME[match(res_FVC_microb_str_ratio$numerator, dict$Name)]
res_FVC_microb_str_ratio$numerator_sub_pathway <- dict$SUB_PATHWAY[match(res_FVC_microb_str_ratio$numerator, dict$CHEMICAL_NAME)]

res_FVC_microb_str_ratio$denominator <- dict$CHEMICAL_NAME[match(res_FVC_microb_str_ratio$denominator, dict$Name)]
res_FVC_microb_str_ratio$denominator_sub_pathway <- dict$SUB_PATHWAY[match(res_FVC_microb_str_ratio$denominator, dict$CHEMICAL_NAME)]

setcolorder(res_FVC_microb_str_ratio, c("Name", "numerator", "numerator_sub_pathway", "denominator", "denominator_sub_pathway"))

saveRDS(res_FVC_microb_str_ratio, file = here(res_dir, "res_FVC_microb_str_ratio.rds"))

## display results
### min p-value = 5.11e-11	
### #sig = 1060/3085
datatable(res_FVC_microb_str_ratio[res_FVC_microb_str_ratio$fdr <= 0.05,], 
          filter = "top") %>% 
          formatSignif(columns = c("beta", "pval", "fdr", "fdr", "se", "tval"), digits = 3)

res_4.3.3_sig <- res_FVC_microb_str_ratio[res_FVC_microb_str_ratio$fdr <= 0.05,]

```

## 4.4 FEV1/FVC
```{r}
ggplot(dat_tmp, aes(x=`FEV1/FVC_most_recent`)) + geom_histogram(bins = 50)
```

### 4.4.1 targeted sph/microb
```{r}
list_analysis <- mets_missing_sph_microb[mets_missing_sph_microb$V1 <= 0.3,"rowname"]
length(list_analysis) #5743

df_sph_microb_ratio_log$Biobank_Subject_ID <- as.integer(df_sph_microb_ratio_log$Biobank_Subject_ID)
pheno_tmp <- merge(dat_tmp, df_sph_microb_ratio_log, by = "Biobank_Subject_ID", all.y = T)

res_4.4.1 <- linreg_res_fnc(outc = "`FEV1/FVC_most_recent`", 
                            mets_list = list_analysis, 
                            dat = "pheno_tmp", 
                            add_covar = " + Age_at_collect + BMI_most_recent + Gender + race + Height +  ics_trim_totnum_5y")

## Save results
res_FEV1_FVC_sph_microb_ratio <- res_4.4.1
res_FEV1_FVC_sph_microb_ratio <- res_FEV1_FVC_sph_microb_ratio %>% separate(Name, sep = "\\.", c("numerator", "denominator"), remove = F)

dict <- mets_info_updated[,c("Name", "CHEMICAL_NAME", "SUB_PATHWAY", "SUPER_PATHWAY")]
res_FEV1_FVC_sph_microb_ratio$numerator <- dict$CHEMICAL_NAME[match(res_FEV1_FVC_sph_microb_ratio$numerator, dict$Name)]
res_FEV1_FVC_sph_microb_ratio$numerator_sub_pathway <- dict$SUB_PATHWAY[match(res_FEV1_FVC_sph_microb_ratio$numerator, dict$CHEMICAL_NAME)]

res_FEV1_FVC_sph_microb_ratio$denominator <- dict$CHEMICAL_NAME[match(res_FEV1_FVC_sph_microb_ratio$denominator, dict$Name)]
res_FEV1_FVC_sph_microb_ratio$denominator_sub_pathway <- dict$SUB_PATHWAY[match(res_FEV1_FVC_sph_microb_ratio$denominator, dict$CHEMICAL_NAME)]

setcolorder(res_FEV1_FVC_sph_microb_ratio, c("Name", "numerator", "numerator_sub_pathway", "denominator", "denominator_sub_pathway"))


saveRDS(res_FEV1_FVC_sph_microb_ratio, file = here(res_dir, "res_FEV1_FVC_sph_microb_ratio.rds"))

## display results
### min p-value = #
### #sig = 0/5743
datatable(res_FEV1_FVC_sph_microb_ratio[res_FEV1_FVC_sph_microb_ratio$fdr <= 0.05,], 
          filter = "top") %>% 
          formatSignif(columns = c("beta", "pval", "fdr", "fdr", "se", "tval"), digits = 3)

res_4.4.1_sig <- res_FEV1_FVC_sph_microb_ratio[res_FEV1_FVC_sph_microb_ratio$fdr <= 0.05,]

```

### 4.4.2 targeted sph/str
```{r}
list_analysis <- mets_missing_sph_str[mets_missing_sph_str$pct_na_met <= 0.3,"Name"]
length(list_analysis) #1827

df_sph_steroid_ratio_log$Biobank_Subject_ID <- as.integer(df_sph_steroid_ratio_log$Biobank_Subject_ID)
pheno_tmp <- merge(dat_tmp, df_sph_steroid_ratio_log, by = "Biobank_Subject_ID", all.y = T)

res_4.4.2 <- linreg_res_fnc(outc = "`FEV1/FVC_most_recent`", 
                            mets_list = list_analysis, 
                            dat = "pheno_tmp", 
                            add_covar = " + Age_at_collect + BMI_most_recent + Gender + race + Height +  ics_trim_totnum_5y")

## Save results
res_FEV1_FVC_sph_str_ratio <- res_4.4.2
res_FEV1_FVC_sph_str_ratio <- res_FEV1_FVC_sph_str_ratio %>% separate(Name, sep = "\\.", c("numerator", "denominator"), remove = F)

dict <- mets_info_updated[,c("Name", "CHEMICAL_NAME", "SUB_PATHWAY", "SUPER_PATHWAY")]
res_FEV1_FVC_sph_str_ratio$numerator <- dict$CHEMICAL_NAME[match(res_FEV1_FVC_sph_str_ratio$numerator, dict$Name)]
res_FEV1_FVC_sph_str_ratio$numerator_sub_pathway <- dict$SUB_PATHWAY[match(res_FEV1_FVC_sph_str_ratio$numerator, dict$CHEMICAL_NAME)]

res_FEV1_FVC_sph_str_ratio$denominator <- dict$CHEMICAL_NAME[match(res_FEV1_FVC_sph_str_ratio$denominator, dict$Name)]
res_FEV1_FVC_sph_str_ratio$denominator_sub_pathway <- dict$SUB_PATHWAY[match(res_FEV1_FVC_sph_str_ratio$denominator, dict$CHEMICAL_NAME)]

setcolorder(res_FEV1_FVC_sph_str_ratio, c("Name", "numerator", "numerator_sub_pathway", "denominator", "denominator_sub_pathway"))


saveRDS(res_FEV1_FVC_sph_str_ratio, file = here(res_dir, "res_FEV1_FVC_sph_str_ratio.rds"))

## display results
### min p-value = #
### #sig = 0/1827
datatable(res_FEV1_FVC_sph_str_ratio[res_FEV1_FVC_sph_str_ratio$fdr <= 0.05,], 
          filter = "top") %>% 
          formatSignif(columns = c("beta", "pval", "fdr", "fdr", "se", "tval"), digits = 3)

res_4.4.2_sig <- res_FEV1_FVC_sph_str_ratio[res_FEV1_FVC_sph_str_ratio$fdr <= 0.05,]

```


### 4.4.3 targeted microb/str
```{r}
list_analysis <- mets_missing_microb_str[mets_missing_microb_str$V1 <= 0.3,"rowname"]
length(list_analysis) #3085

df_microb_str_ratio_log$Biobank_Subject_ID <- as.integer(df_microb_str_ratio_log$Biobank_Subject_ID)
pheno_tmp <- merge(dat_tmp, df_microb_str_ratio_log, by = "Biobank_Subject_ID", all.y = T)

res_4.4.3 <- linreg_res_fnc(outc = "`FEV1/FVC_most_recent`", 
                            mets_list = list_analysis, 
                            dat = "pheno_tmp", 
                            add_covar = " + Age_at_collect + BMI_most_recent + Gender + race + Height +  ics_trim_totnum_5y")

## Save results
res_FEV1_FVC_microb_str_ratio <- res_4.4.3
res_FEV1_FVC_microb_str_ratio <- res_FEV1_FVC_microb_str_ratio %>% separate(Name, sep = "\\.", c("numerator", "denominator"), remove = F)

dict <- mets_info_updated[,c("Name", "CHEMICAL_NAME", "SUB_PATHWAY", "SUPER_PATHWAY")]
res_FEV1_FVC_microb_str_ratio$numerator <- dict$CHEMICAL_NAME[match(res_FEV1_FVC_microb_str_ratio$numerator, dict$Name)]
res_FEV1_FVC_microb_str_ratio$numerator_sub_pathway <- dict$SUB_PATHWAY[match(res_FEV1_FVC_microb_str_ratio$numerator, dict$CHEMICAL_NAME)]

res_FEV1_FVC_microb_str_ratio$denominator <- dict$CHEMICAL_NAME[match(res_FEV1_FVC_microb_str_ratio$denominator, dict$Name)]
res_FEV1_FVC_microb_str_ratio$denominator_sub_pathway <- dict$SUB_PATHWAY[match(res_FEV1_FVC_microb_str_ratio$denominator, dict$CHEMICAL_NAME)]

setcolorder(res_FEV1_FVC_microb_str_ratio, c("Name", "numerator", "numerator_sub_pathway", "denominator", "denominator_sub_pathway"))

saveRDS(res_FEV1_FVC_microb_str_ratio, file = here(res_dir, "res_FEV1_FVC_microb_str_ratio.rds"))

## display results
### min p-value = #	
### #sig = 0/3085
datatable(res_FEV1_FVC_microb_str_ratio[res_FEV1_FVC_microb_str_ratio$fdr <= 0.05,], 
          filter = "top") %>% 
          formatSignif(columns = c("beta", "pval", "fdr", "fdr", "se", "tval"), digits = 3)

res_4.4.3_sig <- res_FEV1_FVC_microb_str_ratio[res_FEV1_FVC_microb_str_ratio$fdr <= 0.05,]

```

## 4.5 OCS use number within 5 years
```{r}
sum(is.na(dat_tmp$ocs_trim_totnum_5y)) #0
ggplot(dat_tmp, aes(x=ocs_trim_totnum_5y)) + geom_histogram(bins = 50)
```

### 4.5.1 targeted sph/microb
```{r}
list_analysis <- mets_missing_sph_microb[mets_missing_sph_microb$V1 <= 0.3,"rowname"]
length(list_analysis) #5743

df_sph_microb_ratio_log$Biobank_Subject_ID <- as.integer(df_sph_microb_ratio_log$Biobank_Subject_ID)
pheno_tmp <- merge(dat_tmp, df_sph_microb_ratio_log, by = "Biobank_Subject_ID", all.y = T)

res_4.5.1 <- poiss_res_fnc(outc = "ocs_trim_totnum_5y", 
                           mets_list = list_analysis, 
                           dat = "pheno_tmp", 
                           add_covar = " + Age_at_collect + BMI_most_recent + Gender + race + ics_trim_totnum_5y")

## Save results
res_ocs_use_sph_microb_ratio <- res_4.5.1
res_ocs_use_sph_microb_ratio <- res_ocs_use_sph_microb_ratio %>% separate(Name, sep = "\\.", c("numerator", "denominator"), remove = F)

dict <- mets_info_updated[,c("Name", "CHEMICAL_NAME", "SUB_PATHWAY", "SUPER_PATHWAY")]
res_ocs_use_sph_microb_ratio$numerator <- dict$CHEMICAL_NAME[match(res_ocs_use_sph_microb_ratio$numerator, dict$Name)]
res_ocs_use_sph_microb_ratio$numerator_sub_pathway <- dict$SUB_PATHWAY[match(res_ocs_use_sph_microb_ratio$numerator, dict$CHEMICAL_NAME)]

res_ocs_use_sph_microb_ratio$denominator <- dict$CHEMICAL_NAME[match(res_ocs_use_sph_microb_ratio$denominator, dict$Name)]
res_ocs_use_sph_microb_ratio$denominator_sub_pathway <- dict$SUB_PATHWAY[match(res_ocs_use_sph_microb_ratio$denominator, dict$CHEMICAL_NAME)]

setcolorder(res_ocs_use_sph_microb_ratio, c("Name", "numerator", "numerator_sub_pathway", "denominator", "denominator_sub_pathway"))


saveRDS(res_ocs_use_sph_microb_ratio, file = here(res_dir, "res_ocs_use_sph_microb_ratio.rds"))

## display results
### min p-value = 6.30e-14
### #sig = 1721/5743
datatable(res_ocs_use_sph_microb_ratio[res_ocs_use_sph_microb_ratio$fdr <= 0.05,], 
          filter = "top") %>% 
          formatSignif(columns = c("beta", "pval", "fdr","se", "zval", "neg_log10_pval"), digits = 3)

res_4.5.1_sig <- res_ocs_use_sph_microb_ratio[res_ocs_use_sph_microb_ratio$fdr <= 0.05,]

```

### 4.5.2 targeted sph/str
```{r}
list_analysis <- mets_missing_sph_str[mets_missing_sph_str$pct_na_met <= 0.3,"Name"]
length(list_analysis) #1827

df_sph_steroid_ratio_log$Biobank_Subject_ID <- as.integer(df_sph_steroid_ratio_log$Biobank_Subject_ID)
pheno_tmp <- merge(dat_tmp, df_sph_steroid_ratio_log, by = "Biobank_Subject_ID", all.y = T)

res_4.5.2 <- poiss_res_fnc(outc = "ocs_trim_totnum_5y", 
                           mets_list = list_analysis, 
                           dat = "pheno_tmp", 
                           add_covar = " + Age_at_collect + BMI_most_recent + Gender + race + ics_trim_totnum_5y")

## Save results
res_ocs_use_sph_str_ratio <- res_4.5.2
res_ocs_use_sph_str_ratio <- res_ocs_use_sph_str_ratio %>% separate(Name, sep = "\\.", c("numerator", "denominator"), remove = F)

dict <- mets_info_updated[,c("Name", "CHEMICAL_NAME", "SUB_PATHWAY", "SUPER_PATHWAY")]
res_ocs_use_sph_str_ratio$numerator <- dict$CHEMICAL_NAME[match(res_ocs_use_sph_str_ratio$numerator, dict$Name)]
res_ocs_use_sph_str_ratio$numerator_sub_pathway <- dict$SUB_PATHWAY[match(res_ocs_use_sph_str_ratio$numerator, dict$CHEMICAL_NAME)]

res_ocs_use_sph_str_ratio$denominator <- dict$CHEMICAL_NAME[match(res_ocs_use_sph_str_ratio$denominator, dict$Name)]
res_ocs_use_sph_str_ratio$denominator_sub_pathway <- dict$SUB_PATHWAY[match(res_ocs_use_sph_str_ratio$denominator, dict$CHEMICAL_NAME)]

setcolorder(res_ocs_use_sph_str_ratio, c("Name", "numerator", "numerator_sub_pathway", "denominator", "denominator_sub_pathway"))


saveRDS(res_ocs_use_sph_str_ratio, file = here(res_dir, "res_ocs_use_sph_str_ratio.rds"))

## display results
### min p-value = 5.15e-33	
### #sig = 1519/1827
datatable(res_ocs_use_sph_str_ratio[res_ocs_use_sph_str_ratio$fdr <= 0.05,], 
          filter = "top") %>% 
          formatSignif(columns = c("beta", "pval", "fdr","se", "zval", "neg_log10_pval"), digits = 3)

res_4.5.2_sig <- res_ocs_use_sph_str_ratio[res_ocs_use_sph_str_ratio$fdr <= 0.05,]


## manhattan plot
res_ocs_use_sph_str_ratio <- as.data.table(res_ocs_use_sph_str_ratio)
res_ocs_use_sph_str_ratio$direction_of_effect <- with(res_ocs_use_sph_str_ratio, ifelse(beta < 0, "-", "+"))
res_ocs_use_sph_str_ratio$CHEMICAL_NAME <- paste(res_ocs_use_sph_str_ratio$numerator, res_ocs_use_sph_str_ratio$denominator, sep = "/")

res_ocs_use_sph_str_ratio <- res_ocs_use_sph_str_ratio[order(denominator_sub_pathway, numerator_sub_pathway, Name)]
res_ocs_use_sph_str_ratio$numerator_sub_pathway <- factor(res_ocs_use_sph_str_ratio$numerator_sub_pathway, 
                                              levels = as.vector(unique(res_ocs_use_sph_str_ratio$numerator_sub_pathway)))
res_ocs_use_sph_str_ratio$Name <- factor(res_ocs_use_sph_str_ratio$Name, levels = as.vector(res_ocs_use_sph_str_ratio$Name))

save_manh_plot(dat = res_ocs_use_sph_str_ratio, 
               title_infig = "sphingolipid/steroid metabolite ratio association results (ocs use)", 
               fig_name = "/res_ocs_use_sph_str_ratio.png", 
               if_lbl = T, lbl_thld = 1e-30, wdth = 29)

```

### 4.5.3 targeted microb/str
```{r}
list_analysis <- mets_missing_microb_str[mets_missing_microb_str$V1 <= 0.3,"rowname"]
length(list_analysis) #3085

df_microb_str_ratio_log$Biobank_Subject_ID <- as.integer(df_microb_str_ratio_log$Biobank_Subject_ID)
pheno_tmp <- merge(dat_tmp, df_microb_str_ratio_log, by = "Biobank_Subject_ID", all.y = T)

res_4.5.3 <- poiss_res_fnc(outc = "ocs_trim_totnum_5y", 
                           mets_list = list_analysis, 
                           dat = "pheno_tmp", 
                           add_covar = " + Age_at_collect + BMI_most_recent + Gender + race + ics_trim_totnum_5y")

## Save results
res_ocs_use_microb_str_ratio <- res_4.5.3
res_ocs_use_microb_str_ratio <- res_ocs_use_microb_str_ratio %>% separate(Name, sep = "\\.", c("numerator", "denominator"), remove = F)

dict <- mets_info_updated[,c("Name", "CHEMICAL_NAME", "SUB_PATHWAY", "SUPER_PATHWAY")]
res_ocs_use_microb_str_ratio$numerator <- dict$CHEMICAL_NAME[match(res_ocs_use_microb_str_ratio$numerator, dict$Name)]
res_ocs_use_microb_str_ratio$numerator_sub_pathway <- dict$SUB_PATHWAY[match(res_ocs_use_microb_str_ratio$numerator, dict$CHEMICAL_NAME)]

res_ocs_use_microb_str_ratio$denominator <- dict$CHEMICAL_NAME[match(res_ocs_use_microb_str_ratio$denominator, dict$Name)]
res_ocs_use_microb_str_ratio$denominator_sub_pathway <- dict$SUB_PATHWAY[match(res_ocs_use_microb_str_ratio$denominator, dict$CHEMICAL_NAME)]

setcolorder(res_ocs_use_microb_str_ratio, c("Name", "numerator", "numerator_sub_pathway", "denominator", "denominator_sub_pathway"))

saveRDS(res_ocs_use_microb_str_ratio, file = here(res_dir, "res_ocs_use_microb_str_ratio.rds"))

## display results
### min p-value = 9.53e-27	
### #sig = 2421/3085
datatable(res_ocs_use_microb_str_ratio[res_ocs_use_microb_str_ratio$fdr <= 0.05,], 
          filter = "top") %>% 
          formatSignif(columns = c("beta", "pval", "fdr","se", "zval", "neg_log10_pval"), digits = 3)

res_4.5.3_sig <- res_ocs_use_microb_str_ratio[res_ocs_use_microb_str_ratio$fdr <= 0.05,]

```


## 4.6 OCS use number after 5 years
```{r}
sum(is.na(dat_tmp$ocs_trim_totnum_5y_after)) #0
ggplot(dat_tmp, aes(x=ocs_trim_totnum_5y_after)) + geom_histogram(bins = 50)
```

### 4.6.1 targeted sph/microb
```{r}
list_analysis <- mets_missing_sph_microb[mets_missing_sph_microb$V1 <= 0.3,"rowname"]
length(list_analysis) #5743

df_sph_microb_ratio_log$Biobank_Subject_ID <- as.integer(df_sph_microb_ratio_log$Biobank_Subject_ID)
pheno_tmp <- merge(dat_tmp, df_sph_microb_ratio_log, by = "Biobank_Subject_ID", all.y = T)

res_4.6.1 <- poiss_res_fnc(outc = "ocs_trim_totnum_5y_after", 
                           mets_list = list_analysis, 
                           dat = "pheno_tmp", 
                           add_covar = " + Age_at_collect + BMI_most_recent + Gender + race + ics_trim_totnum_5y")

## Save results
res_ocs_use_after_sph_microb_ratio <- res_4.6.1
res_ocs_use_after_sph_microb_ratio <- res_ocs_use_after_sph_microb_ratio %>% separate(Name, sep = "\\.", c("numerator", "denominator"), remove = F)

dict <- mets_info_updated[,c("Name", "CHEMICAL_NAME", "SUB_PATHWAY", "SUPER_PATHWAY")]
res_ocs_use_after_sph_microb_ratio$numerator <- dict$CHEMICAL_NAME[match(res_ocs_use_after_sph_microb_ratio$numerator, dict$Name)]
res_ocs_use_after_sph_microb_ratio$numerator_sub_pathway <- dict$SUB_PATHWAY[match(res_ocs_use_after_sph_microb_ratio$numerator, dict$CHEMICAL_NAME)]

res_ocs_use_after_sph_microb_ratio$denominator <- dict$CHEMICAL_NAME[match(res_ocs_use_after_sph_microb_ratio$denominator, dict$Name)]
res_ocs_use_after_sph_microb_ratio$denominator_sub_pathway <- dict$SUB_PATHWAY[match(res_ocs_use_after_sph_microb_ratio$denominator, dict$CHEMICAL_NAME)]

setcolorder(res_ocs_use_after_sph_microb_ratio, c("Name", "numerator", "numerator_sub_pathway", "denominator", "denominator_sub_pathway"))


saveRDS(res_ocs_use_after_sph_microb_ratio, file = here(res_dir, "res_ocs_use_after_sph_microb_ratio.rds"))

## display results
### min p-value = 1.28e-10
### #sig = 1051/5743
datatable(res_ocs_use_after_sph_microb_ratio[res_ocs_use_after_sph_microb_ratio$fdr <= 0.05,], 
          filter = "top") %>% 
          formatSignif(columns = c("beta", "pval", "fdr","se", "zval", "neg_log10_pval"), digits = 3)

res_4.6.1_sig <- res_ocs_use_after_sph_microb_ratio[res_ocs_use_after_sph_microb_ratio$fdr <= 0.05,]

```

### 4.6.2 targeted sph/str
```{r}
list_analysis <- mets_missing_sph_str[mets_missing_sph_str$pct_na_met <= 0.3,"Name"]
length(list_analysis) #1827

df_sph_steroid_ratio_log$Biobank_Subject_ID <- as.integer(df_sph_steroid_ratio_log$Biobank_Subject_ID)
pheno_tmp <- merge(dat_tmp, df_sph_steroid_ratio_log, by = "Biobank_Subject_ID", all.y = T)

res_4.6.2 <- poiss_res_fnc(outc = "ocs_trim_totnum_5y_after", 
                           mets_list = list_analysis, 
                           dat = "pheno_tmp", 
                           add_covar = " + Age_at_collect + BMI_most_recent + Gender + race + ics_trim_totnum_5y")

## Save results
res_ocs_use_after_sph_str_ratio <- res_4.6.2
res_ocs_use_after_sph_str_ratio <- res_ocs_use_after_sph_str_ratio %>% separate(Name, sep = "\\.", c("numerator", "denominator"), remove = F)

dict <- mets_info_updated[,c("Name", "CHEMICAL_NAME", "SUB_PATHWAY", "SUPER_PATHWAY")]
res_ocs_use_after_sph_str_ratio$numerator <- dict$CHEMICAL_NAME[match(res_ocs_use_after_sph_str_ratio$numerator, dict$Name)]
res_ocs_use_after_sph_str_ratio$numerator_sub_pathway <- dict$SUB_PATHWAY[match(res_ocs_use_after_sph_str_ratio$numerator, dict$CHEMICAL_NAME)]

res_ocs_use_after_sph_str_ratio$denominator <- dict$CHEMICAL_NAME[match(res_ocs_use_after_sph_str_ratio$denominator, dict$Name)]
res_ocs_use_after_sph_str_ratio$denominator_sub_pathway <- dict$SUB_PATHWAY[match(res_ocs_use_after_sph_str_ratio$denominator, dict$CHEMICAL_NAME)]

setcolorder(res_ocs_use_after_sph_str_ratio, c("Name", "numerator", "numerator_sub_pathway", "denominator", "denominator_sub_pathway"))


saveRDS(res_ocs_use_after_sph_str_ratio, file = here(res_dir, "res_ocs_use_after_sph_str_ratio.rds"))

## display results
### min p-value = 8.19e-15	
### #sig = 1425/1827
datatable(res_ocs_use_after_sph_str_ratio[res_ocs_use_after_sph_str_ratio$fdr <= 0.05,], 
          filter = "top") %>% 
          formatSignif(columns = c("beta", "pval", "fdr","se", "zval", "neg_log10_pval"), digits = 3)

res_4.6.2_sig <- res_ocs_use_after_sph_str_ratio[res_ocs_use_after_sph_str_ratio$fdr <= 0.05,]

## manhattan plot
res_ocs_use_after_sph_str_ratio <- as.data.table(res_ocs_use_after_sph_str_ratio)
res_ocs_use_after_sph_str_ratio$direction_of_effect <- with(res_ocs_use_after_sph_str_ratio, ifelse(beta < 0, "-", "+"))
res_ocs_use_after_sph_str_ratio$CHEMICAL_NAME <- paste(res_ocs_use_after_sph_str_ratio$numerator, res_ocs_use_after_sph_str_ratio$denominator, sep = "/")

res_ocs_use_after_sph_str_ratio <- res_ocs_use_after_sph_str_ratio[order(denominator_sub_pathway, numerator_sub_pathway, Name)]
res_ocs_use_after_sph_str_ratio$numerator_sub_pathway <- factor(res_ocs_use_after_sph_str_ratio$numerator_sub_pathway, 
                                              levels = as.vector(unique(res_ocs_use_after_sph_str_ratio$numerator_sub_pathway)))
res_ocs_use_after_sph_str_ratio$Name <- factor(res_ocs_use_after_sph_str_ratio$Name, levels = as.vector(res_ocs_use_after_sph_str_ratio$Name))

save_manh_plot(dat = res_ocs_use_after_sph_str_ratio, 
               title_infig = "sphingolipid/steroid metabolite ratio association results (ocs use after)", 
               fig_name = "/res_ocs_use_after_sph_str_ratio.png", 
               if_lbl = T, lbl_thld = 1e-13, wdth = 36)
```


### 4.6.3 targeted microb/str
```{r}
list_analysis <- mets_missing_microb_str[mets_missing_microb_str$V1 <= 0.3,"rowname"]
length(list_analysis) #3085

df_microb_str_ratio_log$Biobank_Subject_ID <- as.integer(df_microb_str_ratio_log$Biobank_Subject_ID)
pheno_tmp <- merge(dat_tmp, df_microb_str_ratio_log, by = "Biobank_Subject_ID", all.y = T)

res_4.6.3 <- poiss_res_fnc(outc = "ocs_trim_totnum_5y_after", 
                           mets_list = list_analysis, 
                           dat = "pheno_tmp", 
                           add_covar = " + Age_at_collect + BMI_most_recent + Gender + race + ics_trim_totnum_5y")

## Save results
res_ocs_use_after_microb_str_ratio <- res_4.6.3
res_ocs_use_after_microb_str_ratio <- res_ocs_use_after_microb_str_ratio %>% separate(Name, sep = "\\.", c("numerator", "denominator"), remove = F)

dict <- mets_info_updated[,c("Name", "CHEMICAL_NAME", "SUB_PATHWAY", "SUPER_PATHWAY")]
res_ocs_use_after_microb_str_ratio$numerator <- dict$CHEMICAL_NAME[match(res_ocs_use_after_microb_str_ratio$numerator, dict$Name)]
res_ocs_use_after_microb_str_ratio$numerator_sub_pathway <- dict$SUB_PATHWAY[match(res_ocs_use_after_microb_str_ratio$numerator, dict$CHEMICAL_NAME)]

res_ocs_use_after_microb_str_ratio$denominator <- dict$CHEMICAL_NAME[match(res_ocs_use_after_microb_str_ratio$denominator, dict$Name)]
res_ocs_use_after_microb_str_ratio$denominator_sub_pathway <- dict$SUB_PATHWAY[match(res_ocs_use_after_microb_str_ratio$denominator, dict$CHEMICAL_NAME)]

setcolorder(res_ocs_use_after_microb_str_ratio, c("Name", "numerator", "numerator_sub_pathway", "denominator", "denominator_sub_pathway"))

saveRDS(res_ocs_use_after_microb_str_ratio, file = here(res_dir, "res_ocs_use_after_microb_str_ratio.rds"))

## display results
### min p-value = 1.41e-15	
### #sig = 2134/3085
datatable(res_ocs_use_after_microb_str_ratio[res_ocs_use_after_microb_str_ratio$fdr <= 0.05,], 
          filter = "top") %>% 
          formatSignif(columns = c("beta", "pval", "fdr","se", "zval", "neg_log10_pval"), digits = 3)

res_4.6.3_sig <- res_ocs_use_after_microb_str_ratio[res_ocs_use_after_microb_str_ratio$fdr <= 0.05,]

```


# 5. Save Data
```{r}

write.csv(res_4.1.1_sig, file = str_c(res_dir, "ODOLLFA_sig_targeted_sph_microb_log_ige.csv"), row.names = FALSE)
write.csv(res_4.1.2_sig, file = str_c(res_dir, "ODOLLFA_sig_targeted_sph_str_log_ige.csv"), row.names = FALSE)
write.csv(res_4.1.3_sig, file = str_c(res_dir, "ODOLLFA_sig_targeted_microb_str_log_ige.csv"), row.names = FALSE)

write.csv(res_4.2.1_sig, file = str_c(res_dir, "ODOLLFA_sig_targeted_sph_microb_FEV1.csv"), row.names = FALSE)
write.csv(res_4.2.2_sig, file = str_c(res_dir, "ODOLLFA_sig_targeted_sph_str_FEV1.csv"), row.names = FALSE)
write.csv(res_4.2.3_sig, file = str_c(res_dir, "ODOLLFA_sig_targeted_microb_str_FEV1.csv"), row.names = FALSE)

write.csv(res_4.3.1_sig, file = str_c(res_dir, "ODOLLFA_sig_targeted_sph_microb_FVC.csv"), row.names = FALSE)
write.csv(res_4.3.2_sig, file = str_c(res_dir, "ODOLLFA_sig_targeted_sph_str_FVC.csv"), row.names = FALSE)
write.csv(res_4.3.3_sig, file = str_c(res_dir, "ODOLLFA_sig_targeted_microb_str_FVC.csv"), row.names = FALSE)

write.csv(res_4.4.1_sig, file = str_c(res_dir, "ODOLLFA_sig_targeted_sph_microb_FEV1_FVC.csv"), row.names = FALSE)
write.csv(res_4.4.2_sig, file = str_c(res_dir, "ODOLLFA_sig_targeted_sph_str_FEV1_FVC.csv"), row.names = FALSE)
write.csv(res_4.4.3_sig, file = str_c(res_dir, "ODOLLFA_sig_targeted_microb_str_FEV1_FVC.csv"), row.names = FALSE)

write.csv(res_4.5.1_sig, file = str_c(res_dir, "ODOLLFA_sig_targeted_sph_microb_ocs_use.csv"), row.names = FALSE)
write.csv(res_4.5.2_sig, file = str_c(res_dir, "ODOLLFA_sig_targeted_sph_str_ocs_use.csv"), row.names = FALSE)
write.csv(res_4.5.3_sig, file = str_c(res_dir, "ODOLLFA_sig_targeted_microb_str_ocs_use.csv"), row.names = FALSE)

write.csv(res_4.6.1_sig, file = str_c(res_dir, "ODOLLFA_sig_targeted_sph_microb_ocs_use_after.csv"), row.names = FALSE)
write.csv(res_4.6.2_sig, file = str_c(res_dir, "ODOLLFA_sig_targeted_sph_str_ocs_use_after.csv"), row.names = FALSE)
write.csv(res_4.6.3_sig, file = str_c(res_dir, "ODOLLFA_sig_targeted_microb_str_ocs_use_after.csv"), row.names = FALSE)

```

# 6. Sig Percent Plot
```{r}
sig_pct_df <- read_xlsx(str_c(dat_dir, "ODOLLFA_sig_pert.xlsx"), sheet = "Sheet1") %>% as.data.table()
colnames(sig_pct_df)[1] <- "Ratio_group"
sig_pct_long_df <-  melt(sig_pct_df, id.vars = "Ratio_group", variable.name = "Phenotypes", value.name = "Pct")

sig_pct_long_df$Pct <- as.numeric(sig_pct_long_df$Pct)

plot1 <- ggplot(sig_pct_long_df, aes(x = Phenotypes, y = Pct, fill = Ratio_group)) +
         geom_bar(stat = "identity", position = position_dodge(width = 0.7), width = 0.7) +
         # ggtitle("Bar Plot with Different Groups") +
         xlab("Asthma Phenotypes") +
         ylab("Significant Percentage") +
         theme(legend.position = "none") +
         theme(panel.background = element_rect(fill = "white"),
          plot.background = element_rect(fill = "white", color = "white"),
          panel.grid.major = element_blank(),
          panel.grid.minor = element_blank(),
          axis.title.y = element_text(face = "bold"),
          axis.text.y = element_text(face = "bold"))

ggsave(file = here::here(fig_dir, "/sig_mets_bar_plot.png"), plot = plot1, width = 10, height = 4)
```


# Session info
```{r}
sessionInfo()

```
